% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{add_horizons_to_quantiles}
\alias{add_horizons_to_quantiles}
\title{Add columns indicating the horizon to the quantiles}
\usage{
add_horizons_to_quantiles(scores, max_weeks_nowcast = 2)

add_horizons_to_quantiles(scores, max_weeks_nowcast = 2)
}
\arguments{
\item{scores}{A tibble containing the quantiled values and the following
required columns: \code{date}and \code{forecast_date}}

\item{max_weeks_nowcast}{An integer indicating the expected maximum number
of weeks where date < forecast date. If the output contains values beyond
this threshold, an error will be thrown which may indicate that the scores
contain scores for calibrated data.}

\item{quantiles}{A tibble containing the quantiled values and the following
required columns: \code{date}, \code{forecast_date}, and \code{period}}
}
\value{
a tibble containing the same columns as \code{quantiles} plus
\code{horizon_weeks} and \code{horizon}

a tibble containing the same columns as \code{scores} plus
\code{horizon_weeks} and \code{horizon}
}
\description{
This function takes in a tibble of quantiles containing the columns
\code{date}, \code{forecast_date} and \code{period} and adds the following columns:
\code{horizon_weeks} (an integer) and \code{horizon} (a string to be used for
categorical grouping of horizons)

This function takes in a tibble of scores on the nowcasts/forecasts
containing the columns \code{date} and \code{forecast_date} and adds the following
columns: \code{horizon_weeks} (an integer) and \code{horizon} (a string to be used for
categorical grouping of horizons)
}
