% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{add_time_indexing}
\alias{add_time_indexing}
\title{Add time indexing to hospital admissions data}
\usage{
add_time_indexing(input_hosp_data)
}
\arguments{
\item{input_hosp_data}{data frame with dates and admissions,
but without time indexing.}
}
\value{
The same data frame, with an added
time index, including NA rows if dates internal
to the timeseries are missing admissions data.
}
\description{
Add time indexing to hospital admissions data
}
\examples{
hosp_data_example <- tibble::tibble(
  date = lubridate::ymd("2024-01-01", "2024-01-02", "2024-01-06"),
  daily_hosp_admits = c(5, 3, 8)
)
hosp_data_w_t <- add_time_indexing(hosp_data_example)
}
