% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{assert_needed_env_vars}
\alias{assert_needed_env_vars}
\title{Assert that needed environment variables are set}
\usage{
assert_needed_env_vars(needed_vars)
}
\arguments{
\item{needed_vars}{Vector of needed environment
variables.}
}
\value{
\code{NULL}, invisibly on success or raise an error.
}
\description{
Assert that needed environment variables are set
}
\examples{

tryCatch(
  assert_needed_env_vars(c(
    "WWEVAL_EXAMPLE_ONE",
    "WWEVAL_EXAMPLE_TWO"
  )),
  error = \(e) print(e)
)

}
