% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{barplot_avg_wis}
\alias{barplot_avg_wis}
\title{Plot average WIS as a barplot.}
\usage{
barplot_avg_wis(average_scores, time_period, fig_file_dir)
}
\arguments{
\item{average_scores}{Scores for models, with
\code{avg_wis} and \code{model} columns.}

\item{time_period}{string indicating the time period covered,
for labeling the plot.}

\item{fig_file_dir}{directory to save figure}
}
\value{
The figure, also saving it to disk as a side effect.
}
\description{
Plot average WIS as a barplot.
}
