% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{benchmark_performance}
\alias{benchmark_performance}
\title{Benchmark performance of wastewater and hospital admissions model
foreast runs}
\usage{
benchmark_performance(
  ww_scores,
  hosp_scores,
  benchmark_dir,
  benchmark_scope,
  overwrite_benchmarking
)
}
\arguments{
\item{ww_scores}{A tibble containing the scores for all target horizon days,
forecast dates, and locations for the wastewater model}

\item{hosp_scores}{A tibble containing the scores for all target horizon
days,forecast dates, and locations for the hospital admissions only model}

\item{benchmark_dir}{A directory indicating where to save the yaml file
containing tables and the git hash for benchmarking}

\item{benchmark_scope}{A string indicating verbally what the scores
are summarized over. Options are "all_forecasts" or "subset_of_forecasts"}

\item{overwrite_benchmarking}{A boolean indicating whether or not to
write the yaml to the benchmarking directory}
}
\value{
a list containing the metadata and 3 tables summarizing forecast
performance overall, by forecast date, and by location
}
\description{
Benchmark performance of wastewater and hospital admissions model
foreast runs
}
