% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{check_any_flags}
\alias{check_any_flags}
\title{Check whether a model run had any flags.}
\usage{
check_any_flags(forecast_dir, forecast_date, run_on_date, run_id)
}
\arguments{
\item{forecast_dir}{Path to a top-level forecast directory.}

\item{forecast_date}{Forecast date, as a string in YYYY-MM-DD format.}

\item{run_on_date}{Date the model was run, as a string in
YYYY-MM-DD format.}

\item{run_id}{run UUID, as a string.}
}
\value{
\code{FALSE} if there were no flags, \code{TRUE} if there were
any flags.
}
\description{
Check whether a model run had any flags.
}
