% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{clean_and_filter_nwss_data}
\alias{clean_and_filter_nwss_data}
\title{Initial subsetting of NWSS data}
\usage{
clean_and_filter_nwss_data(raw_nwss_data)
}
\arguments{
\item{raw_nwss_data}{nwss data from nwss}
}
\value{
nwss_subset_raw which just dplyr::filters based on sample types and wwtp and
returns a subset of the columns
}
\description{
Grab the columns we want and subset to raw wastewater and to only
Wastewater treatment plants (rather than downstream sites, for now).
Transform concentration to copies per mL
}
