% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{clean_flag_df}
\alias{clean_flag_df}
\title{Clean flags from a real-time forecast}
\usage{
clean_flag_df(df)
}
\arguments{
\item{df}{Dataframe of flags}
}
\value{
A cleaned version of the data frame,
with flag names corrected.
}
\description{
Fixes typos and computes needed quantities if absent
}
