% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_secrets.R
\name{cleanup_secrets}
\alias{cleanup_secrets}
\title{Cleans up environment when secrets and API information is no longer needed}
\usage{
cleanup_secrets()
}
\value{
nothing, unsets values of environmental variables set by setup_secrets
}
\description{
Cleans up environment when secrets and API information is no longer needed
}
