% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_summarize_ww_data.R
\name{combine_and_summarize_ww_data}
\alias{combine_and_summarize_ww_data}
\title{Combine and summarize wastewater data}
\usage{
combine_and_summarize_ww_data(forecast_dates, locations, eval_output_subdir)
}
\arguments{
\item{forecast_dates}{The vector of character strings of all the forecast
dates}

\item{locations}{The vector of character strings of all the locations}

\item{eval_output_subdir}{The outer subdirectory of the nested file structure}
}
\value{
ww_metadata: a tibble with a row for each forecast date location and
columns that provide summaries of the wastewater data
}
\description{
This function iterates through the vector of forecast dates and locations,
checks for the presence of wastewater data, and if present loads it in.
It then computes a number of summary metrics on the wastewater data for that
particular location and forecast date to produce a single row of output
metadata. Locations without wastewater data are indicated as such and have
NAs for the wastewater specific entries
}
