% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{compose_and_save_hub_figure}
\alias{compose_and_save_hub_figure}
\title{Make a figure showing comparisons to Hub models across
the entire season and from February onward.}
\usage{
compose_and_save_hub_figure(
  plot_wis_t,
  hist_rwis,
  barplot_wis,
  heatmap_rel_wis,
  qq_plot,
  figure_name,
  fig_file_dir
)
}
\arguments{
\item{plot_wis_t}{average wis over time
across locations for ach model in real-time (feb-mar)}

\item{hist_rwis}{Histogram of relative WIS
across location, forecast_date, day, and model}

\item{barplot_wis}{bar chart in order of average WIS}

\item{heatmap_rel_wis}{heatmap comparing WIS across
forecast dates for each location.}

\item{qq_plot}{qq plot comparing model coverage}

\item{figure_name}{Filename for the figure, without the extension.}

\item{fig_file_dir}{Directory in which to save figures}
}
\value{
a patchwork object containing all the figures combined. As a
side effect, saves the figure to disk both as a png and as an svg.
}
\description{
Make a figure showing comparisons to Hub models across
the entire season and from February onward.
}
