% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_bootstrap_fig}
\alias{compose_bootstrap_fig}
\title{Make a multi-panel figure summarizing relative
performance with bootstrapping}
\usage{
compose_bootstrap_fig(
  abs_score_overall,
  rel_score_overall,
  rel_score_by_time,
  rel_score_by_location
)
}
\arguments{
\item{abs_score_overall}{plot of overall absolute scores
with bootstrapped uncertainty intervals.}

\item{rel_score_overall}{plot of overall relative score
with bootstrapped uncertainty intervals.}

\item{rel_score_by_time}{timeseries plot of relative score
by forecast date with bootstrapped uncertainty intervals.}

\item{rel_score_by_location}{plot of (date-specific) score values by location with bootstrapped uncertainty intervals.}
}
\value{
patchwork object with all the elements combined
}
\description{
Make a multi-panel figure summarizing relative
performance with bootstrapping
}
