% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_clock_time_fig}
\alias{compose_clock_time_fig}
\title{Make a multi-panel figure summarizing wall clock runtime.}
\usage{
compose_clock_time_fig(
  clock_time_dist_full,
  clock_time_dist_used,
  clock_time_versus_sites_full,
  clock_time_versus_sites_used
)
}
\arguments{
\item{clock_time_dist_full}{Distribution of clock time values for all
fits, including non-convergent fits.}

\item{clock_time_dist_used}{Distribution of clock time values for just
convergent fits.}

\item{clock_time_versus_sites_full}{Plot of clock time versus sites for all
fits, including non-convergent fits.}

\item{clock_time_versus_sites_used}{Plot of clock time versus sites just
for convergent fits.}
}
\value{
patchwork object with all the elements combined
}
\description{
Make a multi-panel figure summarizing wall clock runtime.
}
