% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_diagnostic_fig}
\alias{compose_diagnostic_fig}
\title{Make a multi-panel summarizing diagnostics.}
\usage{
compose_diagnostic_fig(
  max_rhat_used,
  max_rhat_not_used,
  ess_bulk_forecast_used,
  ess_bulk_forecast_not_used,
  ess_tail_forecast_used,
  ess_tail_forecast_not_used
)
}
\arguments{
\item{max_rhat_used}{Distribution of max R-hat values for
fits that were used in the analysis.}

\item{max_rhat_not_used}{Distribution of max R-hat values for fits
that were not used (often due to non-convergence)}

\item{ess_bulk_forecast_used}{Distribution of worst bulk ESS for
a scored prediction for fits used in the analysis.}

\item{ess_bulk_forecast_not_used}{Distribution of worst bulk ESS for
a scored prediction for fits not used in the analysis}

\item{ess_tail_forecast_used}{Distribution of worst tail ESS for
a scored prediction for fits used in the analysis.}

\item{ess_tail_forecast_not_used}{Distribution of worst tail ESS for
a scored prediction for fits not used in the analysis.}
}
\value{
patchwork object with all the elements combined
}
\description{
Make a multi-panel summarizing diagnostics.
}
