% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_example_scores_fig}
\alias{compose_example_scores_fig}
\title{Compose figure showing examples of forecasts and their scores.}
\usage{
compose_example_scores_fig(
  score_single_loc1,
  forecast_comparison_nowcast1,
  forecast_comparison_1wk1,
  forecast_comparison_4wk1,
  score_underlay_nowcast1,
  score_underlay_1wk1,
  score_underlay_4wk1,
  score_single_loc2,
  forecast_comparison_nowcast2,
  forecast_comparison_1wk2,
  forecast_comparison_4wk2,
  score_underlay_nowcast2,
  score_underlay_1wk2,
  score_underlay_4wk2,
  score_single_loc3,
  forecast_comparison_nowcast3,
  forecast_comparison_1wk3,
  forecast_comparison_4wk3,
  score_underlay_nowcast3,
  score_underlay_1wk3,
  score_underlay_4wk3
)
}
\arguments{
\item{score_single_loc1}{first states score density plot}

\item{forecast_comparison_nowcast1}{first states nowcast comparison}

\item{forecast_comparison_1wk1}{first states 1 wk forecast comparison}

\item{forecast_comparison_4wk1}{first states 4 wk forecast comparison}

\item{score_underlay_nowcast1}{first states score nowcast underlay}

\item{score_underlay_1wk1}{first states score 1 wk underlay}

\item{score_single_loc2}{second states score density plot}

\item{forecast_comparison_nowcast2}{second states nowcast comparison}

\item{forecast_comparison_1wk2}{second states 1 wk forecast comparison}

\item{forecast_comparison_4wk2}{second states 4 wk forecast comparison}

\item{score_underlay_nowcast2}{second states score nowcast underlay}

\item{score_underlay_1wk2}{second states score 1 wk underlay}

\item{score_underlay_4wk2}{second states score 4wk underlay}

\item{score_single_loc3}{first state's score density plot}

\item{forecast_comparison_nowcast3}{third states nowcast comparison}

\item{forecast_comparison_1wk3}{third states 1 wk forecast comparison}

\item{forecast_comparison_4wk3}{third states 4 wk forecast comparison}

\item{score_underlay_nowcast3}{third states score nowcast underlay}

\item{score_underlay_1wk3}{third states score 1 wk underlay}

\item{score_underlay_4wks1}{first states score 4wk underlay}

\item{score_underlay_4wks3}{third states score 4wk underlay}
}
\value{
ggplot object that is a combination of 3 states overall score
distributions comparing the two model types +
forecast comparisons across horizons with an underlay
indicating the score.
}
\description{
Compose figure showing examples of forecasts and their scores.
}
