% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_hub_fig}
\alias{compose_hub_fig}
\title{Functions for creating multi-panel
composite figures
Make a figure showing comparisons between CFA models
and other COVIDHub models}
\usage{
compose_hub_fig(plot_wis_t, hist_rwis, barplot_wis, heatmap_rel_wis, qq_plot)
}
\arguments{
\item{plot_wis_t}{average wis over time
across locations for ach model in real-time (feb-mar)}

\item{hist_rwis}{Histogram of relative WIS
across location, forecast_date, day, and model}

\item{barplot_wis}{bar chart in order of average WIS}

\item{heatmap_rel_wis}{heatmap comparing WIS across
forecast dates for each location.}

\item{qq_plot}{qq plot comparing model coverage}
}
\value{
a patchwork object containing all the figures combined.
}
\description{
Functions for creating multi-panel
composite figures
Make a figure showing comparisons between CFA models
and other COVIDHub models
}
