% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_pred_actual_fig}
\alias{compose_pred_actual_fig}
\title{Compose figure showing predictions and observations
for three example forecasts.}
\usage{
compose_pred_actual_fig(hosp1, hosp2, hosp3, ww1, ww2, ww3)
}
\arguments{
\item{hosp1}{first hospital admissions predicted-actual figure}

\item{hosp2}{second hospital admissions predicted-actual figure}

\item{hosp3}{third hospital admissions predicted-actual figure}

\item{ww1}{first wastewater predicted-actual figure}

\item{ww2}{second wastewater predicted-actual figure}

\item{ww3}{third wastewater predicted-actual figure}

\item{fig_file_dir}{Path to save figures}
}
\value{
a combined ggplot object
}
\description{
Compose figure showing predictions and observations
for three example forecasts.
}
