% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{compute_flags}
\alias{compute_flags}
\title{Compute diagnostic flags from raw stanfit objects}
\usage{
compute_flags(forecast_dir, forecast_date, run_on_date, run_id)
}
\arguments{
\item{forecast_dir}{Path to a top-level forecast directory.}

\item{forecast_date}{Forecast date, as a string in YYYY-MM-DD format.}

\item{run_on_date}{Date the model was run, as a string in
YYYY-MM-DD format.}

\item{run_id}{run UUID, as a string.}
}
\value{
The flags, as the output of \code{\link[=get_diagnostic_flags]{get_diagnostic_flags()}}.
}
\description{
Compute diagnostic flags from raw stanfit objects
}
