% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_forecast_differences.R
\name{compute_forecast_difference}
\alias{compute_forecast_difference}
\title{Compute the the posterior of the difference between model predictions.}
\usage{
compute_forecast_difference(forecast_date, location, scenario, raw_output_dir)
}
\arguments{
\item{forecast_date}{forecast_date for which to fits recent data
trends. Data for an actual forecast must already have been produced
and exist in \code{raw_output_dir}.}

\item{location}{Location for which to fit trends.}

\item{scenario}{Wastewater data availability scenario to analyze.}

\item{raw_output_dir}{Directory containing raw output \code{.rds} files.
Used to obtain the admissions and wastewater data used in fitting
the forecasting model.}
}
\value{
List containing two tables, one containing
posteriors for the difference by target date, the other
containing the posterior for the total difference across
all target dates.
}
\description{
Compute the the posterior of the difference between model predictions.
}
