% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_forecast_differences.R
\name{compute_forecast_differences}
\alias{compute_forecast_differences}
\title{Compute the the posterior of the difference between model predictions.}
\usage{
compute_forecast_differences(
  forecast_date,
  location,
  scenario,
  raw_output_dir,
  offset = 1
)
}
\arguments{
\item{forecast_date}{forecast_date for which to fits recent data
trends. Data for an actual forecast must already have been produced
and exist in \code{raw_output_dir}.}

\item{location}{Location for which to fit trends.}

\item{scenario}{Wastewater data availability scenario to analyze.}

\item{raw_output_dir}{Directory containing raw output \code{.rds} files.
Used to obtain the admissions and wastewater data used in fitting
the forecasting model.}

\item{offset}{to add to incidences when computing a log difference.
Default \code{1}.}
}
\value{
Posterior of the difference in total incidence
across the full nowcast/forecast period. Saves
the incidence differences by day and the total
incidence difference to disk as a side effect.
}
\description{
Compute the the posterior of the difference between model predictions.
}
