% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_relative_crps.R
\name{compute_relative_crps}
\alias{compute_relative_crps}
\title{Compute relative CRPS}
\usage{
compute_relative_crps(scores, id_cols)
}
\arguments{
\item{scores}{A tibble of scores containing the specified \code{id_cols},
\code{model}, and the scoring metrics from  \code{scoringutils::score()} applied on
samples}

\item{id_cols}{A vector of character strings listing the columns of metadata
we want to keep as identifiers}
}
\value{
a pivoted wide tibble with side-by-side \code{ww} and \code{hosp} scores
for each \code{id_cols} as well as a \code{rel_crps} column for the relative
CRPS
}
\description{
Compute relative CRPS
}
