% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{convert_to_boostrap_input}
\alias{convert_to_boostrap_input}
\title{Prepare bootstrap input from a table of paired absolute and relative retro
CPRS for the wastewater-informed and hospital-admissions only model.}
\usage{
convert_to_boostrap_input(df)
}
\arguments{
\item{df}{table of paired CRPS values, with columns \code{model}, \code{crps},
and \code{rel_crps}. Model column must include the retro wastewater-informed
model (\code{cfa-wwrenewal(retro)}).}
}
\value{
A table that can be passed to \code{\link[=bootstrap_crps_values]{bootstrap_crps_values()}}.
}
\description{
Prepare bootstrap input from a table of paired absolute and relative retro
CPRS for the wastewater-informed and hospital-admissions only model.
}
