% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\docType{data}
\name{covidhub_submissions_raw}
\alias{covidhub_submissions_raw}
\alias{covidhub_truth_data}
\title{Github URL of raw Covidhub submissions files.}
\format{
An object of class \code{character} of length 1.

An object of class \code{character} of length 1.
}
\usage{
covidhub_submissions_raw

covidhub_truth_data
}
\description{
Github URL of raw Covidhub submissions files.

Github URL of Covidhub truth data file
}
\keyword{datasets}
