% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hub_submissions.R
\name{create_hub_submissions}
\alias{create_hub_submissions}
\title{Create hub submission files}
\usage{
create_hub_submissions(
  hosp_quantiles_ww,
  hosp_quantiles_hosp,
  forecast_dates,
  hub_subdir,
  model_name,
  scenario = "status_quo",
  save_files = TRUE
)
}
\arguments{
\item{hosp_quantiles_ww}{a dataframe of the probabilistic estimates of
hospital admissions formatted as quantiles, from the wastewater model,
for multiple forecast dates if present for a particular location}

\item{hosp_quantiles_hosp}{a dataframe of the probabilistic estimates of
hospital admissions formatted as quantiles, from the hospital admissions
model, for multiple forecast dates}

\item{forecast_dates}{vector of forecast dates that we want to create
mock submissions for}

\item{hub_subdir}{character string of the outer directory specifying where to
write the submission files to (which will be named by forecast date)}

\item{model_name}{character string indicating the name of the model. This will
be used to determine both the name of the submission \code{.csv} file and the name of
its enclosing directory, per COVID-19 Forecast Hub formatting.}

\item{scenario}{string indicating which wastewater availability scenario to use when creating
the hub submission file. Default \code{"status_quo"} (all actually available wastewater data).}

\item{save_files}{Save the created submission data frame to disk as a \code{.csv}
file? Boolean, default \code{TRUE}.}
}
\value{
metadata_df a dataframe that has one row per forecast date indicating
thenumber of wastewater submissions in that date and the number of total
locations
}
\description{
This function takes in the combined set of wastewater and hospital admissions
quantiles by forecast date, date, and location, creates a dataframe
formatted as it would have been for submission to the COVID forecast Hub,
and saves this to disk if specified. It returns the metadata on the number
of locations total and the number of locations using the wastewater model
for each forecast date submission saved.
It assumes the wastewater-informed model output quantiles passed in have already been
filtered to remove ones that we would wish to exclude (for convergence
diagnostics or other reasons)
}
