% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mock_submissions.R
\name{create_mock_submission_scores}
\alias{create_mock_submission_scores}
\title{Create mock submission scores}
\usage{
create_mock_submission_scores(
  all_scores,
  name_of_replacement_model = "no_wastewater"
)
}
\arguments{
\item{all_scores}{This a dataframe of scores for all the successful model runs.
Because there are locations in each scenario where the wastewater data is
missing, this is not all the combinations of scenarios, forecast dates, and
locations}

\item{name_of_replacement_model}{the the string identifier in the \code{scenario}
of \code{all_scores} that should be used to replace the missing locations for
each scenario and forecast date. Default is \code{no_wastewater} as we will
typically use the hospital admissions only model for this.}
}
\value{
\code{all_submission_scores} which will be an expanded dataframe with
scores for each day with evaluation data for all combinations of
forecast dates, locations, and scenarios, with the name of replacement
model scores filling in when scores are missing.
}
\description{
For each forecast date, we want to create a mock submission to the Hub
so forecasts for all states + territories for each scenario. Since we
know the scores are independent and we just do a weighted average over them
to get an overall score, we can just take the scores directly from the
"no_wastewater" scenario whenever a wastewater score is missing,
add it to the dataframe, and continue to concatenate until we get a full
score dataframe with all combinations of forecast dates, locations, and
scenarios, labeled according to scenario. We will still know which ones
are from which model with the `model`` column
}
