% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{date_of_ww_data}
\alias{date_of_ww_data}
\title{Date of wastewater data}
\usage{
date_of_ww_data(forecast_date, ww_data_mapping, ww_data_dir)
}
\arguments{
\item{forecast_date}{the forecast date for this iteration}

\item{ww_data_mapping}{a string that tells this function how to pick
data pull dates from forecast dates. This function needs to be configured
for each new string.}

\item{ww_data_dir}{A string indicating the path to the directory containing
time stamped wastewater datasets}
}
\value{
the date to get the ww data from
}
\description{
Date of wastewater data
}
