% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_secrets.R
\name{.get_api_needed}
\alias{.get_api_needed}
\title{Just a list of things the package needs, for convenience}
\usage{
.get_api_needed()
}
\value{
the API token/secret/ID, as a character
}
\description{
Just a list of things the package needs, for convenience
}
\keyword{internal}
