% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{.get_ratio_of_bstrap_means}
\alias{.get_ratio_of_bstrap_means}
\title{Compute the ratio of bootstrapped mean CRPS values for the two
models.}
\usage{
.get_ratio_of_bstrap_means(df, by = NULL)
}
\arguments{
\item{df}{Data frame of bootstrapped CRPS values, as the output of
\code{\link[=.summarize_bstrap_crps]{.summarize_bstrap_crps()}}}

\item{by}{Variables to summarize by when computing ratio of means.
Passed as the \code{.by} argument to \code{\link[dplyr:summarize]{dplyr::summarize()}}. Default \code{NULL}.}
}
\value{
table of the ratios
}
\description{
Compute the ratio of bootstrapped mean CRPS values for the two
models.
}
\keyword{internal}
