% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.path_with_suffix}
\alias{.path_with_suffix}
\title{Construct a path to a serialized object}
\usage{
.path_with_suffix(dir, basename, suffix, ext)
}
\arguments{
\item{dir}{directory containing the object}

\item{basename}{basename of the object}

\item{suffix}{suffix for the object}

\item{ext}{file extension for the object}
}
\description{
Helper function for \code{\link[=save_rds_with_suffix]{save_rds_with_suffix()}},
\code{\link[=read_rds_with_suffix]{read_rds_with_suffix()}}, and
\code{\link[=file_exists_object_with_suffix]{file_exists_object_with_suffix()}}.
}
\keyword{internal}
