% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{.process_bstrap_samples}
\alias{.process_bstrap_samples}
\title{Process a dataframe of bootstrapped CRPS values.}
\usage{
.process_bstrap_samples(df)
}
\arguments{
\item{df}{Output of \code{\link[rsample:bootstraps]{rsample::bootstraps()}} applied to a table of
relative scores for the wastewater model, with columns \code{crps} and
\code{mean_scores_ratio}.}
}
\value{
A tibble of replicate values for \code{crps_hosp} and \code{crps_ww},
organized by bootstrap replicate id.
}
\description{
Process a dataframe of bootstrapped CRPS values.
}
\keyword{internal}
