% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{.summarize_bstrap_crps}
\alias{.summarize_bstrap_crps}
\title{Summarize bootstrapped CRPS}
\usage{
.summarize_bstrap_crps(df, by = NULL)
}
\arguments{
\item{df}{Data frame of replicate CRPS values for the two models,
as the output of \code{\link[=.process_bstrap_samples]{.process_bstrap_samples()}}.}

\item{by}{Additional summarization variables besides \code{"id"}.
\code{c("id", by)} will be passed as the \code{.by} argument to
\code{\link[dplyr:summarize]{dplyr::summarize()}}. Default \code{NULL}.}
}
\value{
A summary of the data frame by replicate dataset \code{id} and optionally
other grouping variables, including the mean CRPS for the wastewater model
for each replicate dataset and group, the mean CRPS for the admissions-only model
for each replicate dataset and group, and the ratio of those two means for each
replicate dataset and group.
}
\description{
Summarize bootstrapped CRPS
}
\keyword{internal}
