% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{drop_first_and_renormalize}
\alias{drop_first_and_renormalize}
\title{Drop the first element of a simplex}
\usage{
drop_first_and_renormalize(x)
}
\arguments{
\item{x}{A numeric vector, sums to 1. Corresponds to a discretized PDF or PMF
(usually the GI distribution).}
}
\value{
A numeric vector, sums to 1.
}
\description{
When this vector corresponds to the generation interval distribution, we
want to drop this first bin. The renewal equation assumes that same-day
infection and onward transmission does not occur, and we assume
everything is 1 indexed not 0 indeced. We need to
manually drop the first element from the PMF vector.
}
