% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fit.R
\name{eval_fit_hosp}
\alias{eval_fit_hosp}
\title{Evaluate Hospitalizations Fit}
\usage{
eval_fit_hosp(config_index, eval_config_path, params_path, output_dir)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}

\item{output_dir}{Path to directory for storing output}
}
\description{
Evaluate Hospitalizations Fit
}
