% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_post_process.R
\name{eval_post_process_hosp}
\alias{eval_post_process_hosp}
\title{Post Process Hospitalizations Model for Evaluation}
\usage{
eval_post_process_hosp(
  config_index,
  eval_config_path,
  params_path,
  max_eval_data_days = 365
)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}

\item{max_eval_data_days}{Maximum number of days of data to pull
when creating evaluation dataset. Default 365.}
}
\description{
Post Process Hospitalizations Model for Evaluation
}
