% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exclude_hosp_outliers.R
\name{exclude_hosp_outliers}
\alias{exclude_hosp_outliers}
\title{Exclude hospital outliers}
\usage{
exclude_hosp_outliers(
  raw_input_hosp_data,
  forecast_date,
  table_of_exclusions,
  col_name_dates_to_exclude = "dates_to_exclude"
)
}
\arguments{
\item{raw_input_hosp_data}{A dataframe containing the location, the date of
admission, and the number of hospital admissions, in that location}

\item{forecast_date}{The forecast date as a character string}

\item{table_of_exclusions}{A table containing the combinations of
locations, forecast dates, and dates to exclude. These will be treated
as missing by the model}

\item{col_name_dates_to_exclude}{The name of the column in the table of
exclusiosn that cooresponds to the date of hospital admissions to exclude,
default is \code{dates_to_exclude}}
}
\value{
A dataframe in the same format as the \code{raw_input_hosp_data} but
with the location-forecast-date-date admissions rows excluded if indicated
in the \code{table_of_exclusions}
}
\description{
Exclude hospital outliers
}
