% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_exists_object_with_suffix}
\alias{file_exists_object_with_suffix}
\title{Check that a file for corresponding to a
programmatically-named project output object
exists.}
\usage{
file_exists_object_with_suffix(object_basename, dir, suffix = "", ext = "rds")
}
\arguments{
\item{object_basename}{Base name for the object, before any
\code{suffix} or the file extension, as a string.}

\item{dir}{Directory in which to look for the object.}

\item{suffix}{Suffix to append to the \code{object_basename} before
the file extension. Default \code{""} (no suffix).}

\item{ext}{Extension for the saved file, without the \code{.}.
Default \code{rds}.}
}
\value{
\code{TRUE} if the file exists, else \code{FALSE}.
}
\description{
Check that a file for corresponding to a
programmatically-named project output object
exists.
}
