% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{filter_sites_by_scenario}
\alias{filter_sites_by_scenario}
\title{Filter sites by scenario}
\usage{
filter_sites_by_scenario(
  init_subset_nwss_data,
  scenario = "Status quo",
  scenario_dir = NA
)
}
\arguments{
\item{init_subset_nwss_data}{a dataframe of the raw NWSS data
filtered to exclude solids and upstream sites}

\item{scenario}{a string indicating what scenario we are running.
Default is "Status quo" which uses all the data we have available}

\item{scenario_dir}{a string indicating the file path where the
scenario csvs will live. Default is NA because we don't need this
for the status quo scenario}
}
\value{
a dataframe that only contains the ww data from
the sites in the list pertaining to the scenario
}
\description{
Filter sites by scenario
}
