% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_recent_trends.R
\name{fit_ww_trend}
\alias{fit_ww_trend}
\title{Fit lab-site-level recent wastewater trends with BRMS.}
\usage{
fit_ww_trend(
  data,
  exp_rate_conc_pop_prior_mean,
  exp_rate_conc_pop_prior_sd,
  sd_exp_rate_conc_prior_mode,
  sd_exp_rate_conc_prior_sd,
  t0_conc_pop_prior_mean,
  t0_conc_pop_prior_sd,
  sd_t0_conc_prior_mode,
  sd_t0_conc_prior_sd,
  log_sd_obs_pop_prior_mean,
  log_sd_obs_pop_prior_sd,
  sd_log_sd_obs_prior_mode,
  sd_log_sd_obs_prior_sd,
  seed,
  ...
)
}
\arguments{
\item{data}{Data to fit. Must have columns \code{"time"},
\verb{'lab_site_index"}, \code{"conc"}, and \code{"cens"}.}

\item{exp_rate_conc_pop_prior_mean}{Mean for the Normal prior on
the population-wide exponential growth rate of wastewater viral
genome concentrations.}

\item{exp_rate_conc_pop_prior_sd}{Standard deviation for the
Normal prior on the population-wide exponential growth rate of
wastewater viral genome concentrations.}

\item{sd_exp_rate_conc_prior_mode}{Mode for the positive-constrained
Normal prior on the standard deviation of wastewater viral
genome concentration exponential growth rates (which governs
the variability in growth rate among lab-sites)}

\item{sd_exp_rate_conc_prior_sd}{Standard deviation for the
positive-constrained Normal prior on the standard deviation
of wastewater viral genome concentration exponential growth
rates (which governs the variability in growth rate
among lab-sites)}

\item{t0_conc_pop_prior_mean}{Mean for the Normal prior on the
population-wide wastewater viral genome concentration at time
t = 0.}

\item{t0_conc_pop_prior_sd}{Standard deviation for the Normal
prior on the population-wide wastewater viral genome concentration
at time t = 0.}

\item{sd_t0_conc_prior_mode}{Mode for the positive-constrained
Normal prior on the standard deviation of wastewater viral genome
concentrations at time t = 0 (which governs the variability in
initial concentration among lab-sites).}

\item{sd_t0_conc_prior_sd}{Standard deviation for the
positive-constrained Normal prior on the standard deviation of
wastewater viral genome concentrations at time t = 0
(which governs the variability in initial concentration
among lab-sites).}

\item{log_sd_obs_pop_prior_mean}{Mean for the log of the
population-wide observation error standard deviation
(observation errors are Normal).}

\item{log_sd_obs_pop_prior_sd}{Standard deviation for the
log of the population-wide observation error
standard deviation (observation errors are Normal).}

\item{sd_log_sd_obs_prior_mode}{Mode for the positive-constrained
Normal prior on the standard deviation of log lab-site-specific
observation error standard deviations (which governs the variability
in observation error scale among lab-sites).}

\item{sd_log_sd_obs_prior_sd}{Standard deviation for the
positive-constrained Normal prior on the standard deviation
observation error standard deviations (which governs the variability
in observation error scale among lab-sites).}

\item{seed}{Seed for cmdnstan's pseudorandom number generator.}

\item{...}{Additional keyword arguments passed to \code{\link[brms:brm]{brms::brm()}}}
}
\description{
Fit lab-site-level recent wastewater trends with BRMS.
}
