% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{flag_ww_outliers}
\alias{flag_ww_outliers}
\title{Flag WW outliers}
\usage{
flag_ww_outliers(
  ww_data,
  rho_threshold = 2,
  log_conc_threshold = 3,
  threshold_n_dps = 1
)
}
\arguments{
\item{ww_data}{data at the lab-site level of WW concentrations}

\item{rho_threshold}{z-score threshold for "jump"}

\item{log_conc_threshold}{z-score threshold for log concentration}

\item{threshold_n_dps}{min number of data points above the LOD per lab-site}
}
\value{
ww_data + columns for outlier flagging
}
\description{
Flag WW outliers
}
