% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{flusight_location_table}
\alias{flusight_location_table}
\title{flusight_location_table dataset}
\format{
A tibble with 58 rows and 3 columns:
\describe{
\item{location_code}{The name of the location}
\item{short_name}{Two letter code for the location}
\item{long_name}{The full name of the location}
}
}
\source{
Originally downloaded from
\url{https://www2.census.gov/geo/docs/reference/state.txt}
}
\usage{
flusight_location_table
}
\description{
This dataset codifies US states and territories.
}
\examples{
data(flusight_location_table)
head(flusight_location_table)
}
\keyword{dataset,}
\keyword{location}
