% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forecast_interval_coverage_plot}
\alias{forecast_interval_coverage_plot}
\title{Plot interval coverage at specified ranges}
\usage{
forecast_interval_coverage_plot(
  forecasts,
  ranges,
  time_period,
  by = c("model", "horizon")
)
}
\arguments{
\item{forecasts}{df of granular (daily) quantile forecasts,
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.}

\item{ranges}{A numeric vector of credible interval ranges to plot,
spanning from 0 to 100.}

\item{time_period}{string indicating time period of fig to save}

\item{by}{Columns by which to stratify. Passed as the \code{by}
argument to \code{\link[scoringutils:get_coverage]{scoringutils::get_coverage()}}. Default \code{c("model", "horizon")}.}
}
\value{
A ggplot2 object
}
\description{
Plot interval coverage at specified ranges
}
