% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{forecast_output_path}
\alias{forecast_output_path}
\title{Standard path to an output directory for a specific forecast}
\usage{
forecast_output_path(output_dir, scenario, forecast_date, model, location)
}
\arguments{
\item{output_dir}{Path to the general output directory.}

\item{scenario}{Wastewater data availability scenario to analyze}

\item{forecast_date}{Forecast date, in as a string in
YYYY-MM-DD format.}

\item{model}{Name of the model fit.}

\item{location}{Forecast location, as as a two-letter USPS
code.}
}
\value{
The path
}
\description{
Standard path to an output directory for a specific forecast
}
