% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_calibration.R
\name{forecast_qq_plot}
\alias{forecast_qq_plot}
\title{Functions for plots that assess forecast calibration
Get quantile-quantile plot}
\usage{
forecast_qq_plot(
  forecasts,
  model_z_order = NULL,
  linewidth = 2,
  reference_linecolor = "gray",
  reference_linewidth = 2,
  reference_linetype = "dashed",
  ...
)
}
\arguments{
\item{forecasts}{df of granular (daily) quantile forecasts}

\item{model_z_order}{z-order in which to overplot the individual models,
ascending (so the last named model is plotted on top).
If \code{NULL} (default), plot the models in the order they appear in
\code{forecasts}.}

\item{linewidth}{\code{width} parameter for the Q-Q lines for
individual models. Default 2.}

\item{reference_linecolor}{\code{color} parameter for the r
eference y = x line. Default \code{"gray"}.}

\item{reference_linewidth}{\code{width} parameter for the
reference y = x line. Default \code{2}.}

\item{reference_linetype}{\code{linetype} parameter for the
reference y = x line. Default \code{"dashed"}.}

\item{...}{additional keyword arguments passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}
for the Q-Q lines for individual models.}
}
\value{
a ggplot object containing a plot of the proportion of data
within each interval for each model.
}
\description{
Adapted from MIT-licensed \code{\link[scoringutils:plot_quantile_coverage]{scoringutils::plot_quantile_coverage()}}.
}
