% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_hub_submissions.R
\name{format_for_hub}
\alias{format_for_hub}
\title{Format quantiles for the current COVID Forecast Hub}
\usage{
format_for_hub(
  quantiles,
  date_col_name = "date",
  value_col_name = "value",
  loc_col_name = "location",
  forecast_date_col_name = "forecast_date",
  quantile_col_name = "quantile"
)
}
\arguments{
\item{quantiles}{dataframe of quantiles containing columns with the specified
names for date, value, location, forecast date, and quantile}

\item{date_col_name}{string indicating the column name for date of the
forecasted quantity, default is \code{date}}

\item{value_col_name}{string indicating the column name for the value of the
prediction at the date, default is \code{value}}

\item{loc_col_name}{string indicating the column name for the state
abbreviation corresponding to the forecast location, default \code{location}}

\item{forecast_date_col_name}{string indicating the column name for the forecast
date, default is \code{forecast_date}}

\item{quantile_col_name}{string indicating the column name for the quantile
value, default is \code{quantile}}
}
\value{
a dataframe in Hub formatting
}
\description{
This function takes in a dataframe of quantiles and their
values and the names of the columns specifying the date, location,
value, quantile, and forecast date, and returns a hub formatted dataframe
}
