% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_all_training_data}
\alias{get_all_training_data}
\title{Get long tibble of data needed for model fitting}
\usage{
get_all_training_data(
  ww_data_raw,
  forecast_date,
  location,
  calibration_time,
  geo_type,
  pull_from_local,
  hosp_data_dir,
  population_data_path,
  forecast_time,
  hosp_reporting_delay,
  ww_geo_type,
  hosp_data_source,
  include_ww,
  train_data_dir,
  model_type,
  write_files = TRUE,
  ...
)
}
\arguments{
\item{ww_data_raw}{NEEDS DOCUMENTATION}

\item{forecast_date}{NEEDS DOCUMENTATION}

\item{location}{NEEDS DOCUMENTATION}

\item{calibration_time}{NEEDS DOCUMENTATION}

\item{geo_type}{NEEDS DOCUMENTATION}

\item{pull_from_local}{NEEDS DOCUMENTATION}

\item{hosp_data_dir}{NEEDS DOCUMENTATION}

\item{population_data_path}{NEEDS DOCUMENTATION}

\item{forecast_time}{NEEDS DOCUMENTATION}

\item{hosp_reporting_delay}{NEEDS DOCUMENTATION}

\item{ww_geo_type}{NEEDS DOCUMENTATION}

\item{hosp_data_source}{NEEDS DOCUMENTATION}

\item{include_ww}{NEEDS DOCUMENTATION}

\item{train_data_dir}{NEEDS DOCUMENTATION}

\item{model_type}{NEEDS DOCUMENTATION}

\item{write_files}{NEEDS DOCUMENTATION}

\item{...}{additional arguments}
}
\value{
a long stacked dataframe with columns corresponding to model specs
and data specs
}
\description{
Get long tibble of data needed for model fitting
}
