% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_avg_scores_model_horizon}
\alias{get_avg_scores_model_horizon}
\title{Get avg of scores by horizon}
\usage{
get_avg_scores_model_horizon(scores, score_type)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{score_type}{string indicating the type of score to summarize over}
}
\value{
table of avg scores by horizon and overall by model
}
\description{
Get avg of scores by horizon
}
