% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_box_plot}
\alias{get_box_plot}
\title{Get box plot by date and scenario}
\usage{
get_box_plot(
  all_scores,
  figure_file_path,
  baseline_scenario = "no_wastewater",
  save_files = TRUE
)
}
\arguments{
\item{all_scores}{the full set of scores for each location, forecast date,
scenario, and date}

\item{figure_file_path}{Outer directory for plots to go in}

\item{baseline_scenario}{the scenario to compute the difference in
crps scores relative to, default is \code{no_wastewater}}

\item{save_files}{Whether or not to write to a folder of plots,
default is \code{TRUE}}
}
\value{
p a ggplot object with a box plot showing the average improvement in
scores by forecast date and scenario for each location
}
\description{
Get box plot by date and scenario
}
