% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_date_locs_excluded.R
\name{get_date_locs_excluded}
\alias{get_date_locs_excluded}
\title{Get forecast date locations that we excluded in the model run}
\usage{
get_date_locs_excluded(dates)
}
\arguments{
\item{dates}{vector of character strings indicating the forecast dates that
we want to pull from our repository to check for exclusions}
}
\value{
a table of locations and forecast dates to exclude because we e
excluded them in real-time
}
\description{
Get forecast date locations that we excluded in the model run
}
