% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_filepath}
\alias{get_filepath}
\title{Get filepath}
\usage{
get_filepath(
  output_subdir,
  scenario,
  forecast_date,
  model_type,
  location,
  output_type,
  file_extension
)
}
\arguments{
\item{output_subdir}{Character string to upper level directory where outputs
live}

\item{scenario}{string indicating the evaluation scenario}

\item{forecast_date}{string indicating the forecast date in "YYYY-MM-DD"
format}

\item{model_type}{string indicating the type of model, either "ww" or "hosp"
for now but can be used for model comparison}

\item{location}{string indicating the location}

\item{output_type}{string indicating the type of output}

\item{file_extension}{string indicating the file extension e.g. ".tsv"}
}
\value{
string of the full filepath
}
\description{
A function to return a character string for each item to either be saved or
read in, based on the nesting that is specified in this function. To change
the nesting, we will change this function
}
