% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_full_scores}
\alias{get_full_scores}
\title{Get the scores for ever day for a particular location and forecast date}
\usage{
get_full_scores(draws, scenario)
}
\arguments{
\item{draws}{a dataframe of the model estimated  quantity you are evaluating
alongside the evaluation data}

\item{scenario}{a string indicating the wastewater data scenario we're
running}
}
\value{
a dataframe containing a score for each day in the nowcast
and forecast period
}
\description{
Uses scoringutils to transform data and predictions using a log transform
with a log shift of 1, and return the default scoring metrics
}
