% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_growth_rate_plot}
\alias{get_growth_rate_plot}
\title{Compare growth rates in wastewater and hospital admissions data}
\usage{
get_growth_rate_plot(
  input_hosp_data,
  input_ww_data,
  location,
  forecast_date,
  rate = "daily",
  align = "center"
)
}
\arguments{
\item{input_hosp_data}{tibble of hospital admissions counts for a
particular location and forecast date}

\item{input_ww_data}{tibble of wastewater concentrations for a particular
location and forecast date}

\item{location}{string indicating the location of interest}

\item{forecast_date}{string indicating the forecast date of interest}

\item{rate}{string indicating whether the plot should be a daily or
weekly growth rates. Default is \code{"daily"}}

\item{align}{string indicating how to align the 7 day rolling average
of the hospital admissions and wastewater data. Default is \code{"center"}}
}
\value{
a ggplot object containing site-lab level exponential growth rates
with the state level hospital admissions growth rates overlaid
}
\description{
Compare growth rates in wastewater and hospital admissions data
}
