% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_heatmap_relative_wis}
\alias{get_heatmap_relative_wis}
\title{Get heatmap of relative WIS}
\usage{
get_heatmap_relative_wis(
  scores,
  figure_file_path,
  time_period,
  baseline_model = "COVIDhub-baseline",
  save_files = TRUE
)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{figure_file_path}{path to save figure}

\item{time_period}{time period that scores are summarized over}

\item{baseline_model}{which model to compute relative WIS compared to, default
is \code{COVIDhub-baseline}}

\item{save_files}{save_files boolean indicating whether or not to save figures, default
is \code{TRUE}}
}
\value{
a ggplot with a heatmap with model on the x-axis, location on the y-axis
and fill by relative WIS score across forecast dates and horizons
}
\description{
Get heatmap of relative WIS
}
