% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{get_hosp_data_sizes}
\alias{get_hosp_data_sizes}
\title{Get hospital data integer sizes for stan}
\usage{
get_hosp_data_sizes(
  input_hosp_data,
  forecast_date,
  forecast_time,
  calibration_time,
  last_hosp_data_date,
  uot,
  hosp_value_col_name = "daily_hosp_admits"
)
}
\arguments{
\item{input_hosp_data}{a dataframe with the input hospital admissions data}

\item{forecast_date}{string indicating the forecast date}

\item{forecast_time}{integer indicating the number of days to make a forecast
for}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to hospital admissions for}

\item{last_hosp_data_date}{string indicating the date of the last observed
hospital admission}

\item{uot}{integer indicating the time of model initialization when there are
no observations}

\item{hosp_value_col_name}{A string represeting the name of the column in the
input_hosp-data that indicates the number of daily hospital admissions,
default is \code{daily_hosp_admits}}
}
\value{
A list containing the integer sizes of the follow variables that
the stan model requires:
ht:  integer indicating horizon time for the model(hospital admissions
nowcast + forecast time in days)
ot: integer indicating the total duration of time that the hospital admissions
model has available calibration data
oht: integer indicating the number of hospital admission observations
n_weeks: number of weeks (rounded up) that hospital admissions are generated
from the model
tot_weeks: number of week(rounded up) that infections are generated for
}
\description{
Get hospital data integer sizes for stan
}
