% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{get_hosp_indices}
\alias{get_hosp_indices}
\title{Get hospital admissions indices}
\usage{
get_hosp_indices(input_hosp_data)
}
\arguments{
\item{input_hosp_data}{a dataframe with the input hospital admissions data}
}
\value{
A list containing the vectors of indices that
the stan model requires:
hosp_times: a vector of integer times corresponding to the times when the
hospital admissions observations were made
}
\description{
Get hospital admissions indices
}
