% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig3.R
\name{get_ind_forecast_score}
\alias{get_ind_forecast_score}
\title{Get an individual forecast score summary for a particular
forecast date and location}
\usage{
get_ind_forecast_score(scores, loc, this_forecast_date)
}
\arguments{
\item{scores}{A tibble of all forecast date- location- model
forecast perfomance scores}

\item{loc}{the location of interest}

\item{this_forecast_date}{the forecast date of interest}
}
\value{
A tibble of mean scores by models
}
\description{
Get an individual forecast score summary for a particular
forecast date and location
}
