% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_input_ww_data}
\alias{get_input_ww_data}
\title{Get input wastewater data}
\usage{
get_input_ww_data(
  forecast_date_i,
  location_i,
  scenario_i,
  scenario_dir,
  ww_data_dir,
  calibration_time,
  last_hosp_data_date,
  ww_data_mapping,
  for_eval = FALSE
)
}
\arguments{
\item{forecast_date_i}{The forecast date for this iteration,
formatted as a character string in IS08601 format (YYYY-MM-DD).}

\item{location_i}{The location (state or other jurisdiction)
for this iteration, formatted as a string (uppercase USPS two-letter
abbreviation, e.g. AK for Alaska, DC for the District of Columbia,
PR for Puerto Rico).}

\item{scenario_i}{The scenario for this iteration, formatted as a
string}

\item{scenario_dir}{A string indicating the path to the directory
containing the csvs with the wwtp ids needed for each scenario}

\item{ww_data_dir}{A string indicating the path to the directory
containing time stamped wastewater datasets}

\item{calibration_time}{The duration of the model calibration period
(relative to the last hospital admissions data point) in units of
model timesteps (typically days).}

\item{last_hosp_data_date}{A date indicating the date of last reported
hospital admission as of the forecast date}

\item{ww_data_mapping}{A string indicating how to map the
forecast date to the wastewater dates (see \code{\link[=date_of_ww_data]{date_of_ww_data()}}
for more details)}

\item{for_eval}{A boolean indicating whether the hospital admissions data
is for evaluation. Default is FALSE which means it will be used to fit
a single model, TRUE means we will combine with multiple locations and a
longer time span than we would fit to.}
}
\value{
a tibble containing the pre-processed wastewater data ready to
be passed into the wwinference function, or \code{NULL} if no input data is available.
}
\description{
Get input wastewater data
}
