% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_last_hosp_data_date}
\alias{get_last_hosp_data_date}
\title{Get last hospital admissions data point date}
\usage{
get_last_hosp_data_date(input_hosp)
}
\arguments{
\item{input_hosp}{the hospital admissions dataset for that location and
forecast date}
}
\value{
a date indicating the last day of observed data
}
\description{
Get last hospital admissions data point date
}
