% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{get_last_hosp_data_date_map}
\alias{get_last_hosp_data_date_map}
\title{Get a map of the location, forecast date, and last hospital admissions data
date}
\usage{
get_last_hosp_data_date_map(df)
}
\arguments{
\item{df}{A tibble containing the following columns: \code{forecast_date},
\code{location},\code{date}, \code{calib_data}.
\code{calib_data} should be \code{NA} for any dates for which there was
not hospital admissions data to fit to.}
}
\value{
A tibble with that maps the unique combinations of \code{location} and
\code{forecast} date to the last hospital admissions data date
\code{last_hosp_data_date}
}
\description{
Get a map of the location, forecast date, and last hospital admissions data
date
}
