% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_model_draws_w_data}
\alias{get_model_draws_w_data}
\title{Get model draws combined with input and evaluation data}
\usage{
get_model_draws_w_data(
  fit_obj_wwinference,
  model_output = c("ww", "hosp"),
  model_type = c("ww", "hosp"),
  forecast_date,
  scenario,
  location,
  eval_data
)
}
\arguments{
\item{fit_obj_wwinference}{wwinference_fit object that is returned when
wwinference::wwinference() is run.}

\item{model_output}{the type of model expected observation you want,
options are "hosp" and "ww"}

\item{model_type}{The type of model, options are "ww" and "hosp"}

\item{forecast_date}{The date the forecast was made}

\item{scenario}{A name for the scenario that the input
data represents, as a string.}

\item{location}{The location for which the model is being run.}

\item{eval_data}{The retrospective dataset used to evaluate the model
(should have data beyond the forecast date)}
}
\value{
a dataframe of model draws subsetted to only the specified output
type, joined with the evaluation data and the input calibration data
}
\description{
Get model draws combined with input and evaluation data
}
