% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filepath_mapping.R
\name{get_model_path}
\alias{get_model_path}
\title{Get model path}
\usage{
get_model_path(model_type, stan_models_dir)
}
\arguments{
\item{model_type}{string specifying the model to be run, options are either
'hosp' for the hospital admissions only model or
'ww' for the site-level infection dynamics model using wastewater}

\item{stan_models_dir}{directory where stan files are located}
}
\value{
string indicating path to correct stan file
}
\description{
Get model path
}
\examples{
model_path <- get_model_path("hosp", system.file("stan",
  package = "cfaforecastrenewalww"
))
}
