% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_object_loader}
\alias{get_object_loader}
\title{Generate a function for loading raw output objects
for a location/forecast date/scenario trio.}
\usage{
get_object_loader(location, forecast_date, scenario, raw_output_dir)
}
\arguments{
\item{location}{Name of the location, as a string.}

\item{forecast_date}{The forecast date, as a date or
in a format coercible by \code{\link[=as.Date]{as.Date()}}.}

\item{scenario}{Name of the scenario, as a string.}

\item{raw_output_dir}{Directory in which to look for
the \code{.rds} file.}
}
\value{
A function for loading objects from \code{.rds} files,
as the output of calling \code{\link[purrr:partial]{purrr::partial()}} on
\code{\link[=read_rds_with_suffix]{read_rds_with_suffix()}}.
}
\description{
Generate a function for loading raw output objects
for a location/forecast date/scenario trio.
}
