% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{get_parameter_display_name}
\alias{get_parameter_display_name}
\title{Convert parameter names from the wwinference Stan model
to display names.}
\usage{
get_parameter_display_name(parameter_name)
}
\arguments{
\item{parameter_name}{Name of the parameter in the wwinference Stan model,
as a string.}
}
\value{
The display name, as a string. Returns the raw parameter name
if no display name is found.
}
\description{
Convert parameter names from the wwinference Stan model
to display names.
}
