% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_posterior.R
\name{get_parameterized_prior_dist}
\alias{get_parameterized_prior_dist}
\title{Get a properly parameterized prior distribution
for selected model parameters.}
\usage{
get_parameterized_prior_dist(parameter_name, prior_params)
}
\arguments{
\item{parameter_name}{Name of the parameter in the wwinference Stan model}

\item{prior_params}{Named list of prior hyperparameter values, as the
output of \code{\link[wwinference:get_params]{wwinference::get_params()}}}
}
\value{
\verb{A distributional} distribution object parameterized by the priors,
e.g. a \code{\link[distributional:dist_normal]{distributional::dist_normal()}} object. Errors if the user requests
an unknown parameter.
}
\description{
Get a properly parameterized prior distribution
for selected model parameters.
}
