% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{get_path_to_ww_quants_or_rerun}
\alias{get_path_to_ww_quants_or_rerun}
\title{Either get the path to the ww quantiles or regenerate them}
\usage{
get_path_to_ww_quants_or_rerun(
  scenarios,
  forecast_dates,
  locations,
  eval_output_subdir,
  ww_quantiles_path,
  rerun_ww_postprocess,
  model_type = "ww",
  output_type = "ww_quantiles"
)
}
\arguments{
\item{scenarios}{vector of character strings indicating scenarios to load in}

\item{forecast_dates}{vector of character strings indicating forecast dates
to load in}

\item{locations}{vector of character strings indicating states to load in}

\item{eval_output_subdir}{string indicating the subdir to find the things
to load in if needed}

\item{ww_quantiles_path}{string indicating the full path to where the
ww_quantiles either are saved or should be saved}

\item{rerun_ww_postprocess}{Boolean indicating whether or not to force
rerun the ww quantiles. If FALSE, then grab the quantiles in the path}

\item{model_type}{string indicating the model type, default is \code{ww}}

\item{output_type}{string indicating the output type, default is \code{ww}}
}
\value{
a character string indicating the path to the wastewater quantiles
}
\description{
Either get the path to the ww quantiles or regenerate them
}
