% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_hub_performance}
\alias{get_plot_hub_performance}
\title{Get plot of overall hub performance}
\usage{
get_plot_hub_performance(
  all_scores,
  cfa_real_time_scores,
  figure_file_path,
  all_time_period,
  real_time_period,
  baseline_model = "COVIDhub-baseline",
  save_files = TRUE
)
}
\arguments{
\item{all_scores}{df with granular (daily) scores from every model,
forecast_date, and location for the entire time period. Includes the
two retrospective models}

\item{cfa_real_time_scores}{df with granular (daily) scores from the
submitted cfa ww model for the time period when it was submitted.}

\item{figure_file_path}{path to directory to save figures}

\item{all_time_period}{string indicating the longer time frame we are
comparing, e.g. "Oct 2023-Mar 2024"}

\item{real_time_period}{string indicating the shorter time frame that
we submitted our model to the hub e.g. "Feb 2024-Mar 2024"}

\item{baseline_model}{which model to compute relative WIS compared to, default
is \code{COVIDhub-baseline}}

\item{save_files}{boolean indicating whether or not to save figures, default
is \code{TRUE}}
}
\value{
a ggplot object containing distributions of WIS scores grouped by
model and the comaprison time period, with the mean plotted alongside the
full distribution
}
\description{
Get plot of overall hub performance
}
