% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_plot_quantile_comparison}
\alias{get_plot_quantile_comparison}
\title{Get plot quantile comparison}
\usage{
get_plot_quantile_comparison(
  hosp_quantiles,
  eval_data,
  figure_file_path,
  days_to_show_forecast = 28,
  save_files = TRUE
)
}
\arguments{
\item{hosp_quantiles}{dataframe of hospital admissions quantiles}

\item{eval_data}{Hospital admissions data for visual comparison}

\item{figure_file_path}{Outer directory for plots to go in}

\item{days_to_show_forecast}{Number of days to show the forecast, default
vlaue is 28}

\item{save_files}{Whether or not to write to a folder of plots,
default is \code{TRUE}}
}
\value{
a ggplot object with forecasts overlaid with evaluation data for
each scenarion in a specific location for visual comparison
}
\description{
Get plot quantile comparison
}
