% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_plot_rel_crps_real_time}
\alias{get_plot_rel_crps_real_time}
\title{Get a plot of the relative crps from the real-time models}
\usage{
get_plot_rel_crps_real_time(rel_scores, fig_file_dir)
}
\arguments{
\item{rel_scores}{tibble containing the relative crps for each forecast
date and location and horizon day}

\item{fig_file_dir}{string indicating the directory to save the figure}
}
\value{
ggplot object of a heatmap of the realtive crps
}
\description{
Get a plot of the relative crps from the real-time models
}
