% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_plot_sites_vs_performance}
\alias{get_plot_sites_vs_performance}
\title{Make a scatterplot comparing number of wastewater
sites to model performance.}
\usage{
get_plot_sites_vs_performance(scores, ww_metadata, fig_file_dir)
}
\arguments{
\item{scores}{data frame of scores}

\item{ww_metadata}{data frame of wastewater metadata.}

\item{fig_file_dir}{directory in which to save the figure.}
}
\value{
The figure, saving it to disk as a side effect.
}
\description{
Make a scatterplot comparing number of wastewater
sites to model performance.
}
