% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_wis_over_time}
\alias{get_plot_wis_over_time}
\title{Get plot of WIS over time}
\usage{
get_plot_wis_over_time(
  all_scores,
  cfa_real_time_scores,
  figure_file_path,
  horizon_time_in_weeks = NULL,
  save_files = TRUE
)
}
\arguments{
\item{all_scores}{Scores from entire time period of interest, including
the retrospective cfa model}

\item{cfa_real_time_scores}{Real-time scores from Feb - Mar for the cfa ww
model submitted to the hub}

\item{figure_file_path}{directory to save figures in}

\item{horizon_time_in_weeks}{horizon time in weeks to summarize over, default
is \code{NULL} which means that the scores are summarized over 4 weeks}

\item{save_files}{bolean indicating whether or not to save the files, default
is \code{TRUE}}
}
\value{
a ggplot object of WIS scores over time colored by model, for the
real-time cfa model from Feb - Mar and the retrospective CFA model over
all time points
}
\description{
Get plot of WIS over time
}
