% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_plot_ww_data}
\alias{get_plot_ww_data}
\title{Plot wastewater evaluation data}
\usage{
get_plot_ww_data(eval_data)
}
\arguments{
\item{eval_data}{Data frame of evaluation data}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot wastewater evaluation data
}
