% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_plot_ww_data_comparison}
\alias{get_plot_ww_data_comparison}
\title{Get plot of wastewater data compared to model draws}
\usage{
get_plot_ww_data_comparison(
  draws_w_data,
  location,
  model_type = "ww",
  n_draws = 100
)
}
\arguments{
\item{draws_w_data}{A long tidy dataframe containing draws from the model of
the estimated wastewater concentrations in each site joined with both the data
the model was calibrated to and the later observed data for evaluating the
future predicted concentrations against.}

\item{location}{the jurisdiction the data is from}

\item{model_type}{type of model the output is from, default is \code{ww}}

\item{n_draws}{number of draws to plot, default = 100}
}
\value{
a ggplot object faceted by site showing the draws
}
\description{
Get plot of wastewater data compared to model draws
}
