% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{get_plot_xy_raw_limits}
\alias{get_plot_xy_raw_limits}
\title{Get the x and y limits of an individual ggplot,
on the raw data scale.}
\usage{
get_plot_xy_raw_limits(plot)
}
\arguments{
\item{plot}{plot whose limits to extract}
}
\value{
The x and y limits, as a named list with
entries xmin, xmax, ymin, and ymax
}
\description{
Adapted from this stackoverflow implementation
https://stackoverflow.com/a/40304848
}
