% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_qq_plot}
\alias{get_qq_plot}
\title{Get quantile-quantile plot}
\usage{
get_qq_plot(scores, figure_file_path, time_period, save_files = TRUE)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{figure_file_path}{path to save figure}

\item{time_period}{time period that scores are summarized over}

\item{save_files}{save_files boolean indicating whether or not to save figures, default
is \code{TRUE}}
}
\value{
a ggplot object containing a plot of the proportion of data within
each interval for each model.
}
\description{
Get quantile-quantile plot
}
