% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_regions_for_mapping}
\alias{get_regions_for_mapping}
\title{: get states mapped to their regions}
\usage{
get_regions_for_mapping()
}
\value{
state abbreviations and region that corr to the WW data source for
the biobot data
}
\description{
: get states mapped to their regions
}
