% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_scenario_site_ids}
\alias{get_scenario_site_ids}
\title{Get scenario site ids}
\usage{
get_scenario_site_ids(
  init_subset_nwss_data,
  scenario = "status_quo",
  scenario_dir = NA
)
}
\arguments{
\item{init_subset_nwss_data}{a dataframe of the raw NWSS data
filtered to exclude solids and upstream sites}

\item{scenario}{a string indicating what scenario we are running.
Default is "status_quo" which uses all the data we have available}

\item{scenario_dir}{a string indicating the file path where the
scenario csvs will live. Default is NA because we don't need this
for the status_quo scenario}
}
\value{
a dataframe that only contains the ww data from
the sites in the list pertaining to the scenario
}
\description{
Get scenario site ids
}
