% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_score_summary_tables}
\alias{get_score_summary_tables}
\title{Make a summary of the with and without wastewater comparison scores}
\usage{
get_score_summary_tables(scores)
}
\arguments{
\item{scores}{a tibble of scores for each model, horizon day, forecast date
and location for the subset of forecasts used in the head-to-head comparison}
}
\value{
a list of two tables with summary scores, one overall and one by
forecast vs nowcast
}
\description{
Make a summary of the with and without wastewater comparison scores
}
