% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_scores_from_quantiles}
\alias{get_scores_from_quantiles}
\title{Get the scores for every day for a location, forecast date, and scenario
from the quantiles during the forecast period}
\usage{
get_scores_from_quantiles(quantiles, scenario, metrics = NULL)
}
\arguments{
\item{quantiles}{a dataframe of the model estimated quantiles alongside
the data you are evaluating against, during the nowcast and forecast period
only}

\item{scenario}{a string indicating the wastewater data scenario we're
running}

\item{metrics}{Vector of scoring metrics to output, passed as the
\code{metrics} argument to \code{\link[scoringutils:score]{scoringutils::score()}}.
Default is NULL which will
include all scoring metrics for quantiles by default, including
\code{c("interval_score", "coverage", "dispersion", "bias")}.}
}
\value{
a dataframe containing a score for each day in the nowcast
and forecast period
}
\description{
Uses scoringutils to transform data and predictions using a log transform
with a log shift of 1, and return the default scoring metrics
}
