% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_secrets.R
\name{get_secret}
\alias{get_secret}
\title{Interface for getting information needed for API interfacing}
\usage{
get_secret(secret)
}
\arguments{
\item{secret}{character, name of the API token/secret/ID needed}
}
\value{
the API token/secret/ID, as a character
}
\description{
Interface for getting information needed for API interfacing
}
