% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{get_shared_xy_raw_limits}
\alias{get_shared_xy_raw_limits}
\title{Get shared (spanning) x and y limits for a set of
ggplot objects, on the raw data scale.}
\usage{
get_shared_xy_raw_limits(list_of_plots)
}
\arguments{
\item{list_of_plots}{List of ggplot objects for which to obtain
the spanning limits}
}
\value{
The shared limits, as a named list with entries
entries xmin, xmax, ymin, and ymax
}
\description{
Get shared (spanning) x and y limits for a set of
ggplot objects, on the raw data scale.
}
