% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{get_site_county_map}
\alias{get_site_county_map}
\title{Get site county mapp}
\usage{
get_site_county_map(nwss, county_site_map_path)
}
\arguments{
\item{nwss}{wastewater data}

\item{county_site_map_path}{path to a county site map CSV file;
if it exists, it will be used, otherwise the file will be created and placed here}
}
\value{
a mapping from the unique county fips in the ww data to the major county
}
\description{
Get site county mapp
}
