% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{get_stan_data_list}
\alias{get_stan_data_list}
\title{Get stan data}
\usage{
get_stan_data_list(
  model_type,
  forecast_date,
  forecast_time,
  calibration_time,
  input_ww_data,
  input_hosp_data,
  generation_interval,
  inf_to_hosp,
  infection_feedback_pmf,
  params,
  compute_likelihood = 1,
  ww_outlier_col_name = "flag_as_ww_outlier",
  lod_col_name = "below_LOD",
  ww_measurement_col_name = "ww",
  ww_value_lod_col_name = "lod_sewage",
  hosp_value_col_name = "daily_hosp_admits"
)
}
\arguments{
\item{model_type}{string indicating which model we are getting data for
Options are \code{ww} or \code{hosp}}

\item{forecast_date}{string indicating the forecast date in ISO 8601 convention
e.g. YYYY-MM-DD}

\item{forecast_time}{integer indicating the number of days to make a forecast
for}

\item{calibration_time}{integer indicating the max duration in days that
the model is calibrated to hospital admissions for}

\item{input_ww_data}{a dataframe with the input wastewater data}

\item{input_hosp_data}{a dataframe with the input hospital admissions data}

\item{generation_interval}{a vector with a zero-truncated normalized pmf of
the generation interval}

\item{inf_to_hosp}{a vector with a normalized pmf of the delay from infection
to hospital admissions}

\item{infection_feedback_pmf}{a vector with a normalized pmf dictating the
delay of infection feedback}

\item{params}{a dataframe of parameter names and numeric values}

\item{compute_likelihood}{indicator variable telling stan whether or not to
compute the likelihood, default = \code{1}}

\item{ww_outlier_col_name}{A string representing the name of the
column in the input_ww_data that provides a 0 if the data point is not an
outlier to be excluded from the model fit, or a 1 if it is to be excluded
default value is \code{flag_as_ww_outlier}}

\item{lod_col_name}{A string representing the name of the
column in the input_ww_data that provides a 0 if the data point is not above
the LOD and a 1 if the data is below the LOD, default value is \code{below_LOD}}

\item{ww_measurement_col_name}{A string representing the name of the column
in the input_ww_data that indicates the wastewater measurement value in
natural scale, default is \code{ww}}

\item{ww_value_lod_col_name}{A string representing the name of the column
in the input_ww_data that indicates the value of the LOD in natural scale,
default is \code{lod_sewage}}

\item{hosp_value_col_name}{A string representing the name of the column in the
input_hosp-data that indicates the number of daily hospital admissions,
default is \code{daily_hosp_admits}}
}
\value{
a list named variables to pass to stan
}
\description{
Get stan data
}
