% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_state_level_quantiles}
\alias{get_state_level_quantiles}
\title{Get quantiles for state-level generated quantities}
\usage{
get_state_level_quantiles(draws)
}
\arguments{
\item{draws}{a dataframe containing all the draws from the model estimated
state-level quantities}
}
\value{
a dataframe containing the quantile value for the quantiles
required for the Hub submission
}
\description{
Get quantiles for state-level generated quantities
}
