% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_stats_improved_forecasts}
\alias{get_stats_improved_forecasts}
\title{Get stats on number of improved forecasts}
\usage{
get_stats_improved_forecasts(
  scores,
  threshold,
  target_model,
  baseline_model,
  metric_to_compare
)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{threshold}{numeric indicating fold change for considering a forecast
improved or worse relative to baseline, e.g. 1.1}

\item{target_model}{Name of the target model}

\item{baseline_model}{Name of the baseline model}

\item{metric_to_compare}{Metric for which to compute
relative scores. One of \code{"wis"} or \code{"crps"}. Not case-sensitive.
Passed as the \code{metric_to_compare} argument
to \code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} via
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}
}
\value{
table of the number of states with improvements, number of overall
forecasts with improvements, number that got worse, etc.
}
\description{
Get stats on number of improved forecasts
}
