% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{get_subpop_data}
\alias{get_subpop_data}
\title{Get subpopulation data}
\usage{
get_subpop_data(add_auxiliary_site, state_pop, pop_ww, n_ww_sites)
}
\arguments{
\item{add_auxiliary_site}{Boolean indicating whether to add another
subpopulation in addition to the wastewater sites to estimate R(t) of}

\item{state_pop}{The state population size}

\item{pop_ww}{The population size in each of the wastewater sites}

\item{n_ww_sites}{The number of wastewater sites}
}
\value{
A list containing the necessary integers and vectors that stan
needs to estiamte infection dynamics for each subpopulation
}
\description{
Get subpopulation data
}
\examples{
subpop_data <- get_subpop_data(TRUE, 100000, c(1000, 500), 2)
}
