% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_summary_metadata}
\alias{get_summary_metadata}
\title{Get a summary table of the number of forecasts excluded for each reason}
\usage{
get_summary_metadata(metadata)
}
\arguments{
\item{metadata}{a tibble containing metadata for each forecast date location}
}
\value{
a 1 row tibble with the number of forecasts for each category
}
\description{
Get a summary table of the number of forecasts excluded for each reason
}
