% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{get_table_sufficient_ww}
\alias{get_table_sufficient_ww}
\title{Get table of location-forecast dates with sufficient wastewater}
\usage{
get_table_sufficient_ww(
  path_to_ww_vintaged_data,
  delay_thres = 21,
  n_dps_thres = 5,
  prop_below_lod_thres = 0.5,
  sd_thres = 0.1,
  mean_log_ww_value_thres = -4
)
}
\arguments{
\item{path_to_ww_vintaged_data}{A character string indicating the path
to where only the vintaged wastewater data are saved}

\item{delay_thres}{The maximum number of days of delay between the last
wastewater data point and the forecat date, before we would flag a state as
having insufficient wastewater data to inform a forecast. Default is 21}

\item{n_dps_thres}{The threshold number of data points within a single site
within a state before we would flag the state as having insufficient
wastewater data to inform a forecast. Default is 5}

\item{prop_below_lod_thres}{The threshold proportion of wastewater data
points that can be below the LOD. If greater than this proportion of points
are below the LOD, we flag the state as having insufficient wastewater data.
Default is 0.5}

\item{sd_thres}{The minimum standard deviation between wastewater data points
within a site. This is intended to catch when a site reports all the same
values. Default is 0.1}

\item{mean_log_ww_value_thres}{The minimum value of the log of the ww
concentration, default is -4}
}
\value{
table_of_loc_dates_w_ww a tibble containing the location,
forecast_date, and column stating the wastewater data was sufficient for all
locations and forecast dates where ww data was deemed sufficient
}
\description{
This function takes in a large dataframe containing the quantiled estimated
and forecasted wastewater concentrations for each site and lab in each
location and forecast date, joined with the observed data on each day, in
each site and lab, from each forecast date. We will use the data to
get a table of location-forecast-dates where wastewater data was considered
sufficient to inform a forecast, based on the critera that was used for
the Hub submissions.
}
