% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_data_for_stan.R
\name{get_ww_data_sizes}
\alias{get_ww_data_sizes}
\title{Get the integer sizes of the wastewater input data}
\usage{
get_ww_data_sizes(ww_data, lod_col_name = "below_LOD")
}
\arguments{
\item{ww_data}{Input wastewater dataframe containing one row
per observation, with outliers already removed}

\item{lod_col_name}{A string representing the name of the
column in the input_ww_data that provides a 0 if the data point is not above
the LOD and a 1 if the data is below the LOD, default value is \code{below_LOD}}
}
\value{
A list containing the integer sizes of the follow variables that
the stan model requires:
owt: number of wastewater observations
n_censored: number of censored wastewater observations (below the LOD)
n_uncensored: number of uncensored wastewter observations (above the LOD)
n_ww_sites: number of wastewater sites
n_ww_lab_sites: number of unique wastewater site-lab combinations
}
\description{
Get the integer sizes of the wastewater input data
}
