% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{heatmap_scores_by_loc_date}
\alias{heatmap_scores_by_loc_date}
\title{Plot a heatmap of avg forecast performance
by location and forecast date}
\usage{
heatmap_scores_by_loc_date(scores, metric, models_to_plot)
}
\arguments{
\item{scores}{A tibble of daily scores by forecast date,
location, and model}

\item{metric}{Metric to plot. One of \code{"wis"} or \code{"crps"}}

\item{models_to_plot}{Character vector of models to plot}
}
\value{
a ggplot object
}
\description{
Plot a heatmap of avg forecast performance
by location and forecast date
}
