% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{hub_average_score_table}
\alias{hub_average_score_table}
\title{Make summary table of WIS scores in Hub models overall}
\usage{
hub_average_score_table(scores)
}
\arguments{
\item{scores}{quantile based scores from the hub}
}
\value{
A table with the average scores of each model,
where model is an ordered factor.
}
\description{
Make summary table of WIS scores in Hub models overall
}
