% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_summarize_ww_data.R
\name{load_data_and_summarize}
\alias{load_data_and_summarize}
\title{Load data and summarize metadata}
\usage{
load_data_and_summarize(fp_hosp, fp_ww, this_forecast_date, this_location)
}
\arguments{
\item{fp_hosp}{Character string indicating the filepath for the hospital
admissions dataset, expected as a tsv}

\item{fp_ww}{Character string indicating the filepath for the hospital
admisisons dataset, expected as a tsv}

\item{this_forecast_date}{Character string indicating the forecast date in
ISO8601 format (YYYY-MM-DD)}

\item{this_location}{Character string indicating location, formatted as
state abbreviation}
}
\value{
this_ww_metadata: a 1 row dataframe with metadata on the wastewater
summary statistics for that forecast date and location
}
\description{
This function pulls in an individual hospital admissions and wastewater
dataset for a particular forecast date and location. It computes summary
statistics on the wastewater characteristics for that state on that
forecast date. It will return a single row dataframe with the relevant
metadata.
}
