% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{load_real_time_forecast}
\alias{load_real_time_forecast}
\title{Load a real-time forecast}
\usage{
load_real_time_forecast(
  output_dir,
  forecast_date,
  location,
  model_type,
  forecast_output_type,
  table_of_run_ids
)
}
\arguments{
\item{output_dir}{Output directory containing real-time
forecasts.}

\item{forecast_date}{Forecast date, as a string.}

\item{location}{Forecast location, as a string.}

\item{model_type}{Model type, as a string. One of
\code{"hosp"} or \code{"ww"}.}

\item{forecast_output_type}{Type of forecast output to pull.
One of \code{"quantiles"} o \code{"draws"}.}

\item{table_of_run_ids}{Table mapping forecast dates
to chosen run unique ids and run dates.}
}
\value{
The forecast, as a tibble.
}
\description{
Loads a forecast for a single location, date,
model type, and run.
}
