% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{make_baseline_score_table}
\alias{make_baseline_score_table}
\title{Make baseline score table}
\usage{
make_baseline_score_table(
  all_ww_scores,
  baseline_score_table_dir,
  overwrite_table = FALSE
)
}
\arguments{
\item{all_ww_scores}{the table of scores for all dates from all locations
from the wastewater informed model}

\item{baseline_score_table_dir}{character string indicating the directory
to save the baseline score tables}

\item{overwrite_table}{boolean indicating whether or not to overwrite the
current baseline table, default is FALSE.}
}
\value{
a table containing the summarized outputs from scoring utils for the
wastewater model across dates and locations
}
\description{
This function makes a wide table with the average score of a pipeline run
summarized across all locations. The point of this is to get an approximate
estimate of the performance of the pipeline/model across locations when we
are iterating on model development. This way, we can use this score as a
baseline and aim to add new features to the model only if the overall score
of the forecast performance is improved
}
