% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_crps_underlay_fig}
\alias{make_crps_underlay_fig}
\title{Make CRPS underlay figure}
\usage{
make_crps_underlay_fig(scores, loc_to_plot, horizon_to_plot)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
the ouput of \code{scoringutils::score()} on samples.}

\item{loc_to_plot}{A  string indicating the state abbreviations of the state
to plot}

\item{horizon_to_plot}{A string indicating what horizon period to plot,
either \code{nowcast}, \verb{1 wk}, or \verb{4 wks}}
}
\value{
A ggplot object containing a bar chart of the crps score averaged
across the horizon for each forecast date, colored by the model type
}
\description{
Make CRPS underlay figure
}
