% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_up.R
\name{make_df}
\alias{make_df}
\title{Make dataframe}
\usage{
make_df(config)
}
\arguments{
\item{config}{a config file with a list of locations, forecast_dates, and
model types}
}
\value{
an expanded dataframe that contains 3 columns with all combinations
of locations, forecast_dates and model types
}
\description{
Make dataframe
}
