% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig2.R
\name{make_fig2}
\alias{make_fig2}
\title{Make figure 2}
\usage{
make_fig2(hosp1, hosp2, hosp3, ct1, ct2, ct3, fig_file_dir)
}
\arguments{
\item{hosp1}{first hospital admissions forecast}

\item{hosp2}{second}

\item{hosp3}{third}

\item{ct1}{first faceted fit to wastewater data in each site}

\item{ct2}{second}

\item{ct3}{third}

\item{fig_file_dir}{Path to save figures}
}
\value{
a combined ggplot object
}
\description{
Make figure 2
}
