% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig2_hosp_t}
\alias{make_fig2_hosp_t}
\title{Make a plot comparing the fit and forecasted hospital admissions
from the wastewater and hospital admissions model for a few
example states}
\usage{
make_fig2_hosp_t(hosp_quantiles, locs_to_plot, date_to_plot)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated hospital admissions
data, the evaluation hospital admissions data, and the quantiles of the
calibrated and forecasted admissions}

\item{locs_to_plot}{A vector of strings indicating the state abbreviations
for which states to plot}

\item{date_to_plot}{A character string indicating what forecast date to plot}
}
\value{
a ggplot object containing a faceted vertical plot of the
forecasts produced for each state, comparing the wastewater and hospital
admissions models
}
\description{
Make a plot comparing the fit and forecasted hospital admissions
from the wastewater and hospital admissions model for a few
example states
}
