% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig3_crps_density}
\alias{make_fig3_crps_density}
\title{Make a CRPS density plot for a subset of locations}
\usage{
make_fig3_crps_density(scores)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
the ouput of \code{scoringutils::score()} on samples.}
}
\value{
a ggplot object that is a vertical facet of violin plots colored
by model type and broken down my horizon
}
\description{
Make a CRPS density plot for a subset of locations
}
