% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig3.R
\name{make_fig3_pct_better_w_ww}
\alias{make_fig3_pct_better_w_ww}
\title{Make a figure of the percent of locations with a better forecast with ww
over time}
\usage{
make_fig3_pct_better_w_ww(scores, eval_hosp_data, days_to_show_prev_data = 14)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
the ouput of \code{scoringutils::score()} on samples.}

\item{eval_hosp_data}{The retrospective hospital admissions data for each
state and for all dates, used to generate national hospital admissions}

\item{days_to_show_prev_data}{An ingeger indicating how many days before the
last forecast date to show the data, default is \code{14}}
}
\value{
a ggplot object containing a stacked bar chart of the percent of
states with imprved forecasts from wastewater by forecast date alongside
daily national hospital admissions
}
\description{
Make a figure of the percent of locations with a better forecast with ww
over time
}
