% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig3_rel_crps_by_location}
\alias{make_fig3_rel_crps_by_location}
\title{Make figure that stratifies scores by location across forecast dates}
\usage{
make_fig3_rel_crps_by_location(scores)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
the ouput of \code{scoringutils::score()} on samples.}
}
\value{
A ggplot object containing plots of the distribution of relative
CRPS scores by location, across forecast dates, colored by location
}
\description{
Make figure that stratifies scores by location across forecast dates
}
