% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4_rel_wis_by_location}
\alias{make_fig4_rel_wis_by_location}
\title{Make figure that stratifies scores by location across forecast dates}
\usage{
make_fig4_rel_wis_by_location(wis_scores)
}
\arguments{
\item{wis_scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}
}
\value{
A ggplot object containing plots of the distribution of relative
WIS scores by location, across forecast dates, colored by location
}
\description{
Make figure that stratifies scores by location across forecast dates
}
