% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4_rel_wis_over_time}
\alias{make_fig4_rel_wis_over_time}
\title{Make a CRPS density plot for a subset of locations}
\usage{
make_fig4_rel_wis_over_time(wis_scores)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}
}
\value{
a ggplot object that is a vertical facet of violin plots colored
by model type and broken down my horizon
}
\description{
Make a CRPS density plot for a subset of locations
}
