% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4_results_table}
\alias{make_fig4_results_table}
\title{Make a summary of the with and without wastewater comparison scores}
\usage{
make_fig4_results_table(scores)
}
\arguments{
\item{scores}{a tibble of scores for each model, horizon day, forecast date
and location for the subset of forecasts used in the head-to-head comparison}
}
\value{
a list of two tables with summary scores, one overall and one by
forecast vs nowcast
}
\description{
Make a summary of the with and without wastewater comparison scores
}
