% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_heatmap_rel_wis}
\alias{make_fig5_heatmap_rel_wis}
\title{Get a plot of the relative wis from the real-time models}
\usage{
make_fig5_heatmap_rel_wis(rel_scores, time_period, analysis_type, fig_file_dir)
}
\arguments{
\item{rel_scores}{tibble containing the relative wis for each forecast
date and location and horizon day}

\item{time_period}{string indicating dates of analysis, either "Feb-Mar",
or "Oct-Mar"}

\item{analysis_type}{string indicating whether analysis is Real-time}

\item{fig_file_dir}{string indicating the directory to save the figure}
}
\value{
ggplot object of a heatmap of the realtive wis
}
\description{
Get a plot of the relative wis from the real-time models
}
