% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_heatmap_relative_wis}
\alias{make_fig5_heatmap_relative_wis}
\title{Make a heatmap of relative WIS across locations}
\usage{
make_fig5_heatmap_relative_wis(
  scores,
  figure_file_path,
  time_period,
  baseline_model = "COVIDhub-baseline"
)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{figure_file_path}{path to save figure}

\item{time_period}{time period that scores are summarized over}

\item{baseline_model}{which model to compute relative WIS compared to, default
is \code{COVIDhub-baseline}}
}
\value{
a ggplot with a heatmap with model on the x-axis, location on the y-axis
and fill by relative WIS score across forecast dates and horizons
}
\description{
Make a heatmap of relative WIS across locations
}
