% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig_phase_comparison}
\alias{make_fig_phase_comparison}
\title{Make phase comparison figure}
\usage{
make_fig_phase_comparison(hosp_quantiles, loc_to_plot)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated hospital admissions
data, the evaluation hospital admissions data, and the quantiles of the
calibrated and forecasted admissions}

\item{loc_to_plot}{A  string indicating the state abbreviations of the state
to plot}
}
\value{
a ggplot object containing the hospital admissions for evaluation
colored by epidemic phase, for visual comparison to the trend in the data
}
\description{
Make phase comparison figure
}
