% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_hosp_forecast_comp_fig}
\alias{make_hosp_forecast_comp_fig}
\title{Make hospital forecast comparison figure}
\usage{
make_hosp_forecast_comp_fig(
  hosp_quantiles,
  loc_to_plot,
  horizon_to_plot,
  days_to_show_prev_data = 14
)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated hospital admissions
data, the evaluation hospital admissions data, and the quantiles of the
calibrated and forecasted admissions}

\item{loc_to_plot}{A  string indicating the state abbreviations of the state
to plot}

\item{horizon_to_plot}{A string indicating what horizon period to plot,
one of \code{nowcast}, \verb{1 wk}, or \verb{4 wks}}

\item{days_to_show_prev_data}{An ingeger indicating how many days before the
last forecast date to show the data, default is \code{14}}
}
\value{
A ggplot object containing a plot of the retrospective hospital
admissions data compared to the nowcasted/forecasted quantiles and median
for the specified horizon to plot, colored by the model type
}
\description{
Make hospital forecast comparison figure
}
