% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{make_hospital_onset_delay_pmf}
\alias{make_hospital_onset_delay_pmf}
\title{Make hospital onset delay pmf}
\usage{
make_hospital_onset_delay_pmf(
  neg_binom_mu = 6.98665,
  neg_binom_size = 2.490848
)
}
\arguments{
\item{neg_binom_mu}{NEEDS DOCUMENTATION.}

\item{neg_binom_size}{NEEDS DOCUMENTATION.}
}
\value{
pmf of distribution from symptom onset to hospital admission
}
\description{
Uses the parameter estimates from cfa-parameter-estimates,
which is based on Danache et al linelist data from symptom onset to hospital
admission. See below:
https://journals.plos.org/plosone/article?id=10.1371/journal.pone.0261428
}
