% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delay_distribs.R
\name{make_incubation_period_pmf}
\alias{make_incubation_period_pmf}
\title{Make incubation period pmf}
\usage{
make_incubation_period_pmf(
  backward_scale = 3.6,
  backward_shape = 1.5,
  r = 0.15
)
}
\arguments{
\item{backward_scale}{NEEDS DOCUMENTATION.}

\item{backward_shape}{NEEDS DOCUMENTATION.}

\item{r}{NEEDS DOCUMENTATION.}
}
\value{
pmf of incubation period
}
\description{
This uses NNH's logic and workflow to make a pmf corresponding to
the incubation period for COVID after Omicron used in Park et al 2023
(which we also use for the
lognormal generation interval). These estimates are from early Omicron.
}
