% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_plot_coverage_range}
\alias{make_plot_coverage_range}
\title{Plot coverage at specified ranges}
\usage{
make_plot_coverage_range(
  scores_quantiles,
  ranges,
  fig_file_dir = NULL,
  write_files = FALSE
)
}
\arguments{
\item{scores_quantiles}{A tibble of scores by location, forecast date,
date and model, containing the outputs of \code{scoringutils::score()} on
quantiles plus metadata transformed into a tibble.}

\item{ranges}{A numeric vector of credible interval ranges to plot,
spanning from 0 to 100.}

\item{fig_file_dir}{string indicating directory to save figure in,
default is \code{NULL}}

\item{write_files}{boolean indicating whether to save file, default is FALSE}
}
\value{
A ggplot2 object
}
\description{
Plot coverage at specified ranges
}
