% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_sfig_crps_by_phase}
\alias{make_sfig_crps_by_phase}
\title{Make figure that plots distribution of crps scores stratified by
epidemic phase}
\usage{
make_sfig_crps_by_phase(scores)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}
}
\value{
A ggplot object containing plots of the distribution of
CRPS scores, colored by model, stratified by epidemic phase, across
locations and forecast dates
}
\description{
Make figure that plots distribution of crps scores stratified by
epidemic phase
}
