% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_example_scores.R
\name{multi_location_crps_figure}
\alias{multi_location_crps_figure}
\title{Generate bar plots of CRPS for each model in different locations}
\usage{
multi_location_crps_figure(scores, locs_to_plot, fig_file_dir)
}
\arguments{
\item{scores}{tibble of crps scores by location, forecast date, model,
horizon day}

\item{locs_to_plot}{Vector of strings indicating the locations to plot,
as two-letter USPS abbreviations.}

\item{fig_file_dir}{string indicating directory to save fig in}
}
\value{
Figure showing CRPS for multiple locations.
}
\description{
Generate bar plots of CRPS for each model in different locations
}
