% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_predicted_actual_examples.R
\name{multi_location_pred_actual_fig}
\alias{multi_location_pred_actual_fig}
\title{Make figure 2}
\usage{
multi_location_pred_actual_fig(
  hosp1,
  hosp2,
  hosp3,
  ww1,
  ww2,
  ww3,
  fig_file_dir
)
}
\arguments{
\item{hosp1}{first hospital admissions predicted-actual figure}

\item{hosp2}{second hospital admissions predicted-actual figure}

\item{hosp3}{third hospital admissions predicted-actual figure}

\item{ww1}{first wastewater predicted-actual figure}

\item{ww2}{second wastewater predicted-actual figure}

\item{ww3}{third wastewater predicted-actual figure}

\item{fig_file_dir}{Path to save figures}
}
\value{
a combined ggplot object
}
\description{
Make figure 2
}
