% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{order_col}
\alias{order_col}
\title{Light wrapper function for converting a dataframe column
to an ordered factors.}
\usage{
order_col(df, col, levels = NULL)
}
\arguments{
\item{df}{Data frame to transform}

\item{col}{column to transform}

\item{levels}{levels for the column, in ascending order.
If \code{NULL} (default) use the current ordering in the column.}
}
\value{
A copy of the data frame with the \code{col} column transformed
into an ordered factor with levels given by \code{levels}.
}
\description{
Wraps \code{\link[dplyr:mutate]{dplyr::mutate()}}. Useful for pipe chains.
}
\examples{

df <- tibble::tibble(
  x = c("b", "c", "a", "c", "b", "a", "a"),
  y = rnorm(7),
  z = 5
)

new_df <- df |>
  order_col("x", c("c", "b", "a")) |>
  dplyr::select("x", "z") |>
  dplyr::arrange(x)

new_df
}
