% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{order_horizons}
\alias{order_horizons}
\title{Get the order of the categorical horizons for plotting}
\usage{
order_horizons(df)
}
\arguments{
\item{df}{a dataframe containing the column name \code{horizon} which will
contain character strings indicating the horizon (by week, nowcast period,
or overall) of the score/quantile/sample.}
}
\value{
a dataframe containing the same columns as \code{df} but with the
addition of \code{fig_order} and reordered in terms of \code{fig_order} for plotting.
}
\description{
Get the order of the categorical horizons for plotting
}
