% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{order_periods}
\alias{order_periods}
\title{Get the order of the time periods for plotting}
\usage{
order_periods(df)
}
\arguments{
\item{df}{a dataframe containing the column name \code{period} which will
contain character strings indicating what period the scores correspond to}
}
\value{
a dataframe containing the same columns as \code{df} but now ordered
in terms of the specified order.
}
\description{
Get the order of the time periods for plotting
}
