% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchmarking.R
\name{plot_benchmarks}
\alias{plot_benchmarks}
\title{Plot benchmark summaries}
\usage{
plot_benchmarks(
  grouping_var,
  benchmark_scope,
  benchmark_dir,
  score_to_plot = "crps",
  write_files = TRUE
)
}
\arguments{
\item{grouping_var}{The variable to plot by, so in this case either
\code{location} or \code{forecast_date}}

\item{benchmark_scope}{The scope of the benchmarking (so in this case
either \code{subset_forecasts} or \code{all_forecasts})}

\item{benchmark_dir}{The directory where the benchmark tables live}

\item{score_to_plot}{Which of the scores saved in the benchmarking
tables to plot, options are \code{crps}, \code{bias}, and \code{ae}, defualt is \code{crps}}

\item{write_files}{Boolean indicating whether or not to save the plots
to disk, default is TRUE}
}
\value{
a ggplot object containing a bar chart colored by package/model
version and faceted by the grouping variable, with the height of the bar
indicating the scores
}
\description{
Plot benchmark summaries
}
