% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bias.R
\name{plot_bias_t}
\alias{plot_bias_t}
\title{Get a plot of bias over time}
\usage{
plot_bias_t(scores)
}
\arguments{
\item{scores}{a complete df of scores with a column for bias}
}
\value{
a plot of bias over time averaged across locations and
forecast dates, separated by model
}
\description{
Get a plot of bias over time
}
