% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_bootstrapped_scores.R
\name{plot_bootstrapped_score_values}
\alias{plot_bootstrapped_score_values}
\title{Plot bootstrapped absolute CRPS values as pointintervals.}
\usage{
plot_bootstrapped_score_values(replicates, point_estimates)
}
\arguments{
\item{replicates}{Data frame of bootstrapped replicates,
as the output of \code{\link[=bootstrap_crps_values]{bootstrap_crps_values()}}.}

\item{point_estimates}{Data frame of point estimates,
with columns \code{"model"} and \code{"crps"}.}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot bootstrapped absolute CRPS values as pointintervals.
}
