% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{plot_crps_t}
\alias{plot_crps_t}
\title{Plot average CRPS over time for model comparison}
\usage{
plot_crps_t(scores, horizon_time_in_weeks = NULL)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}

\item{horizon_time_in_weeks}{horizon time in weeks to summarize over, default
is \code{NULL} which means that the scores are summarized over the nowcast period
and the 4 week forecast period}
}
\value{
a ggplot object plotting the magnitude of the avg crps across
locations at each forecast date
}
\description{
Plot average CRPS over time for model comparison
}
