% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_example_scores.R
\name{plot_crps_underlay}
\alias{plot_crps_underlay}
\title{Make CRPS underlay figure}
\usage{
plot_crps_underlay(
  scores,
  loc_to_plot,
  horizon_to_plot,
  horizon_days_ahead,
  days_to_shift = 0
)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}

\item{loc_to_plot}{A  string indicating the state abbreviations of the state
to plot}

\item{horizon_to_plot}{A string indicating what horizon period to plot,
one of \code{nowcast}, \verb{1 wk}, or \verb{4 wks}}

\item{horizon_days_ahead}{An integer corresponding to the horizon days to
plot}

\item{days_to_shift}{An integer corresponding to the number of days to shift
the x axis of the underly plot to line up with the corresponding forecast
horizon, default is 0}
}
\value{
A ggplot object containing a bar chart of the crps score averaged
across the horizon for each forecast date, colored by the model type
}
\description{
Make CRPS underlay figure
}
