% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_example_scores.R
\name{plot_forecast_comparison_t}
\alias{plot_forecast_comparison_t}
\title{Compare two hospital admissions forecasts over time.}
\usage{
plot_forecast_comparison_t(
  hosp_quantiles,
  loc_to_plot,
  horizon_to_plot,
  horizon_days_ahead,
  days_to_show_prev_data = 14
)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated hospital
admissions data, the evaluation hospital admissions data, and
the quantiles of the calibrated and forecasted admissions}

\item{loc_to_plot}{A string indicating the location to plot, as a
two-letter USPS abbreviation.}

\item{horizon_to_plot}{A string indicating what horizon period to plot,
one of \code{nowcast}, \verb{1 wk}, or \verb{4 wks}}

\item{horizon_days_ahead}{An integer corresponding to the horizon days to
plot}

\item{days_to_show_prev_data}{An ingeger indicating how many days before the
last forecast date to show the data, default is \code{14}}
}
\value{
A ggplot object containing a plot of the retrospective hospital
admissions data compared to the nowcasted/forecasted quantiles and median
for the specified horizon to plot, colored by the model type
}
\description{
Compare two hospital admissions forecasts over time.
}
