% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_heatmap_metadata_hub}
\alias{plot_heatmap_metadata_hub}
\title{Plot a heatmap of the metadata of Hub models submitted}
\usage{
plot_heatmap_metadata_hub(metadata, analysis_type)
}
\arguments{
\item{metadata}{a tibble of location -forecast date metadata}

\item{analysis_type}{string indicating whether this is the
real-time or retro analysis, which dictates how metadata is gathered}
}
\value{
a ggplot object with a heatmap colored by reason for excluding
}
\description{
Plot a heatmap of the metadata of Hub models submitted
}
