% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_heatmap_metadata_hub_real_time}
\alias{plot_heatmap_metadata_hub_real_time}
\title{Plot model submission decisions in real-time}
\usage{
plot_heatmap_metadata_hub_real_time(dates, locations, hosp_substitution_table)
}
\arguments{
\item{dates}{forecast dates to plot}

\item{locations}{locations to plot}

\item{hosp_subtitution_table}{table of dates and locations
for which the hospital admissions-only model was substituted in
real time.}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot model submission decisions in real-time
}
