% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_heatmap_metadata_retro}
\alias{plot_heatmap_metadata_retro}
\title{Plot a heatmap of the metadata of reasons for excluding
forecasts from analysis}
\usage{
plot_heatmap_metadata_retro(metadata)
}
\arguments{
\item{metadata}{a tibble of location -forecast date metadata}
}
\value{
a ggplot object with a heatmap colored by reason for excluding
}
\description{
Plot a heatmap of the metadata of reasons for excluding
forecasts from analysis
}
