% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{plot_heatmap_relative_wis}
\alias{plot_heatmap_relative_wis}
\title{Make a heatmap of relative WIS across locations}
\usage{
plot_heatmap_relative_wis(
  scores,
  time_period,
  models_to_show,
  baseline_model = "COVIDhub-4_week_ensemble"
)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{time_period}{time period that scores are summarized over}

\item{models_to_show}{A vector of charcter strings indicating which models
from the COVID-19 forecast hub to include in the plot.}

\item{baseline_model}{which model to compute relative WIS compared to, default
is \verb{COVIDhub-4_week_ensemble}}
}
\value{
a ggplot with a heatmap with model on the x-axis, location on the y-axis
and fill by relative WIS score across forecast dates and horizons
}
\description{
Make a heatmap of relative WIS across locations
}
