% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_min_ess_distribution}
\alias{plot_min_ess_distribution}
\title{Plot a distribution of minimum ESS values.}
\usage{
plot_min_ess_distribution(diagnostic_extrema, which)
}
\arguments{
\item{diagnostic_extrema}{data frame of diagnostic extrema,
as the output of \code{\link[=extract_diagnostic_extrema]{extract_diagnostic_extrema()}}, collated via
\code{\link[=combine_outputs]{combine_outputs()}}.}

\item{which}{Plot bulk ESS or tail ESS? Must be one of \code{"bulk"}
or \code{"tail"}.}
}
\value{
The plot.
}
\description{
Plot a distribution of minimum ESS values.
}
