% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/figure_predicted_actual_examples.R
\name{plot_pred_actual_ww}
\alias{plot_pred_actual_ww}
\title{Make concentration fit and forecast figure}
\usage{
plot_pred_actual_ww(
  ww_quantiles,
  loc_to_plot,
  date_to_plot,
  n_forecast_days = 28,
  n_calib_days = 90,
  max_n_site_labs_to_show = 3,
  site_lab_names_to_show = NULL
)
}
\arguments{
\item{ww_quantiles}{A tibble containing the calibrated wastewater
concentrations, the evaluation wastewater concentration data, and the
quantiles of the calibrated and forecasted wastewater concentrations}

\item{loc_to_plot}{A character string indicating the state abbreviation
for which state to plot, can only be one state}

\item{date_to_plot}{A character string indicating what forecast date to plot,
in IS08601 format YYYY-MM-DD}

\item{n_forecast_days}{An integer indicating the number of days to show the
forecast for, default is \code{28}}

\item{n_calib_days}{An integer indicating the number of days to show the
calibration data for, default is \code{90}}

\item{max_n_site_labs_to_show}{An integer indicating the maximum number
of site-labs to show in the figure, default is \code{3}}

\item{site_lab_names_to_show}{a vector of character strings indicating
the site lab names to be displayed in the plot. If NULL, the first
\code{max_n_site_labs_to_show} or all are displayed. Default is \code{NULL}.}
}
\value{
A ggplot object containing a faceted horizontal plot of the
calibrated and forecasted wastewater concentrations for 3 or fewer
site-lab combinations for a single state
}
\description{
Make concentration fit and forecast figure
}
