% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_posterior.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{Plot posterior draws for a given parameter against
a known marginal prior distribution.}
\usage{
plot_prior_posterior(
  draws_long,
  prior_params,
  fit_id_col = "forecast_date",
  model_type_col = "model_type",
  variable_name = ".variable",
  value_name = ".value",
  row_facet = "location",
  custom_x_transform = list(),
  custom_xlim = list()
)
}
\arguments{
\item{draws_long}{long-format posterior draws, as the
output of \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}} with auxiliary metadata
columns as below.}

\item{prior_params}{Named list of prior hyperparameter values, as the
output of \code{\link[wwinference:get_params]{wwinference::get_params()}}}

\item{fit_id_col}{Name of a column in \code{draws_long} that uniquely
identifies individual model fits within a facet. Default \code{"forecast_date"}.}

\item{model_type_col}{Name of a column in \code{draws_long} that
identifies model types, so that \code{\link[=scale_fill_model]{scale_fill_model()}} can be applied.
Default \code{"model_type"}.}

\item{variable_name}{Name for the variable column.
Default \code{".variable"}, matching  \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}}.}

\item{value_name}{Name for the value column.
Default \code{".value"}, matching  \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}}.}

\item{row_facet}{Column to facet rows by. Default \code{"location"}.}

\item{custom_x_transform}{Optional named list mapping parameter names
to custom x transforms to use for that parameter's panels in the plot.
If a parameter name is not matched, the identity transform will be used.
Default \code{list()}, which implies using the identity transform for all
panels.}

\item{custom_xlim}{Optional named list mapping parameter names
to vectors specificing custom x limits to use for that parameter's
panels in the plot. If a parameter name is not matched, x limits will
be deferred to ggplot. Default \code{list()}, which implies deferring limits
for all parameters.}
}
\value{
The plot.
}
\description{
Plot posterior draws for a given parameter against
a known marginal prior distribution.
}
