% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools.R
\name{plot_quantiles}
\alias{plot_quantiles}
\title{Plot a timeseries of quantiles}
\usage{
plot_quantiles(
  data,
  time_column,
  observation_column,
  quantile_level_column,
  linesize = 2,
  pointsize = 4,
  pointcolor = "darkblue",
  linecolor = "darkblue"
)
}
\arguments{
\item{data}{timeseries of quantiles as tidy data,
with one row per timepoint per quantile level}

\item{time_column}{name of the column in \code{data}
containing timepoints}

\item{observation_column}{name of the column
in data containing observed values at the
given quantile levels}

\item{quantile_level_column}{name of the column
in \code{data} giving the quantile level (e.g.
\code{0.01} for the 0.01 quantile / 1st percentile)}

\item{linesize}{\code{size} parameter passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.
Default 2.}

\item{pointsize}{\code{size} parameter passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
Default 4.}

\item{pointcolor}{\code{color} parameter passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}
Default "darkblue".}

\item{linecolor}{\code{color} parameter passed to \code{\link[ggplot2:geom_path]{ggplot2::geom_line()}}.
Default "darkblue.}
}
\value{
the resultant plot, as a ggplot objec
}
\description{
Plot a timeseries of quantiles
}
