% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_example_scores.R
\name{plot_rel_crps_by_horizon}
\alias{plot_rel_crps_by_horizon}
\title{Make head to head CRPS distribution comparison
plot for a single location}
\usage{
plot_rel_crps_by_horizon(
  scores,
  loc_to_plot,
  baseline = "cfa-hosponlyrenewal(retro)",
  horizons_to_show = c("nowcast", "1 wk", "4 wks", "overall")
)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date,
date and model,
containing the outputs of \code{scoringutils::score()} on samples
plus metadata
transformed into a tibble.}

\item{loc_to_plot}{A string indicating the location to plot, as a
two-letter USPS abbreviation.}

\item{baseline}{Name of the model to use as a baseline. Default
\code{"cfa-hosponlyrenewal(retro)"}}

\item{horizons_to_show}{A vector of strings indicating the names of the
\code{horizon} that we want to show on the plot, must be a subset of
\code{nowcast}, \verb{1 wk}, \verb{2 wks},\verb{3 wks}, \verb{4 wks} and \code{overall}}
}
\value{
a ggplot object containing violin plots comparing the
distribution of crps across forecast dates for a single location,
grouped by horizon and colored by model
}
\description{
Make head to head CRPS distribution comparison
plot for a single location
}
