% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{plot_rel_crps_by_location}
\alias{plot_rel_crps_by_location}
\title{Make figure that stratifies scores by location across forecast dates}
\usage{
plot_rel_crps_by_location(scores, target_model, baseline_model)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}

\item{target_model}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS}
}
\value{
A ggplot object containing plots of the distribution of relative
CRPS scores by location, across forecast dates, colored by location
}
\description{
Make figure that stratifies scores by location across forecast dates
}
