% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{plot_rel_crps_distribution}
\alias{plot_rel_crps_distribution}
\title{Get a dotsinterval plot of the relative CRPS distribution
for individual location/forecast-date forecast problems.}
\usage{
plot_rel_crps_distribution(scores, target_model, baseline_model, fig_file_dir)
}
\arguments{
\item{scores}{table of scores by horizon day,
forecast date, and location}

\item{target_model}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS}

\item{fig_file_dir}{directory to save figure in}
}
\value{
ggplot object of distribution of relative CRPS scores
}
\description{
Get a dotsinterval plot of the relative CRPS distribution
for individual location/forecast-date forecast problems.
}
