% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{plot_rel_crps_distribution_t}
\alias{plot_rel_crps_distribution_t}
\title{Make a CRPS dotsinterval plot stratified by time}
\usage{
plot_rel_crps_distribution_t(scores, target_model, baseline_model)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date,
date and model,
containing the outputs of \code{scoringutils::score()} on samples,
plus metadata, transformed into a tibble.}

\item{target_model}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS
computation.}
}
\value{
a ggplot object with the distributions plotted by
forecast date.
}
\description{
Make a CRPS dotsinterval plot stratified by time
}
