% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{plot_rel_crps_heatmap}
\alias{plot_rel_crps_heatmap}
\title{Plot a heatmap of the relative crps by locations and forecast date
for the head-to-head comparison}
\usage{
plot_rel_crps_heatmap(scores, target_model, baseline_model, fig_file_dir)
}
\arguments{
\item{scores}{A tibble of daily scores by forecast date, location,
and model}

\item{target_model}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS}

\item{fig_file_dir}{A string indicating the directory to save
the figures in}
}
\value{
a ggplot object
}
\description{
Plot a heatmap of the relative crps by locations and forecast date
for the head-to-head comparison
}
