% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{plot_rel_score_distribution}
\alias{plot_rel_score_distribution}
\title{Get a dotsinterval plot of a relative score distribution
for individual location/forecast-date forecast problems.}
\usage{
plot_rel_score_distribution(
  scores,
  target_model,
  baseline_model,
  metric_to_compare
)
}
\arguments{
\item{scores}{table of scores by horizon day,
forecast date, and location}

\item{target_model}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS}

\item{metric_to_compare}{Metric for which to compute the
relative score. One of \code{"wis"} or \code{"crps"}. Passed as the
\code{metric_to_compare} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} via
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}
}
\value{
ggplot object of distribution of relative CRPS scores
}
\description{
Get a dotsinterval plot of a relative score distribution
for individual location/forecast-date forecast problems.
}
