% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{plot_rel_score_dists_by_horizon}
\alias{plot_rel_score_dists_by_horizon}
\title{Make figure that stratifies across location and forecast dates}
\usage{
plot_rel_score_dists_by_horizon(
  scores,
  target_model,
  baseline_model,
  metric_to_compare,
  horizons_to_show = c("nowcast", "1 wk", "4 wks", "overall")
)
}
\arguments{
\item{scores}{output of \code{\link[scoringutils:score]{scoringutils::score()}}.}

\item{target_model}{Model for which to plot relative scores}

\item{baseline_model}{Baseline model for the relative scores}

\item{metric_to_compare}{Metric for which to compute the
relative score. One of \code{"wis"} or \code{"crps"}. Passed as the
\code{metric_to_compare} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} via
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}

\item{horizons_to_show}{A vector of strings indicating the names of the
\code{horizon} that we want to show on the plot, must be a subset of
\code{nowcast}, \verb{1 wk}, \verb{2 wks},\verb{3 wks}, \verb{4 wks} and \code{overall}}
}
\value{
A ggplot object containing plots of the distribution
of relative scores across location and forecast dates
}
\description{
Make figure that stratifies across location and forecast dates
}
