% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{plot_rel_score_heatmap}
\alias{plot_rel_score_heatmap}
\title{Plot a heatmap of the relative crps by locations and forecast date
for the head-to-head comparison}
\usage{
plot_rel_score_heatmap(
  scores,
  target_models,
  baseline_model,
  metric_to_compare
)
}
\arguments{
\item{scores}{A tibble of daily scores by forecast date, location,
and model}

\item{target_models}{Model for which to plot relative CRPS}

\item{baseline_model}{Baseline model for the relative CRPS}

\item{metric_to_compare}{Metric for which to compute the
relative score. One of \code{"wis"} or \code{"crps"}. Passed as the
\code{metric_to_compare} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} via
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}
}
\value{
a ggplot object
}
\description{
Plot a heatmap of the relative crps by locations and forecast date
for the head-to-head comparison
}
