% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{plot_rel_score_t}
\alias{plot_rel_score_t}
\title{Plot a timeseries of relative scores versus time}
\usage{
plot_rel_score_t(
  scores,
  target_model,
  baseline_model,
  metric_to_compare,
  x = "forecast_date",
  by = NULL,
  retain_baseline = FALSE,
  model_z_order = NULL
)
}
\arguments{
\item{scores}{Output of \code{\link[scoringutils:score]{scoringutils::score()}}, not yet summarized.}

\item{target_model}{Model for which to plot relative scores}

\item{baseline_model}{Baseline model for the relative scores
computation.}

\item{metric_to_compare}{Metric for which to plot relative
scores. One of \code{"wis"} or \code{"crps"}. Not case-sensitive.}

\item{x}{Name of the column defining the x axis (typically time)
Default \code{"forecast_date"}.}

\item{by}{additional variables defining a single score to plot.
Passed along with the variable defined by \code{x} as the \code{by} argument to
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}. Default \code{NULL}
(do not summarize over additional variables).}

\item{retain_baseline}{Retain the baseline model (with its
relative scores of 1) alongside the target models(s)? Default
\code{FALSE}.}

\item{model_z_order}{z-order in which to overplot the individual models,
ascending (so the last named model is plotted on top). If \code{NULL} (default),
plot the the models in order of overall score, so that the lowest (best) scoring
models are on top.}
}
\value{
a ggplot object with the distributions plotted by
forecast date.
}
\description{
Plot a timeseries of relative scores versus time
}
