% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictive.R
\name{plot_ribbon_hosp_quantiles}
\alias{plot_ribbon_hosp_quantiles}
\title{Make a plot comparing the fit and forecasted hospital admissions
from the wastewater-informed and hospital admissions-only models.}
\usage{
plot_ribbon_hosp_quantiles(
  hosp_quantiles,
  loc_to_plot,
  date_to_plot,
  n_forecast_days = 28,
  n_calib_days = 90
)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated
hospital admissions data, the evaluation hospital
admissions data, and the quantiles of the calibrated
and forecasted admissions, for either or both the
wastewater model and the hospital admissions-only model.}

\item{loc_to_plot}{A  string indicating the state abbreviation
of the state to plot.}

\item{date_to_plot}{A character string indicating what
forecast date to plot in IS08601 format (YYYY-MM-DD).}

\item{n_forecast_days}{An integer indicating the number of days
to show the forecast for, default is \code{28}.}

\item{n_calib_days}{An integer indicating the number of days
to show the calibration data for, default is \code{90}.}
}
\value{
a ggplot object with all model forecasts plotted against
observed data and models differentiated by fill/line color.
}
\description{
Make a plot comparing the fit and forecasted hospital admissions
from the wastewater-informed and hospital admissions-only models.
}
