% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{plot_score_by_horizon_t}
\alias{plot_score_by_horizon_t}
\title{Plot scores by horizon over time}
\usage{
plot_score_by_horizon_t(
  scores,
  score_type,
  score_display_name = toupper(score_type)
)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{score_type}{Score to display, as its column name in \code{scores}.}

\item{score_display_name}{Display name for the score. If not specified,
use the value of \code{score_type} in uppercase.}
}
\value{
plot of scores over time faceted by horizon
}
\description{
Plot scores by horizon over time
}
