% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{plot_score_model_loc}
\alias{plot_score_model_loc}
\title{Generate bar plots of probabilitistic score for each model in
different locations.}
\usage{
plot_score_model_loc(scores, locs_to_plot)
}
\arguments{
\item{scores}{tibble of crps scores by location, forecast date, model,
horizon day}

\item{locs_to_plot}{Vector of strings indicating the locations to plot,
as two-letter USPS abbreviations.}
}
\value{
Figure showing CRPS for multiple locations.
}
\description{
Light wrapper of \code{\link[=plot_score_decomposed_bars]{plot_score_decomposed_bars()}}.
}
