% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictive.R
\name{plot_spaghetti_hosp_draws}
\alias{plot_spaghetti_hosp_draws}
\title{Plots for visualizing postperior predictive draws or quantiles
compared to observed data.
Spaghetti plot of hospital admissions data compared to model draws}
\usage{
plot_spaghetti_hosp_draws(
  draws_w_data,
  location,
  model_type = c("ww", "hosp"),
  n_draws = 100
)
}
\arguments{
\item{draws_w_data}{A long tidy dataframe containing draws from the
model of the estimated hospital admissions joined with both the data
the model was calibrated to and the later observed data.}

\item{location}{the jursidiction the data is from}

\item{model_type}{type of model the output is from, options are
"ww" or "hosp"}

\item{n_draws}{number of draws to plot, default = 100}
}
\value{
a ggplot object showing posterior predictive draws for
hospital admissions alongside the calibration and evaluation data
}
\description{
Plots for visualizing postperior predictive draws or quantiles
compared to observed data.
Spaghetti plot of hospital admissions data compared to model draws
}
