% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictive.R
\name{plot_spaghetti_ww_draws}
\alias{plot_spaghetti_ww_draws}
\title{Get plot of wastewater data compared to model posterior predictive
draws}
\usage{
plot_spaghetti_ww_draws(
  draws_w_data,
  location,
  model_type = "ww",
  n_draws = 100
)
}
\arguments{
\item{draws_w_data}{A long tidy dataframe containing posterior
predictivedraws from the for wastewater concentration joined to
both the calibration and evaluation data.}

\item{location}{the jurisdiction the data is from}

\item{model_type}{type of model the output is from, default is \code{ww}}

\item{n_draws}{number of draws to plot, default = 100}
}
\value{
a ggplot object faceted by site showing the draws
}
\description{
Get plot of wastewater data compared to model posterior predictive
draws
}
