% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_submit_info_retro}
\alias{plot_submit_info_retro}
\title{Plot a heatmap of the metadata of Hub models "submitted" in the
retrospective analysis}
\usage{
plot_submit_info_retro(metadata)
}
\arguments{
\item{metadata}{a tibble of location and forecast date metadata}
}
\value{
The plot.
}
\description{
Plot a heatmap of the metadata of Hub models "submitted" in the
retrospective analysis
}
