% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_total_admissions}
\alias{plot_total_admissions}
\title{Make a figure of overall admissions (summed across locations) for context}
\usage{
plot_total_admissions(eval_hosp_data, first_forecast_date, last_forecast_date)
}
\arguments{
\item{eval_hosp_data}{Hospital admissions data for evaluating against
for all locations}

\item{first_forecast_date}{The first forecast date we are evaluating}

\item{last_forecast_date}{The last forecast date we are evaluating}
}
\value{
ggplot object displaying a timeseries of total
hospital admissions.
}
\description{
Make a figure of overall admissions (summed across locations) for context
}
