% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_recent_trend_fits.R
\name{process_recent_trend_fits}
\alias{process_recent_trend_fits}
\title{Process BRMS log-linear fits of recent trends}
\usage{
process_recent_trend_fits(
  hosp_fit,
  ww_fit,
  forecast_date,
  location,
  scenario,
  processed_output_dir,
  figure_ext = "pdf",
  n_lab_sites_plot = 10,
  max_rhat = 1.02,
  min_neff_ratio = 0.05,
  max_frac_divergent = 0.05
)
}
\arguments{
\item{hosp_fit}{\link[brms:brmsfit-class]{brms::brmsfit} object for the hospital admissions trend.}

\item{ww_fit}{\link[brms:brmsfit-class]{brms::brmsfit} object for the wastewater trend.}

\item{forecast_date}{forecast_date for the trend fits.}

\item{location}{Location for the trend fits.}

\item{scenario}{Wastewater data availability scenario for the
trend fits.}

\item{processed_output_dir}{Base directory for processed
output (processed output will be saved in a subdirectory specific
to the \code{forecast_date}, \code{location} and \code{scenario}. See
\code{\link[=forecast_output_path]{forecast_output_path()}}.}

\item{figure_ext}{File extension for figures, without the \code{.},
e.g. \code{"pdf"} or \code{"png"}. Default \code{"pdf"}.}

\item{n_lab_sites_plot}{Maximum number of lab-sites for which to
plot wastewater trends. Default \code{10}.}

\item{max_rhat}{maximum allowable r-hat value. Default 1.01}

\item{min_neff_ratio}{maximum allowable ratio of effective sample size
to nominal sample size. Default 0.1.}

\item{max_frac_divergent}{maximum allowable fraction of divergent
transitions. Default 0.01 (1\%).}
}
\value{
NULL, invisibly saving postprocessing results to
disk as a side effect.
}
\description{
Process BRMS log-linear fits of recent trends
}
