% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_postprocess.R
\name{process_successful_fit}
\alias{process_successful_fit}
\title{Postprocess a successful eval fitting job.}
\usage{
process_successful_fit(
  wwinference_fit_obj,
  stan_fit_obj,
  model,
  location,
  forecast_date,
  scenario,
  output_dir,
  raw_output_dir,
  input_hosp_data_wweval,
  input_ww_data_wweval,
  eval_hosp_data,
  eval_ww_data
)
}
\arguments{
\item{wwinference_fit_obj}{Fit object to process, as the output of
\code{\link[wwinference:wwinference]{wwinference::wwinference()}}.}

\item{stan_fit_obj}{Corresponding stan fit object.}

\item{model}{Which model the object represents. One of \code{"ww"} and \code{"hosp"}.}

\item{location}{location for the forecasting problem.}

\item{forecast_date}{"as-of" date for the forecasting problem.}

\item{scenario}{data availability scenario for the forecasting problem.}

\item{output_dir}{Directory in which to save processed output.}

\item{raw_output_dir}{Directory in which to archive objects from the environment
as serialized \code{.rds} files.}

\item{input_hosp_data_wweval}{Input hospital admissions data in legacy wweval format.}

\item{input_ww_data_wweval}{Input wastewater data in legacy \code{wweval} format.}

\item{eval_hosp_data}{Evaluation hospital admissions data in newer \code{wwinference} format.}

\item{eval_ww_data}{Evaluation wastewater admissions data in newer \code{wwinference} format.}
}
\value{
Nothing, saving results to disk as a side effect.
}
\description{
Postprocess a successful eval fitting job.
}
