% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{pull_hub_forecasts}
\alias{pull_hub_forecasts}
\title{Pull hub submissions and create a scorable table
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.}
\usage{
pull_hub_forecasts(
  model_name,
  dates,
  eval_data,
  locations = NULL,
  hub_subdir = NA,
  pull_from_github = TRUE,
  submissions_path = NULL
)
}
\arguments{
\item{model_name}{a vector of character strings indicating
the names of the models to score.}

\item{dates}{a vector of dates indicating the dates of the
submissions to score.}

\item{eval_data}{Table of evaluation data, as the output of
\code{\link[=get_input_hosp_data]{get_input_hosp_data()}}}

\item{locations}{a vector of character strings indicating the locations
to score}

\item{hub_subdir}{path where the retrospective hub submissions are saved
locally since these are not on COVIDHub github}

\item{pull_from_github}{boolean indicating whether or not to pull
from github}

\item{submissions_path}{url pointing to the "data-processed" folder on
the COVIDhub github, which is where team's submissions are located.
If not provided, use the original COVIDHub on github.}
}
\value{
a dataframe containing all of the scores for all models,
forecast dates (indicated by dates), locations, target end dates, and
quantiles
}
\description{
Pull hub submissions and create a scorable table
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
