% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{qq_plot_by_model}
\alias{qq_plot_by_model}
\title{Make a quantile quantile plot for the Hub}
\usage{
qq_plot_by_model(scores, models_to_show, time_period)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{models_to_show}{A vector of charcter strings indicating which models
from the COVID-19 forecast hub to include in the plot.}

\item{time_period}{time period that scores are summarized over}
}
\value{
a ggplot object containing a plot of the proportion of data within
each interval for each model.
}
\description{
Make a quantile quantile plot for the Hub
}
