% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{query_and_select_models}
\alias{query_and_select_models}
\title{Query Zoltar for models to include in the analysis}
\usage{
query_and_select_models(
  prop_dates_for_incl_hub,
  prop_locs_for_incl_hub,
  forecast_dates,
  locations,
  project_name = "COVID-19 Forecasts"
)
}
\arguments{
\item{prop_dates_for_incl_hub}{Numeric greater than 0 and less
than or equal to 1 indicating the inclusion threshold for the
proportion of forecast dates that a model must have submitted
forecasts to be included in analysis.}

\item{prop_locs_for_incl_hub}{Numeric less than 1 indicating
the inclusion threshold for the proportion of the locations
we expect that a model must have subbmited forecasts for to
be included in analysis.}

\item{forecast_dates}{vector of dates formatted in ISO8601 convention
(YYYY-MM-DD) indicating the forecast dates for the analysis}

\item{locations}{vector of state abbreviations that we want to
ensure the submitting teams have produced forecasts for.}

\item{project_name}{name of the Zoltar project, default is
\code{"COVID-19 Forecasts"}.}
}
\value{
a vector of character strings indicating the
unique model names that fit the inclusion criteria
}
\description{
This function uses the \code{zoltr} R package to connect to the Zoltar
database, which contains forecasts from the COVID Hub forecast
project, and query it for the specified forecast dates.
It queries for all dates, computes the proportion of dates for
which the model has submitted, filters for models that have
submitted for greater than the specified proportion of forecast
dates for inclusion, and returns the vector of model names.
}
