% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{read_rds_with_suffix}
\alias{read_rds_with_suffix}
\title{Load an object from an \code{.rds} serialized file
in a given \code{output_dir}, with programmatic naming.}
\usage{
read_rds_with_suffix(object_basename, dir, suffix = "", ext = "rds")
}
\arguments{
\item{object_basename}{Base name for the object \code{.rds}, before any
\code{suffix} or the file extension, as a string.}

\item{dir}{Directory from which to load the object.}

\item{suffix}{Suffix to append to the \code{object_basename} before
the file extension. Default \code{""} (no suffix).}

\item{ext}{Extension for the saved file, without the \code{.}.
Default \code{rds}.}
}
\value{
The loaded object.
}
\description{
Load an object from an \code{.rds} serialized file
in a given \code{output_dir}, with programmatic naming.
}
