% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample_model.R
\name{sample_model}
\alias{sample_model}
\title{Fit the model}
\usage{
sample_model(
  standata,
  compiled_model,
  init_lists,
  iter_warmup = 250,
  iter_sampling = 250,
  max_treedepth = 12,
  adapt_delta = 0.95,
  n_chains = 4,
  seed = 123
)
}
\arguments{
\item{standata}{a list of elements to pass to stan}

\item{compiled_model}{the compiled model object}

\item{init_lists}{nested list of initial parameter values for each chain}

\item{iter_warmup}{number of iterations to save in MCMC sampling,
default = 250}

\item{iter_sampling}{number of iterations to save in MCMC sampling,
default = 250}

\item{max_treedepth}{maximum treedepth of MCMC sampling, defauly = 12}

\item{adapt_delta}{MCMC accaptance probability, default = 0.95}

\item{n_chains}{number of independent MCMC chains to run, default = 4}

\item{seed}{seed of random number generator default = 123}
}
\value{
a list containing draws, diagnostics, and summary_diagnostics
}
\description{
Fit the model
}
