% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{save_only_ww_data}
\alias{save_only_ww_data}
\title{Save only the wastewater data from the output quantiles}
\usage{
save_only_ww_data(path_to_ww_quantiles, ww_data_dir)
}
\arguments{
\item{path_to_ww_quantiles}{Path to where the ww quantiles are saved}

\item{ww_data_dir}{directory where wastewater data is to be saved}
}
\value{
path to only the data
}
\description{
Save only the wastewater data from the output quantiles
}
