% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{save_rds_with_suffix}
\alias{save_rds_with_suffix}
\title{Save an object to .rds in a given \code{dir},
with programmatic naming.}
\usage{
save_rds_with_suffix(object, dir, basename = NULL, suffix = "", ext = "rds")
}
\arguments{
\item{object}{Object to save as an \code{.rds} file.}

\item{dir}{Directory in which to save the \code{.rds} file.}

\item{basename}{Base name for the \code{.rds}, before any
\code{suffix} or the file extension, as a string. If \code{NULL},
use the name of the object in the R environment. Default \code{NULL}.}

\item{suffix}{Suffix to append to the \code{basename} before
the file extension. Default \code{""} (no suffix).}

\item{ext}{Extension for the saved file, without the \code{.}.
Default \code{rds}.}
}
\value{
Nothing, saving the object as a side effect.
}
\description{
By default, names the \code{.rds} file according to
the object name in the R environment, with an
optional string suffix appended.
So for example a variable named \code{my_var} with
suffix equal to \code{"my_suffix"} would be saved to disk
as \verb{<output_dir>/my_var_my_suffix.rds}. Alternatively,
a custom \code{basename} can be supplied. The \code{suffix}
will still be added.
}
