% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{save_timestamped_nwss_data}
\alias{save_timestamped_nwss_data}
\title{Save time stampped NWSS data}
\usage{
save_timestamped_nwss_data(ww_path_to_save)
}
\arguments{
\item{ww_path_to_save}{filepath to save the time stamped wastewater datasets
default is data/ww_data
keys}
}
\value{
ww_data_path
}
\description{
Save time stampped NWSS data
}
