% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{save_ww_quantiles}
\alias{save_ww_quantiles}
\title{Save the combined wastewater quantiles}
\usage{
save_ww_quantiles(ww_quantiles, ww_output_path)
}
\arguments{
\item{ww_quantiles}{full set of vintaged wastewater data with fits}

\item{ww_output_path}{path to where to save this}
}
\description{
Save the combined wastewater quantiles
}
