% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_horizon}
\alias{scale_fill_horizon}
\alias{scale_color_horizon}
\title{ggplot discrete scales for forecast horizons}
\usage{
scale_fill_horizon(...)

scale_color_horizon(...)
}
\arguments{
\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_*_manual()}}.}
}
\value{
the ggplot scale
}
\description{
ggplot discrete scales for forecast horizons
}
