% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_model}
\alias{scale_fill_model}
\alias{scale_color_model}
\alias{scale_shape_model}
\title{ggplot discrete scales for forecast models}
\usage{
scale_fill_model(...)

scale_color_model(...)

scale_shape_model(...)
}
\arguments{
\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_fill_manual]{ggplot2::scale_*_manual()}}.}
}
\value{
the ggplot scale
}
\description{
ggplot discrete scales for forecast models
}
