% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_x_weekly_iso_date}
\alias{scale_x_weekly_iso_date}
\title{Default x axis date scale}
\usage{
scale_x_weekly_iso_date(...)
}
\arguments{
\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_date]{ggplot2::scale_x_date()}}.}
}
\value{
The scale.
}
\description{
Dates formatted in ISO YYYY-MM-DD format, weekly breaks.
}
