% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{score_samples}
\alias{score_samples}
\title{Get the scores for every day for a particular location and forecast date}
\usage{
score_samples(draws, scenario, offset, metrics = sample_metrics)
}
\arguments{
\item{draws}{a dataframe of the model estimated  quantity you are evaluating
alongside the evaluation data}

\item{scenario}{a string indicating the wastewater data scenario we're
running}

\item{offset}{Offset to use when transforming forecasts
with \code{\link[scoringutils:log_shift]{scoringutils::log_shift()}} via
\code{\link[scoringutils:transform_forecasts]{scoringutils::transform_forecasts()}}.}

\item{metrics}{Vector of scoring metrics to output, passed as the
\code{metrics} argument to \code{\link[scoringutils:score]{scoringutils::score()}}. Default
\link{sample_metrics}}
}
\value{
a dataframe containing a score for each day in the nowcast
and forecast period
}
\description{
Uses scoringutils to transform data and predictions using a log transform
with a log shift of 1, and return the default scoring metrics
}
