% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{select_like}
\alias{select_like}
\title{Select columns from one dataframe based on the column
spec of a second dataframe}
\usage{
select_like(df, template_df)
}
\arguments{
\item{df}{Dataframe from which to select columns}

\item{template_df}{Dataframe to use as a template
for the column specification}
}
\value{
The result of calling \code{\link[dplyr:select]{dplyr::select()}} on
\code{df}, raising an error if not all of the columns from
\code{template_df} can be found.
}
\description{
Select columns from one dataframe based on the column
spec of a second dataframe
}
