% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_secrets.R
\name{setup_secrets}
\alias{setup_secrets}
\title{Handles values needed to retrieve data from APIs to run pipeline}
\usage{
setup_secrets(yaml_path)
}
\arguments{
\item{yaml_path}{character, file path to a YAML containing the API keys, tokens, and IDs}
}
\value{
nothing, sets values as environmental variables
}
\description{
Handles values needed to retrieve data from APIs to run pipeline
}
