% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pre_processing.R
\name{subsample_sites}
\alias{subsample_sites}
\title{Subsample wastewater sites}
\usage{
subsample_sites(ww_data, prop_sites = 0.2, sampled_sites = NULL)
}
\arguments{
\item{ww_data}{Site-level wastewater data, extracted and trnsformed from
the raw NWSS data to contain only the relevant columns. This will be for all
time points}

\item{prop_sites}{Proportion of all sites that we want to keep in the input
data. default is 0.2}

\item{sampled_sites}{The list of sites that you want to keep, so that we
can compare the exact same subsample if desired. Default is NULL, then
random sampling is done according to the proportion}
}
\value{
a data frame structured the same way as the ww_data but with only
prop_sites included. These sites are chosen randomly
}
\description{
Subsample wastewater sites
}
