% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_predictive.R
\name{three_location_forecast_fig}
\alias{three_location_forecast_fig}
\title{Make a figure showing three example hospital admissions
forecast timeseries and associated wastewater concentration
timeseries.}
\usage{
three_location_forecast_fig(hosp1, hosp2, hosp3, ww_conc1, ww_conc2, ww_conc3)
}
\arguments{
\item{hosp1}{first hospital admissions forecast timseries}

\item{hosp2}{second hospital admissions forecast timseries}

\item{hosp3}{third hospital admissions forecast timseries}

\item{ww_conc1}{wastewater data and model fit associated to \code{hosp1}}

\item{ww_conc2}{wastewater data and model fit associated to \code{hosp2}}

\item{ww_conc3}{wastewater data and model fit associated to \code{hosp3}}
}
\value{
a combined grob, as the output of \code{\link[patchwork:plot_layout]{patchwork::plot_layout()}}.
}
\description{
Make a figure showing three example hospital admissions
forecast timeseries and associated wastewater concentration
timeseries.
}
