% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_rds_with_suffix}
\alias{to_rds_with_suffix}
\title{Save an object to .rds in a given \code{output_dir},
with programmatic naming.}
\usage{
to_rds_with_suffix(
  object,
  output_dir,
  save_basename = NULL,
  save_suffix = "",
  ext = "rds"
)
}
\arguments{
\item{object}{Object to save as an \code{.rds} file.}

\item{output_dir}{Directory in which to save the \code{.rds} file.}

\item{save_basename}{Base name for the \code{.rds}, before any
\code{suffix} or the file extension, as a string. If \code{NULL},
use the name of the object in the R environment. Default \code{NULL}.}

\item{save_suffix}{Suffix to append to the save_basename. Default
\code{""} (no suffix).}

\item{ext}{Extension for the saved file, without the \code{.}.
Default \code{rds}.}
}
\value{
Nothing, saving the object as a side effect.
}
\description{
By default, names the \code{.rds} file according to
the object name in the R environment, with an
optional string suffix appended.
So for example a variable named \code{my_var} with
suffix equal to \code{"my_suffix"} would be saved to disk
as \verb{<output_dir>/my_var_my_suffix.rds}. Alternatively,
a custom \code{save_basename} can be supplied. The \code{suffix}
will still be added.
}
