% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{wis_barplot}
\alias{wis_barplot}
\title{Make plot of WIS scores in Hub models overall}
\usage{
wis_barplot(scores, time_period)
}
\arguments{
\item{scores}{quantile based scores from the hub}

\item{time_period}{string indicating which time period to make the plot for}
}
\value{
A plot ordered by average wis over the time period
}
\description{
Make plot of WIS scores in Hub models overall
}
