% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{add_horizons}
\alias{add_horizons}
\title{Add columns indicating the horizon to the scores}
\usage{
add_horizons(df)
}
\arguments{
\item{df}{A tibble containing either forecasts or scores (or both)
and the following required columns: \code{date},\code{forecast_date},
\code{last_hosp_data_date}}
}
\value{
a tibble containing the same columns as \code{df} plus
\code{horizon_days} and \code{horizon}
}
\description{
This function takes in a tibble of scores on the nowcasts/forecasts
containing the columns \code{date}, \code{forecast_date}, and \code{last_hosp_data_date},
and adds the following columns: \code{horizon_days} (an integer) and
\code{horizon} (a string to be used for categorical grouping of horizons)
}
