% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R
\name{add_horizons_to_quantiles}
\alias{add_horizons_to_quantiles}
\title{Add columns indicating the horizon to the quantiles}
\usage{
add_horizons_to_quantiles(quantiles)
}
\arguments{
\item{quantiles}{A tibble containing the quantiled values and the following
required columns: \code{date}, \code{forecast_date}, and \code{period}}
}
\value{
a tibble containing the same columns as \code{quantiles} plus
\code{horizon_weeks} and \code{horizon}
}
\description{
This function takes in a tibble of quantiles containing the columns
\code{date}, \code{forecast_date} and \code{period} and adds the following columns:
\code{horizon_weeks} (an integer) and \code{horizon} (a string to be used for
categorical grouping of horizons)
}
