% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{bootstrap_crps_values}
\alias{bootstrap_crps_values}
\title{Create bootstrapped replicates to estimate uncertainty in
the mean CRPS for the admissions-only and wastewater-informed
models.}
\usage{
bootstrap_crps_values(scores, n_replicates, by = NULL)
}
\arguments{
\item{scores}{data frame of relative scores for the wastewater
model with the admissions-only model as a baseline.}

\item{n_replicates}{Number of bootstrap replicate datasets to produce.}

\item{by}{Optional grouping column for bootstrap replicate datasets. Default
\code{NULL} (no grouping, produce replicates of the entire dataset).}
}
\value{
Tibble with the grouping column specified in \code{by} (if any)
and four other columns:
\itemize{
\item \code{id}: unique (within-group) identifier of the replicate dataset.
\item \code{bstrap_crps_ww}: mean CRPS for the wastewater model in that replicate dataset.
\item \code{bstrap_crps_hosp}: mean CRPS for the admissions-only model in that replicate dataset.
sampled for that replicate.
\item \code{bstrap_rel_crps}: ratio (ww / hosp) of the mean CRPS values for the two models
in that replicate dataset.
}
}
\description{
Produces bootstrapped values  mean CRPS
for each value of the provided grouping variable. Each
}
