% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{combine_hub_and_local_wis}
\alias{combine_hub_and_local_wis}
\title{Combine the hosp only real-time scores and cfa real-
time scores from github}
\usage{
combine_hub_and_local_wis(cfa_real_time_scores, real_time_wis_hosp_only)
}
\arguments{
\item{cfa_real_time_scores}{Hub formatted scores for
only the ww model, \code{cfa-wwrenewal(real-time)}}

\item{real_time_wis_hosp_only}{hosp only
wis scores calculated from local data}
}
\value{
a df formatted the same way as the real_time_wis_hosp_only scores,
but for both models
}
\description{
Combine the hosp only real-time scores and cfa real-
time scores from github
}
