% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_outputs.R
\name{combine_outputs}
\alias{combine_outputs}
\title{Combine outputs}
\usage{
combine_outputs(
  output_type,
  scenarios,
  forecast_dates,
  locations,
  eval_output_subdir,
  model_type,
  strict = FALSE
)
}
\arguments{
\item{output_type}{the type of output that is saved, one of \code{"scores"},
\code{"ww_quantiles"}, \code{"scores_quantiles"}, \code{"hosp_quantiles"},\code{"errors"},
\code{"ww_data_flags"} or \code{"flags"}.}

\item{scenarios}{The vector of character strings of all the scenarios}

\item{forecast_dates}{The vector of character strings of all the forecast dates}

\item{locations}{The vector of character strings of all the locations}

\item{eval_output_subdir}{The outer subdirectory of the nested file structure}

\item{model_type}{The type of model, either \code{"ww"} or \code{"hosp"}}

\item{strict}{Error if any of the expected outputs to combine are missing?
Boolean, default \code{FALSE}.}
}
\value{
combined_output: a tibble with output types for all combinations of
forecast_dates, locations, and scenarios, or \code{NULL} if no outputs found
and \code{strict} is \code{FALSE}
}
\description{
This function is a helper function specific to the current nested file
structure specified by the functions that save the quantiles and scores
for each of the model runs to disk. The function takes in the
vectors of scenarios, forecast dates, and locations with output model runs,
checks if they exist in the current file structure, and if they do
loads them in and binds them together to create one large dataframe with
all of the outputs row binded
}
