% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_rel_performance_fig}
\alias{compose_rel_performance_fig}
\title{Make a multi-panel figure summarizing relative
performance.}
\usage{
compose_rel_performance_fig(
  rel_score_heatmap,
  rel_score_dist,
  abs_score_by_time,
  total_admissions,
  rel_score_dist_by_time,
  rel_score_dist_by_location,
  time_period
)
}
\arguments{
\item{rel_score_heatmap}{heatmap of relative score by forecast date
and location}

\item{rel_score_dist}{distibution plot comparing overall distribution
of relative scores across forecast_date, date, location, and model}

\item{abs_score_by_time}{timeseries plot of absolute score across
by forecast date}

\item{rel_score_dist_by_location}{plot of the distribution of (date-
specific) relative score values by location.}

\item{time_period}{string to save fig as, either "real_time" or
"all_time"}

\item{natl_admissions}{timeseries plot of total hospital
admissions by day}

\item{rel_score_dist_by_date}{plot of the distribution of (location
-specific) relative score values by forecast_dat.}

\item{fig_file_dir}{Path to save figures}
}
\value{
ggplot object with all the elements combined
}
\description{
Make a multi-panel figure summarizing relative
performance.
}
