% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{compose_rel_performance_fig}
\alias{compose_rel_performance_fig}
\title{Make a multi-panel figure summarizing relative
performance.}
\usage{
compose_rel_performance_fig(
  rel_crps_heatmap,
  rel_crps_dist,
  abs_crps_over_time,
  total_admissions,
  rel_crps_dist_over_time,
  rel_crps_by_location,
  time_period,
  fig_file_dir
)
}
\arguments{
\item{rel_crps_heatmap}{heatmap of relative crps by forecast date
and location}

\item{rel_crps_dist}{distibution plot comparing overall distribution
of crps scores across forecast_date, date, location, and model}

\item{abs_crps_over_time}{avg crps across locations by forecast date}

\item{rel_crps_dist_over_time}{plots of the distribution of location
-specific relative crps values over time (by forecast date).}

\item{rel_crps_by_location}{avg crps across forecast dates by state}

\item{time_period}{string to save fig as, either "real_time" or
"all_time"}

\item{fig_file_dir}{Path to save figures}

\item{natl_admissions}{total admissions by day}
}
\value{
ggplot object with all the elements combined
}
\description{
Make a multi-panel figure summarizing relative
performance.
}
