% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/composite_figures.R
\name{compose_rel_performance_fig}
\alias{compose_rel_performance_fig}
\title{Make a multi-panel figure summarizing relative
performance.}
\usage{
compose_rel_performance_fig(
  rel_score_heatmap,
  rel_score_dist,
  rel_score_by_time,
  total_admissions,
  abs_scores_by_time,
  abs_scores_by_location
)
}
\arguments{
\item{rel_score_heatmap}{heatmap of relative score by forecast date
and location}

\item{rel_score_dist}{distibution plot comparing overall distribution
of relative scores by forecast_date and location}

\item{rel_score_by_time}{timeseries plot of relative score
by forecast date.}

\item{total_admissions}{timeseries plot of total hospital
admissions by day.}

\item{abs_scores_by_time}{plot (location-specific) score values by forecast_date.}

\item{abs_scores_by_location}{plot of (date-specific) score values by location.}
}
\value{
ggplot object with all the elements combined
}
\description{
Make a multi-panel figure summarizing relative
performance.
}
