% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{compute_flags}
\alias{compute_flags}
\title{Compute diagnostic flags from raw stanfit objects}
\usage{
compute_flags(forecast_dir, forecast_date, run_on_date, run_id)
}
\description{
Compute diagnostic flags from raw stanfit objects
}
