% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_forecast_differences.R
\name{compute_forecast_differences}
\alias{compute_forecast_differences}
\title{Compute the posterior of the difference between model predictions.}
\usage{
compute_forecast_differences(
  forecast_date,
  location,
  scenario,
  raw_output_dir,
  log_diff_offset
)
}
\arguments{
\item{forecast_date}{Forecast date for which to compute differences.}

\item{location}{Location for which to compute differences.}

\item{scenario}{Wastewater data availability scenario to analyze.}

\item{raw_output_dir}{Directory containing raw output \code{.rds} files.
Used to obtain the admissions and wastewater data used in fitting
the forecasting model.}

\item{log_diff_offset}{Offset to add to incidences when
computing a log difference.}
}
\value{
Posterior of the difference in total incidence
across the full nowcast/forecast period. Saves
the incidence differences by day and the total
incidence difference to disk as a side effect.
}
\description{
Compute the posterior of the difference between model predictions.
}
