% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{.relative_metric_display_name}
\alias{.relative_metric_display_name}
\title{Internal function for display names of relative metrics}
\usage{
.relative_metric_display_name(metric_to_compare)
}
\arguments{
\item{metric_to_compare}{Name of the absolute metric,
as it would be passed as a \code{metric_to_compare} argument to
\code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}}.}
}
\value{
The relative metric display name, as a character string.
}
\description{
Internal function for display names of relative metrics
}
