% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{.target_model_relative_scores}
\alias{.target_model_relative_scores}
\title{internal function for two-model relative score computations}
\usage{
.target_model_relative_scores(scores, target_model, baseline_model, by = NULL)
}
\arguments{
\item{target_model}{Target model (numerator for the relative scores)}

\item{baseline_model}{Baseline model
(denominator for the relative scores)}

\item{by}{columns to summarize by. Passed as the \code{by} argument
to \code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}

\item{score}{table of scores, as the output of
\code{\link[scoringutils:score]{scoringutils::score()}}}
}
\value{
The relative scores for the target model, as a table.
}
\description{
internal function for two-model relative score computations
}
\keyword{internal}
