% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{.target_model_relative_scores}
\alias{.target_model_relative_scores}
\title{internal function for two-model relative score computations}
\usage{
.target_model_relative_scores(
  scores,
  target_model,
  baseline_model,
  metric_to_compare,
  by = NULL
)
}
\arguments{
\item{scores}{table of scores, as the output of
\code{\link[scoringutils:score]{scoringutils::score()}}}

\item{target_model}{Target model (numerator for the relative scores)}

\item{baseline_model}{Baseline model
(denominator for the relative scores)}

\item{metric_to_compare}{Metric for which to compute
relative scores. One of \code{"wis"} or \code{"crps"}. Not case-sensitive.
Passed as the \code{metric_to_compare} argument
to \code{\link[scoringutils:get_pairwise_comparisons]{scoringutils::get_pairwise_comparisons()}} via
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}

\item{by}{columns to summarize by. Passed as the \code{by} argument
to \code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.}
}
\value{
The relative scores for the target model, as a table.
}
\description{
internal function for two-model relative score computations
}
\keyword{internal}
