% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_fit.R
\name{eval_fit}
\alias{eval_fit}
\title{Fit model for evaluation}
\usage{
eval_fit(config_index, eval_config_path, params_path, model)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}

\item{model}{Model to fit. One of \code{"ww"} and \code{"hosp"}}
}
\description{
Fit model for evaluation
}
