% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_post_process.R
\name{eval_post_process_ww}
\alias{eval_post_process_ww}
\title{Postprocess Wastewater Model for Evaluation}
\usage{
eval_post_process_ww(config_index, eval_config_path, params_path)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}
}
\description{
Postprocess Wastewater Model for Evaluation
}
