% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_postprocess.R
\name{eval_postprocess}
\alias{eval_postprocess}
\title{Postprocess model for evaluation.}
\usage{
eval_postprocess(
  config_index,
  eval_config_path,
  params_path,
  model,
  max_eval_data_days = 365
)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}

\item{model}{model to postprocess. One of \code{"hosp"} and \code{"ww"}.}

\item{max_eval_data_days}{Maximum number of days of data to pull
when creating evaluation dataset. Default 365.}
}
\value{
NULL, saving plots and tables to disk as side effects.
}
\description{
Raw output is saved to disk in the \code{raw_output_dir} specified in the eval config
as serialized \code{.rds} files. Processed output is saved in a structured directory
format in the \code{output_dir} specified in the eval config. See the \code{\link[=save_table]{save_table()}},
\code{\link[=to_rds_with_suffix]{to_rds_with_suffix()}}, and \code{\link[=get_raw_output_suffix]{get_raw_output_suffix()}} functions for more details.
}
\details{
The bulk of the postprocessing for successful model fits is handled
by the \code{\link[=postprocess_successful_fit]{postprocess_successful_fit()}} function, which is called within this
function provided the target model fit was indeed successful, but there are
some postprocessing tasks we wish to perform regardless of
whether the model fit was successful.
}
