% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_post_process.R
\name{eval_postprocess}
\alias{eval_postprocess}
\title{placeholder wrapper for testing refactor}
\usage{
eval_postprocess(
  config_index,
  eval_config_path,
  params_path,
  model,
  max_eval_data_days = 365
)
}
\description{
placeholder wrapper for testing refactor
}
