% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{eval_postprocess}
\alias{eval_postprocess}
\title{Postprocess model for evaluation.}
\usage{
eval_postprocess(
  forecast_date,
  eval_date,
  location,
  model,
  scenario,
  hosp_data_dir,
  ww_data_dir,
  ww_data_mapping,
  scenario_dir,
  output_dir,
  raw_output_dir,
  max_eval_data_days = 365
)
}
\arguments{
\item{forecast_date}{As-of date for the forecast.}

\item{eval_date}{As-of date for evaluation data to use.}

\item{location}{Location to forecast.}

\item{model}{Model to postprocess. One of \code{"ww"} and \code{"hosp"}}

\item{scenario}{Wastewater data availability scenario to analyze.}

\item{hosp_data_dir}{Path to a directory containing vintaged hospital
admissions data in date-stamped .csv files.}

\item{ww_data_dir}{Path to a directory containing vintaged wastewater
data in date-stamped .csv files.}

\item{ww_data_mapping}{String associating forecast dates to wastewater
vintage dates. Passed to \code{\link[=date_of_ww_data]{date_of_ww_data()}}.}

\item{scenario_dir}{Path to a directory containing .csv files that
define wastewater data availability scenarios.}

\item{output_dir}{Path to a direcotry in which to save general
postprocess output.}

\item{raw_output_dir}{Path to a directory in which to save
raw output as serialized \code{.rds} files.}

\item{max_eval_data_days}{Maximum number of days of data to pull
when creating evaluation dataset. Default 365.}
}
\value{
NULL, saving plots and tables to disk as side effects.
}
\description{
Raw output is saved to disk in the \code{raw_output_dir} specified in the eval config
as serialized \code{.rds} files. Processed output is saved in a structured directory
format in the \code{output_dir} specified in the eval config. See the \code{\link[=save_table]{save_table()}},
\code{\link[=to_rds_with_suffix]{to_rds_with_suffix()}}, and \code{\link[=get_raw_output_suffix]{get_raw_output_suffix()}} functions for more details.
}
\details{
The bulk of the postprocessing for successful model fits is handled
by the \code{\link[=postprocess_successful_fit]{postprocess_successful_fit()}} function, which is called within this
function provided the target model fit was indeed successful, but there are
some postprocessing tasks we wish to perform regardless of
whether the model fit was successful.
}
