% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eval_postprocess.R
\name{eval_postprocess}
\alias{eval_postprocess}
\title{Postprocess model for evaluation.}
\usage{
eval_postprocess(
  config_index,
  eval_config_path,
  params_path,
  model,
  max_eval_data_days = 365
)
}
\arguments{
\item{config_index}{Index of eval_config to evaluate}

\item{eval_config_path}{Path to eval_config (created with \code{write_eval_config})}

\item{params_path}{Path to params.toml}

\item{model}{model to postprocess. One of \code{"hosp"} and \code{"ww"}.}

\item{max_eval_data_days}{Maximum number of days of data to pull
when creating evaluation dataset. Default 365.}
}
\value{
NULL, saving plots and tables to disk as side effects.
}
\description{
Postprocess model for evaluation.
}
