% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_diagnostic_extrema}
\alias{extract_diagnostic_extrema}
\title{Compute the maximum rhat and minimum bulk and tail ESS for
all or a subset of parameters in a stan fit object.}
\usage{
extract_diagnostic_extrema(stanfit, variables = NULL)
}
\arguments{
\item{stanfit}{CmdStanR fit object}

\item{variables}{parameter(s) to consider. Passed as the
\code{variables} argument to \link[cmdstanr:CmdStanFit$summary]{cmdstanr::CmdStanFit$summary}.
Default NULL (consider all).}
}
\value{
tibble with the summary
}
\description{
Compute the maximum rhat and minimum bulk and tail ESS for
all or a subset of parameters in a stan fit object.
}
