% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diagnostics.R
\name{extract_parameter_diagnostics}
\alias{extract_parameter_diagnostics}
\title{Compute the maximum rhat and minimum bulk and tail ESS from an array of parameters}
\usage{
extract_parameter_diagnostics(stanfit, parameter, index_subset = NULL)
}
\arguments{
\item{stanfit}{CmdStanR fit object}

\item{parameter}{name of the array-valued parameter}

\item{index_subject}{optional subset of indices to consider,
e.g. \code{10:15}
from the overall array}
}
\value{
tibble with the summary
}
\description{
Compute the maximum rhat and minimum bulk and tail ESS from an array of parameters
}
