% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_recent_trends.R
\name{fit_hosp_trend}
\alias{fit_hosp_trend}
\title{Fit a hosp trend with BRMS}
\usage{
fit_hosp_trend(
  data,
  exp_rate_hosp_prior_mean,
  exp_rate_hosp_prior_sd,
  log_nb_conc_prior_mean,
  log_nb_conc_prior_sd,
  seed,
  control,
  ...
)
}
\arguments{
\item{data}{Data to fit. Must have columns \code{"hosp"} and \code{"time"}.}

\item{exp_rate_hosp_prior_mean}{Mean for the Normal prior on
the exponential growth rate of admissions.}

\item{exp_rate_hosp_prior_sd}{Standard deviation for the Normal prior on
the exponential growth rate of admissions.}

\item{log_nb_conc_prior_mean}{Mean for the Normal prior on
the log of the negative binomial concentration parameter ("phi").}

\item{log_nb_conc_prior_sd}{Standard deviation for the Normal prior on
the log of the negative binomial concentration parameter ("phi").}

\item{seed}{Seed for cmdnstan's pseudorandom number generator.}

\item{control}{list of control parameters passed to stan via
\code{\link[brms:brm]{brms::brm()}}. Default \code{NULL}.}

\item{...}{Additional keyword arguments passed to \code{\link[brms:brm]{brms::brm()}}}
}
\description{
Fit a hosp trend with BRMS
}
