% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit_recent_trends.R
\name{fit_recent_trends}
\alias{fit_recent_trends}
\title{Fit recent hospital admission and wastewater
trends using BRMS}
\usage{
fit_recent_trends(
  forecast_date,
  location,
  scenario,
  raw_output_dir,
  hosp_prior_params,
  ww_prior_params,
  hosp_lookback_days,
  ww_lookback_days,
  seed,
  control = NULL,
  ...
)
}
\arguments{
\item{forecast_date}{forecast_date for which to fit recent data
trends. Data for an actual forecast must already have been produced
and exist in \code{raw_output_dir}.}

\item{location}{Location for which to fit trends.}

\item{scenario}{Wastewater data availability scenario to analyze.}

\item{raw_output_dir}{Directory containing raw output \code{.rds} files.
Used to obtain the admissions and wastewater data used in fitting
the forecasting model.}

\item{hosp_prior_params}{Named list of values for prior
hyperparameters needed by \code{\link[=fit_hosp_trend]{fit_hosp_trend()}}.}

\item{ww_prior_params}{Named list of values for prior hyperparameters
needed by \code{\link[=fit_ww_trend]{fit_ww_trend()}}.}

\item{hosp_lookback_days}{Number of days to look back when fitting the
trend in hospital admissions.}

\item{ww_lookback_days}{Number of days to look back when fitting the
trend in wastewater concentrations.}

\item{seed}{Seed for Stan's pseudorandom number generator.}

\item{control}{list of control parameters passed to stan via
\code{\link[brms:brm]{brms::brm()}}. Default \code{NULL}.}

\item{...}{Additional keyword arguments passed to \code{\link[brms:brm]{brms::brm()}}
via \code{\link[=fit_hosp_trend]{fit_hosp_trend()}}.}
}
\value{
The two fit objects, as a list, saving raw fits to disk
as a side effect.
}
\description{
Fit recent hospital admission and wastewater
trends using BRMS
}
