% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forecast_interval_coverage_plot}
\alias{forecast_interval_coverage_plot}
\title{Plot interval coverage at specified ranges}
\usage{
forecast_interval_coverage_plot(
  forecasts,
  ranges,
  time_period,
  fig_file_dir = NULL,
  write_files = FALSE
)
}
\arguments{
\item{forecasts}{df of granular (daily) quantile forecasts}

\item{ranges}{A numeric vector of credible interval ranges to plot,
spanning from 0 to 100.}

\item{time_period}{string indicating time period of fig to save}

\item{fig_file_dir}{string indicating directory to save figure in,
default is \code{NULL}}

\item{write_files}{boolean indicating whether to save file, default is FALSE}
}
\value{
A ggplot2 object
}
\description{
Plot interval coverage at specified ranges
}
