% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forecast_qq_plot}
\alias{forecast_qq_plot}
\title{Get quantile-quantile plot}
\usage{
forecast_qq_plot(forecasts, time_period)
}
\arguments{
\item{forecasts}{df of granular (daily) quantile forecasts}

\item{time_period}{time period that scores are summarized over}
}
\value{
a ggplot object containing a plot of the proportion of data within
each interval for each model.
}
\description{
Get quantile-quantile plot
}
