% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{forecast_qq_plot}
\alias{forecast_qq_plot}
\title{Get quantile-quantile plot}
\usage{
forecast_qq_plot(
  forecasts,
  time_period,
  fig_file_dir = NULL,
  save_files = TRUE
)
}
\arguments{
\item{forecasts}{df of granular (daily) quantile forecasts}

\item{time_period}{time period that scores are summarized over}

\item{save_files}{save_files boolean indicating whether or not to save figures, default
is \code{TRUE}}

\item{figure_file_path}{path to save figure}
}
\value{
a ggplot object containing a plot of the proportion of data within
each interval for each model.
}
\description{
Get quantile-quantile plot
}
