% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{format_scores_for_comparison}
\alias{format_scores_for_comparison}
\title{Format the hosp only real time scores for comparison to the other real
time models}
\usage{
format_scores_for_comparison(real_time_scores)
}
\arguments{
\item{real_time_scores}{the set of real time scores gathered from local
pull}
}
\value{
a tibble formatted as the other real time scores for the real
time hosp only model
}
\description{
Format the hosp only real time scores for comparison to the other real
time models
}
