% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{geom_decomposed_scores}
\alias{geom_decomposed_scores}
\title{Composite geometric object for barplots of scores decomposed into
undeprediction, overprediction, and dispersion components}
\usage{
geom_decomposed_scores(dispersion_alpha = 0.6, ...)
}
\arguments{
\item{dispersion_alpha}{Alpha parameter for the dispersion tile.
Default 0.5}

\item{...}{keyword arguments passed to \code{\link[ggplot2:geom_tile]{ggplot2::geom_tile()}}.}
}
\description{
Composite geometric object for barplots of scores decomposed into
undeprediction, overprediction, and dispersion components
}
