% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_summarize_ww_data.R
\name{get_add_ww_metadata}
\alias{get_add_ww_metadata}
\title{Get additional wastewater metadata}
\usage{
get_add_ww_metadata(
  granular_ww_metadata,
  ww_forecast_date_locs_to_excl,
  convergence_df,
  table_of_loc_dates_w_ww,
  include_manual_exclusions = FALSE
)
}
\arguments{
\item{granular_ww_metadata}{a tibble with a row for each forecast date
location and columns that provide summaries of the wastewater data. This
is focused on input data. It is the output of
"combine_and_summarize_ww_data".}

\item{ww_forecast_date_locs_to_excl}{table of forecast date-locations
to manually exclude}

\item{convergence_df}{table containing convergence flags for every forecast
date location combination}

\item{table_of_loc_dates_w_ww}{table containing wastewater metadata
for every location-forecast date with wastewater}

\item{include_manual_exclusions}{boolean indicating whether or not the
ww metadata should include manual exclusions, default is FALSE bc isn't
used in main retro head to head analysis.}
}
\value{
a tibble with a number of additional columns indicating whether
or not there were any flags for manual exclusions, convergence issues,
or wastewater quality issues
}
\description{
This function takes the summary statistics from just the input wastewater
data (granular_ww_metadata) and joins the downstream metadata about
convergenece and manual exclusions and combines them all into one table
}
