% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run_diagnostic_flags.R
\name{get_convergence_df}
\alias{get_convergence_df}
\title{Get convergence dataframe}
\usage{
get_convergence_df(all_flags, scenario)
}
\arguments{
\item{all_flags}{a dataframe containing the flags for each location,
forecast_date, and scenario}

\item{scenario}{The scenario to filter to, since some eval output will include multiple
scenarios}
}
\value{
a dataframe with a column \code{any_flags} indicating whether any of the
flags in the original full descriptive set of congerence flags are TRUE.
}
\description{
This function takes the larger dataframe of convergence
flags for each location, forecast date, and scenario and checks if any of
the flags are TRUE, and returns a dataframe with just a column indicating
whether any flags are true
}
