% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/model_run_diagnostic_flags.R
\name{get_diagnostic_flags}
\alias{get_diagnostic_flags}
\title{Get diagnostic flags}
\usage{
get_diagnostic_flags(
  stan_fit_object,
  n_chains,
  iter_sampling,
  ebmfi_tolerance = 0.2,
  divergences_tolerance = 0.01,
  p_high_rhat_tolerance = 0.05,
  max_tree_depth_tol = 0.01
)
}
\arguments{
\item{stan_fit_object}{The R6 Cmdstan Object fit object}

\item{n_chains}{number of chains fun}

\item{iter_sampling}{number of sampling iterations}

\item{ebmfi_tolerance}{Tolerance for EBMFI (bayesian missing information)}

\item{divergences_tolerance}{tolerance for proportion of sampling iterations
that are divergent}

\item{p_high_rhat_tolerance}{tolerance for proportion of parameters rhats>1.05}

\item{max_tree_depth_tol}{tolerance for proportion of iterations that exceed
the maximum tree depth}
}
\value{
flag_df: dataframe containing columns for each of the flags,
if any flags are TRUE that indicates some model issue
}
\description{
This function takes in the output from a cmdstanr$sample() function (the
fit object) and a series of diagnostic toleraances and returns
a dataframe containing flags for whether any of the diagnostic thresholds
were exceeded, which would indicate that the model did not properly
converge
}
