% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_full_scores}
\alias{get_full_scores}
\title{Get the scores for ever day for a particular location and forecast date}
\usage{
get_full_scores(draws, scenario, metrics = NULL)
}
\arguments{
\item{draws}{a dataframe of the model estimated  quantity you are evaluating
alongside the evaluation data}

\item{scenario}{a string indicating the wastewater data scenario we're
running}

\item{metrics}{Vector of scoring metrics to output, passed as the
\code{metrics} argument to \code{\link[scoringutils:score]{scoringutils::score()}}. Default is NULL,
which returns all options for samples including:
\code{c("crps", "dss", "bias", "mad", "ae_median", "se_mean")}.}
}
\value{
a dataframe containing a score for each day in the nowcast
and forecast period
}
\description{
Logs the truth data and forecasts and scores using scoringutils
}
