% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{get_full_scores}
\alias{get_full_scores}
\title{Get the scores for ever day for a particular location and forecast date}
\usage{
get_full_scores(
  draws,
  scenario,
  metrics = c("crps", "dss", "bias", "mad", "ae_median", "se_mean"),
  ...
)
}
\arguments{
\item{draws}{a dataframe of the model estimated  quantity you are evaluating
alongside the evaluation data}

\item{scenario}{a string indicating the wastewater data scenario we're
running}

\item{metrics}{Vector of scoring metrics to output, passed as the
\code{metrics} argument to \code{\link[scoringutils:score]{scoringutils::score()}}. Default
\code{c("crps", "dss", "bias", "mad", "ae_median", "se_mean")}.}

\item{...}{Additional named arguments passed to \code{\link[scoringutils:score]{scoringutils::score()}}.}
}
\value{
a dataframe containing a score for each day in the nowcast
and forecast period
}
\description{
Get the scores for ever day for a particular location and forecast date
}
