% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_heatmap_metadata}
\alias{get_heatmap_metadata}
\title{Get a heatmap of the metadata of reasons for excluding forecasts from analysis}
\usage{
get_heatmap_metadata(metadata, fig_file_dir)
}
\arguments{
\item{metadata}{a tibble of location -forecast date metadata}
}
\value{
a ggplot object with a heatmap colored by reason for excluding
}
\description{
Get a heatmap of the metadata of reasons for excluding forecasts from analysis
}
