% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_heatmap_metadata_hub}
\alias{get_heatmap_metadata_hub}
\title{Get a heatmap of the metadata of Hub models submitted}
\usage{
get_heatmap_metadata_hub(metadata, fig_file_dir)
}
\arguments{
\item{metadata}{a tibble of location -forecast date metadata}

\item{fig_file_dir}{string indicating where to save figs}
}
\value{
a ggplot object with a heatmap colored by reason for excluding
}
\description{
Get a heatmap of the metadata of Hub models submitted
}
