% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_heatmap_scores}
\alias{get_heatmap_scores}
\title{Get heatmap of scores by scenario}
\usage{
get_heatmap_scores(mock_submission_scores, score_metric = "crps")
}
\arguments{
\item{mock_submission_scores}{The dataframe containing the daily scores for a
all of the forecast dates and locations}

\item{score_metric}{A string indicating the score metric to plot,
default is "crps"}
}
\value{
a ggplot object that generates a heatmap colored by score, across
forecast dates (x axis) and locations (y-axis)
}
\description{
Get heatmap of scores by scenario
}
