% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_input_data.R
\name{get_input_hosp_data}
\alias{get_input_hosp_data}
\title{Get input hospital admissions data}
\usage{
get_input_hosp_data(
  forecast_date_i,
  location_i,
  hosp_data_dir,
  calibration_time,
  load_from_covidcast = FALSE
)
}
\arguments{
\item{forecast_date_i}{The forecast date for this iteration}

\item{location_i}{The location (state) for this iteration}

\item{hosp_data_dir}{A string indicating the path to the directory containing
time stamped hospital admissions datasets}

\item{calibration_time}{A numeric indicating the duration of model
calibration (based on the last hospital admissions data point)}

\item{load_from_covidcast}{boolean indicating whether or not the hospital
admissions datasets should be loaded directly from covidcast.
\code{default = FALSE} because we are assuming that we have already created a
folder with time stamped datasets}
}
\value{
a dataframe containing the cleaned hospital admissions needed as
an input to the stan model for the specified forecast date and location
}
\description{
Get input hospital admissions data
}
