% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{get_loc_rel_crps}
\alias{get_loc_rel_crps}
\title{Get the mean relative crps for a location}
\usage{
get_loc_rel_crps(scores, locs)
}
\arguments{
\item{scores}{tibble of scores by day forecast day model}

\item{locs}{loc to get mean relative score for}
}
\value{
table of mean relative score for each location
}
\description{
Get the mean relative crps for a location
}
