% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_n_states_improved_plot}
\alias{get_n_states_improved_plot}
\title{Get bar chart of the number of states improved compared to without wastewater
over each forecast date}
\usage{
get_n_states_improved_plot(
  all_scores,
  figure_file_path,
  baseline_scenario = "no_wastewater",
  threshold_for_improvement = 1,
  save_files = TRUE
)
}
\arguments{
\item{all_scores}{the full set of scores for each location, forecast date,
scenario, and date}

\item{figure_file_path}{Outer directory for plots to go in}

\item{baseline_scenario}{the scenario to compute the difference in
crps scores relative to, default is \code{no_wastewater}}

\item{threshold_for_improvement}{the relative reduction that we will consider
as an improvement over the baseline,}

\item{save_files}{Whether or not to write to a folder of plots,
default is \code{TRUE}}
}
\value{
p a ggplot object with a bar chart showing the number of
jurisdictions with an improvement compared to the forecast without
wastewater under each scenario.
}
\description{
Get bar chart of the number of states improved compared to without wastewater
over each forecast date
}
