% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_overall_rel_crps_rt}
\alias{get_overall_rel_crps_rt}
\title{Get the overall relative crps for the real time scores}
\usage{
get_overall_rel_crps_rt(all_scores)
}
\arguments{
\item{all_scores}{a tibble of the scores for both models in real-time}
}
\value{
A tibble summarized the relative crps across all locs
and forecast dates produced in real-time
}
\description{
Get the overall relative crps for the real time scores
}
