% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_plot_bias_over_time}
\alias{get_plot_bias_over_time}
\title{Get a plot of bias over time}
\usage{
get_plot_bias_over_time(scores, fig_subscript)
}
\arguments{
\item{scores}{a complete df of scores with a column for bias}

\item{fig_subscript}{subscript to name the figure}
}
\value{
a plot of bias over time averaged across locations and forecast dates,
separated by model
}
\description{
Get a plot of bias over time
}
