% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_plot_hub_perf_heatmap}
\alias{get_plot_hub_perf_heatmap}
\title{Plot a heatmap of the avg forecast performance by locations and forecast date
for the Hub submissions}
\usage{
get_plot_hub_perf_heatmap(scores, fig_file_dir)
}
\arguments{
\item{scores}{A tibble of daily scores by forecast date, location, and model}

\item{fig_file_dir}{A string indicating the directory to save the figures in}
}
\value{
a ggplot object
}
\description{
Plot a heatmap of the avg forecast performance by locations and forecast date
for the Hub submissions
}
