% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_raw_scores}
\alias{get_plot_raw_scores}
\title{Get plot raw scores}
\usage{
get_plot_raw_scores(all_scores, score_metric = "crps")
}
\arguments{
\item{all_scores}{a dataframe containing daily scores for each scenario,
and forecast date, for a single location}

\item{score_metric}{A string indicating the score metric to plot,
default is "crps"}
}
\value{
a ggplot object plotting all of the scores over time
}
\description{
Get plot raw scores
}
