% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{get_plot_rel_crps_distrib}
\alias{get_plot_rel_crps_distrib}
\title{Get a density plot of the relative CRPS distribution}
\usage{
get_plot_rel_crps_distrib(scores, fig_file_dir)
}
\arguments{
\item{scores}{tibble of scores by horizon day, forecast date, and location}

\item{fig_file_dir}{directory to save figure in}
}
\value{
ggplot object of distribution of relative CRPS scores
}
\description{
Get a density plot of the relative CRPS distribution
}
