% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{get_plot_rel_wis_distrib}
\alias{get_plot_rel_wis_distrib}
\title{Get a density plot of the relative WIS distribution}
\usage{
get_plot_rel_wis_distrib(wis_scores)
}
\arguments{
\item{scores}{tibble of scores by horizon day, forecast date, and location}
}
\value{
ggplot object of distribution of relative CRPS scores
}
\description{
Get a density plot of the relative WIS distribution
}
