% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_plot_score_by_horizon_t}
\alias{get_plot_score_by_horizon_t}
\title{Get plot of scores by horizon over time}
\usage{
get_plot_score_by_horizon_t(scores, score_type, fig_file_dir)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{score_type}{string indicating the type of score to summarize over}

\item{fig_file_dir}{Path to save figures}
}
\value{
plot of scores over time faceted by horizon
}
\description{
Get plot of scores by horizon over time
}
