% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_plot_scores_and_forecasts}
\alias{get_plot_scores_and_forecasts}
\title{Got a plot of forecasts and scores for a single
location and date}
\usage{
get_plot_scores_and_forecasts(scores_single_loc_date, eval_output_subdir)
}
\arguments{
\item{scores_single_loc_date}{Scores for a single location
and data, as the output of \code{\link[scoringutils:score]{scoringutils::score()}}.}

\item{eval_output_subdir}{Subdirectory in which to save the plot.}
}
\description{
Got a plot of forecasts and scores for a single
location and date
}
