% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_scores_w_data}
\alias{get_plot_scores_w_data}
\title{Get plot scores with evaluation data overlaid}
\usage{
get_plot_scores_w_data(all_scores, eval_data, score_metric = "crps")
}
\arguments{
\item{all_scores}{a dataframe containing daily scores for each scenario,
and forecast date, for a single location}

\item{eval_data}{a dataframe containing the hospital admissions data
that the forecasts are evaluated against}

\item{score_metric}{A string indicating the score metric to plot,
default is "crps"}
}
\value{
a ggplot object plotting the scores summarized by forecast date
and scenario over time with the data overlaid
}
\description{
Get plot scores with evaluation data overlaid
}
