% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{get_plot_summarized_scores}
\alias{get_plot_summarized_scores}
\title{Get plot summarized scores}
\usage{
get_plot_summarized_scores(all_scores, score_metric = "crps")
}
\arguments{
\item{all_scores}{a dataframe containing daily scores for each scenario,
and forecast date, for a single location}

\item{score_metric}{A string indicating the score metric to plot,
default is "crps"}
}
\value{
a ggplot object plotting the scores summarized by forecast date
and scenario over time
}
\description{
Get plot summarized scores
}
