% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{get_plot_theme}
\alias{get_plot_theme}
\title{Get standardized plot theme to add to figures}
\usage{
get_plot_theme(
  x_axis_text_size = 8,
  y_axis_text_size = 8,
  x_axis_title_size = 12,
  y_axis_title_size = 12,
  legend_title_size = 10,
  legend_text_size = 8,
  facet_x_text_size = 8,
  facet_y_text_size = 8,
  plot_title_size = 10,
  x_axis_dates = FALSE
)
}
\arguments{
\item{x_axis_text_size}{integer indicating the size of x axis text, to be
passed to theme. Default is \code{8}}

\item{y_axis_text_size}{integer indicating the size of y axis text, to be
passed to theme. Default is \code{8}}

\item{x_axis_title_size}{integer indicating the size of x axis title, to be
passed to theme. Default is \code{10}}

\item{y_axis_title_size}{integer indicating the size of y axis title, to be
passed to theme. Default is \code{10}}

\item{legend_title_size}{integer indicating the size of the legend title
Default is \code{10}}

\item{legend_text_size}{integer indicating the size of the legend items
Default is \code{10}}

\item{facet_x_text_size}{integer indicating the size of facet title on the
x-xaxis, to be passed to theme. Default is \code{8}}

\item{facet_y_text_size}{integer indicating the size of facet title on the
x-xaxis, to be passed to theme. Default is \code{8}}

\item{plot_title_size}{integer indicating thesize of plot title, to be passed
to theme. Default is \code{10}}

\item{x_axis_dates}{boolean indicating whether or not the x axis are dates.
If they are dates, we will rotate x axis tick text 45 degrees. Default is
\code{FALSE}}
}
\value{
a theme object to add to a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object to specify
line size and formatting
}
\description{
Get standardized plot theme to add to figures
}
