% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_raw_output_suffix}
\alias{get_raw_output_suffix}
\title{Generate a standard-format output suffix for saving raw output
for a location/forecast date/scenario trio.}
\usage{
get_raw_output_suffix(location, forecast_date, scenario)
}
\arguments{
\item{location}{Name of the location, as a string.}

\item{forecast_date}{The forecast date, as a date or
in a format coercible by \code{\link[=as.Date]{as.Date()}}.}

\item{scenario}{Name of the scenario, as a string.}
}
\value{
The output suffix, as a string.
}
\description{
Generate a standard-format output suffix for saving raw output
for a location/forecast date/scenario trio.
}
