% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{get_rel_wis_all_time}
\alias{get_rel_wis_all_time}
\title{Get the relative wis from the hub formatted}
\usage{
get_rel_wis_all_time(all_scores)
}
\arguments{
\item{all_scores}{a tibble of the scores for both models, formatted
like the hub}
}
\value{
A tibble of the relative wis at each forecast date, location, and
horizon day
}
\description{
Get the relative wis from the hub formatted
}
