% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_rel_wis_real_time}
\alias{get_rel_wis_real_time}
\title{Get the relative wis for the real time scores}
\usage{
get_rel_wis_real_time(all_scores)
}
\arguments{
\item{all_scores}{a tibble of the scores for both models in real-time}
}
\value{
A tibble of the relative wis at each forecast date, location, and
horizon day
}
\description{
Get the relative wis for the real time scores
}
