% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{get_state_level_ww_quantiles}
\alias{get_state_level_ww_quantiles}
\title{Get quantiles for site-lab level wastewater}
\usage{
get_state_level_ww_quantiles(ww_draws)
}
\arguments{
\item{ww_draws}{a dataframe containing all the draws from the model estimated
site-lab level cocnentrations}
}
\value{
a dataframe containing the quantile value for the quantiles
required for the Hub submission for each site lab in the state
}
\description{
Get quantiles for site-lab level wastewater
}
