% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_stats_imp_forecasts_wis}
\alias{get_stats_imp_forecasts_wis}
\title{Get stats on number of improved forecasts in real time from wis}
\usage{
get_stats_imp_forecasts_wis(scores, threshold)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{threshold}{numeric indicating fold change for considering a forecast
improved or worse relative to baseline, e.g. 1.1}
}
\value{
table of the number of states with improvements, number of overall
forecasts with improvements, number that got worse, etc.
}
\description{
Get stats on number of improved forecasts in real time from wis
}
