% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplement_analyses_and_figs.R
\name{get_stats_improved_forecasts}
\alias{get_stats_improved_forecasts}
\title{Get stats on number of improved forecasts}
\usage{
get_stats_improved_forecasts(scores, threshold)
}
\arguments{
\item{scores}{tibble of scores for every location, forecast date, and horizon}

\item{threshold}{numeric between 0 and 1 indicating the threshold to call
something better or worse}
}
\value{
table of the number of states with improvements, number of overall
forecasts with improvements, number that got worse, etc.
}
\description{
Get stats on number of improved forecasts
}
