% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig3.R
\name{get_summary_table_fig3}
\alias{get_summary_table_fig3}
\title{Get a summary of a single states crps}
\usage{
get_summary_table_fig3(scores, locs_to_plot, fig_file_dir)
}
\arguments{
\item{scores}{tibble of crps scores by location, forecast date, model,
horizon day}

\item{locs_to_plot}{the locations we want summaries for}

\item{fig_file_dir}{string indicating directory to save fig in}
}
\value{
a table with mean crps for each model and the relative crps
}
\description{
Get a summary of a single states crps
}
