% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_and_summarize_ww_data.R
\name{get_summary_ww_table}
\alias{get_summary_ww_table}
\title{Get wastewater data summary tables}
\usage{
get_summary_ww_table(ww_metadata)
}
\arguments{
\item{ww_metadata}{a tibble containing one row for every forecast date
location containing metrics averaged across that forecast date location
about the wastewater data}
}
\value{
a list containing a summary overall table, a summary by forecast
date and a summary by state
}
\description{
Get wastewater data summary tables
}
