% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{get_table_sufficient_ww}
\alias{get_table_sufficient_ww}
\title{Get table of location-forecast dates with sufficient wastewater}
\usage{
get_table_sufficient_ww(combined_ww_data_flags)
}
\arguments{
\item{combined_ww_data_flags}{A large tibble that consists of row bound
ww data flags for each individual run}
}
\value{
table_of_loc_dates_w_ww a tibble containing the location,
forecast_date, and column stating the wastewater data was sufficient for all
locations and forecast dates where ww data was deemed sufficient
}
\description{
This function takes in a large tibble with all the combined wastewater data
flags from each of the model runs, and produces a table with a row for each
model run indicating whether the wastewater data is sufficient, based
on the thresholds defined in \code{\link[=get_ww_data_flags]{get_ww_data_flags()}}
}
