% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{get_ww_data_flags}
\alias{get_ww_data_flags}
\title{Get table of location-forecast dates with sufficient wastewater}
\usage{
get_ww_data_flags(
  input_ww_data,
  forecast_date,
  delay_thres = 21,
  n_dps_thres = 5,
  prop_below_lod_thres = 0.5,
  sd_thres = 0.1,
  mean_log_ww_value_thres = -4
)
}
\arguments{
\item{input_ww_data}{A tibble with the wastewater data used to fit the
model for a particular location, forecast date, and scenario. Required
columns are \code{ww}, \code{date}, \code{below_lod}, and \code{location}}

\item{forecast_date}{A string indicating the date of the forecast for
this particular fit in ISO8 format (YYYY-MM-DD)}

\item{delay_thres}{The maximum number of days of delay between the last
wastewater data point and the forecast date, before we would flag a state as
having insufficient wastewater data to inform a forecast. Default is 21}

\item{n_dps_thres}{The threshold number of data points within a single site
within a state before we would flag the state as having insufficient
wastewater data to inform a forecast. Default is 5}

\item{prop_below_lod_thres}{The threshold proportion of wastewater data
points that can be below the LOD. If greater than this proportion of points
are below the LOD, we flag the state as having insufficient wastewater data.
Default is 0.5}

\item{sd_thres}{The minimum standard deviation between wastewater data points
within a site. This is intended to catch when a site reports all the same
values. Default is 0.1}

\item{mean_log_ww_value_thres}{The minimum value of the log of the ww
concentration, default is -4}
}
\value{
a tibble containing the location,
forecast_date, and 5 rows indicating the flags for the input wastewater data
associated with this individual run
}
\description{
This function takes in a the input wastewater data from an individual
location, forecast date, and scenario and produces a table with metadata
about the quality of the wastewater data, including values of the
number of data points, the mean value of the wastewater concentraiton,
and the proportion of data below the LOD. The thresholds passed to this
function are used to generate the boolean flags, indicated in the \code{name}
and \code{value} columns.
}
