% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/additional_figures.R
\name{heatmap_scores_by_loc_date}
\alias{heatmap_scores_by_loc_date}
\title{Plot a heatmap of avg forecast performance
by location and forecast date}
\usage{
heatmap_scores_by_loc_date(scores, metric, fig_file_dir)
}
\arguments{
\item{scores}{A tibble of daily scores by forecast date, location, and model}

\item{metric}{Metric to plot. One of \code{"wis"} or \code{"crps"}}

\item{fig_file_dir}{A string indicating the directory to save the figures in}
}
\value{
a ggplot object
}
\description{
Plot a heatmap of avg forecast performance
by location and forecast date
}
