% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score.R
\name{load_real_time_quantile_fcsts}
\alias{load_real_time_quantile_fcsts}
\title{Load real-time quantile outputs}
\usage{
load_real_time_quantile_fcsts(
  real_time_output_dir,
  table_of_run_ids,
  locations,
  eval_data,
  model_type
)
}
\arguments{
\item{real_time_output_dir}{A string indicating the upper
level directory where the real-time outputs live locally}

\item{table_of_run_ids}{A tibble containing the forecast date, run id,
and date run for each of the production runs}

\item{locations}{A vector of character strings indicating
the locations to pull, this should be all jurisdictions}

\item{eval_data}{a tibble of hospital admissions evaluation
data to be used for scoring.}

\item{model_type}{String indicating model type to load.
One of \code{"ww"} or \code{"hosp"}.}

\item{dates}{A vector of forecast dates to pull}
}
\value{
The forecasts, as the output of
\code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}
}
\description{
Load real-time quantile outputs
}
