% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig2_ct}
\alias{make_fig2_ct}
\title{Make concentration fit and forecast figure}
\usage{
make_fig2_ct(
  ww_quantiles,
  loc_to_plot,
  date_to_plot,
  n_forecast_days = 28,
  n_calib_days = 90,
  max_n_site_labs_to_show = 3
)
}
\arguments{
\item{ww_quantiles}{A tibble containing the calibrated wastewater
concentrations, the evaluation wastewater concentration data, and the
quantiles of the calibrated and forecasted wastewater concentrations}

\item{loc_to_plot}{A character string indicating the state abbreviation
for which state to plot, can only be one state}

\item{date_to_plot}{A character string indicating what forecast date to plot,
in IS08601 format YYYY-MM-DD}

\item{n_forecast_days}{An integer indicating the number of days to show the
forecast for, default is \code{28}}

\item{n_calib_days}{An integer indicating the number of days to show the
calibration data for, default is \code{90}}

\item{max_n_site_labs_to_show}{An integer indicating the maximum number
of site-labs to show in the figure, default is \code{3}}
}
\value{
A ggplot object containing a faceted horizontal plot of the
calibrated and forecasted wastewater concentrations for 3 or fewer
site-lab combinations for a single state
}
\description{
Make concentration fit and forecast figure
}
