% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig2_hosp_t}
\alias{make_fig2_hosp_t}
\title{Make a plot comparing the fit and forecasted hospital admissions
from the wastewater and hospital admissions model for a few
example states}
\usage{
make_fig2_hosp_t(
  hosp_quantiles,
  locs_to_plot,
  date_to_plot,
  n_forecast_days = 28,
  n_calib_days = 90
)
}
\arguments{
\item{hosp_quantiles}{A tibble containing the calibrated hospital admissions
data, the evaluation hospital admissions data, and the quantiles of the
calibrated and forecasted admissions}

\item{locs_to_plot}{A vector of strings indicating the state abbreviations
for which states to plot}

\item{date_to_plot}{A character string indicating what forecast date to plot,
in IS08601 format YYYY-MM-DD}

\item{n_forecast_days}{An integer indicating the number of days to show the
forecast for, default is \code{28}}

\item{n_calib_days}{An integer indicating the number of days to show the
calibration data for, default is \code{90}}
}
\value{
a ggplot object containing a faceted vertical plot of the
forecasts produced for each state, comparing the wastewater and hospital
admissions models
}
\description{
Make a plot comparing the fit and forecasted hospital admissions
from the wastewater and hospital admissions model for a few
example states
}
