% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig3.R
\name{make_fig3}
\alias{make_fig3}
\title{Title}
\usage{
make_fig3(
  fig3_crps_single_loc1,
  fig3_forecast_comparison_nowcast1,
  fig3_forecast_comparison_1wk1,
  fig3_forecast_comparison_4wks1,
  fig3_crps_underlay_nowcast1,
  fig3_crps_underlay_1wk1,
  fig3_crps_underlay_4wks1,
  fig3_crps_single_loc2,
  fig3_forecast_comparison_nowcast2,
  fig3_forecast_comparison_1wk2,
  fig3_forecast_comparison_4wks2,
  fig3_crps_underlay_nowcast2,
  fig3_crps_underlay_1wk2,
  fig3_crps_underlay_4wks2,
  fig3_crps_single_loc3,
  fig3_forecast_comparison_nowcast3,
  fig3_forecast_comparison_1wk3,
  fig3_forecast_comparison_4wks3,
  fig3_crps_underlay_nowcast3,
  fig3_crps_underlay_1wk3,
  fig3_crps_underlay_4wks3,
  fig_file_dir
)
}
\arguments{
\item{fig3_crps_single_loc1}{first states crps density plot}

\item{fig3_forecast_comparison_nowcast1}{first states nowcast comparison}

\item{fig3_forecast_comparison_1wk1}{first states 1 wk forecast comparison}

\item{fig3_forecast_comparison_4wks1}{first states 4 wk forecast comparison}

\item{fig3_crps_underlay_nowcast1}{first states crps nowcast underlay}

\item{fig3_crps_underlay_1wk1}{first states crps 1 wk underlay}

\item{fig3_crps_underlay_4wks1}{first states crps 4wk underlay}

\item{fig3_crps_single_loc2}{second states crps density plot}

\item{fig3_forecast_comparison_nowcast2}{second states nowcast comparison}

\item{fig3_forecast_comparison_1wk2}{second states 1 wk forecast comparison}

\item{fig3_forecast_comparison_4wks2}{second states 4 wk forecast comparison}

\item{fig3_crps_underlay_nowcast2}{second states crps nowcast underlay}

\item{fig3_crps_underlay_1wk2}{second states crps 1 wk underlay}

\item{fig3_crps_underlay_4wks2}{second states crps 4wk underlay}

\item{fig3_crps_single_loc3}{first state's crps density plot}

\item{fig3_forecast_comparison_nowcast3}{third states nowcast comparison}

\item{fig3_forecast_comparison_1wk3}{third states 1 wk forecast comparison}

\item{fig3_forecast_comparison_4wks3}{third states 4 wk forecast comparison}

\item{fig3_crps_underlay_nowcast3}{third states crps nowcast underlay}

\item{fig3_crps_underlay_1wk3}{third states crps 1 wk underlay}

\item{fig3_crps_underlay_4wks3}{third states crps 4wk underlay}

\item{fig_file_dir}{Path to save figures}
}
\value{
ggplot object that is a combination of 3 states overall crps
distributions comparing the two model types +
forecast comparisons across horizons with an underlay indicating the crps
score
}
\description{
Title
}
