% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manuscript_figs.R
\name{make_fig3_rel_crps_overall}
\alias{make_fig3_rel_crps_overall}
\title{Make figure that stratifies across location and forecast dates}
\usage{
make_fig3_rel_crps_overall(scores)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
the ouput of \code{scoringutils::score()} on samples.}
}
\value{
A ggplot object containing plots of the distribution of relative
CRPS scores across location and forecast dates
}
\description{
Make figure that stratifies across location and forecast dates
}
