% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4}
\alias{make_fig4}
\title{Make Figure 4}
\usage{
make_fig4(
  fig4_rel_crps_heatmap,
  fig4_rel_crps_hist,
  fig4_avg_crps,
  fig4_natl_admissions,
  fig4_rel_crps_over_time,
  fig4_rel_crps_by_location,
  time_period,
  fig_file_dir
)
}
\arguments{
\item{fig4_rel_crps_heatmap}{heatmap of relative crps by forecast date
and location}

\item{fig4_rel_crps_hist}{histogram comparing overall distribution
of crps scores across forecast_date, date, location, and model}

\item{fig4_avg_crps}{avg crps across locations by forecast date}

\item{fig4_natl_admissions}{national admissions by day}

\item{fig4_rel_crps_over_time}{relative crps across locations by forecast
date}

\item{fig4_rel_crps_by_location}{avg crps across forecast dates by state}

\item{time_period}{string to save fig as, either "real_time" or "all_time"}

\item{fig_file_dir}{Path to save figures}
}
\value{
ggplot object with all the elements combined
}
\description{
Make Figure 4
}
