% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4}
\alias{make_fig4}
\title{Make Figure 4}
\usage{
make_fig4(
  fig4_rel_crps_overall,
  fig4_avg_crps,
  fig4_natl_admissions,
  fig4_rel_crps_over_time,
  fig4_rel_crps_by_location,
  fig4_qq_plot_overall,
  fig4_plot_coverage_range,
  fig_file_dir
)
}
\arguments{
\item{fig4_rel_crps_overall}{density plot comparing overall distribution
of crps scores across forecast_date, date, location, and model}

\item{fig4_avg_crps}{avg crps across locations by forecast date}

\item{fig4_natl_admissions}{national admissions by day}

\item{fig4_rel_crps_over_time}{relative crps across locations by forecast
date}

\item{fig4_rel_crps_by_location}{avg crps across forecast dates by state}

\item{fig4_qq_plot_overall}{overall qq plot}

\item{fig4_plot_coverage_range}{interval coverage plots at 3 intervals}

\item{fig_file_dir}{Path to save figures}
}
\value{
ggplot object with all the elements combined
}
\description{
Make Figure 4
}
