% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4_admissions_overall}
\alias{make_fig4_admissions_overall}
\title{Make a figure of overall admissions for context}
\usage{
make_fig4_admissions_overall(
  eval_hosp_data,
  first_forecast_date,
  last_forecast_date
)
}
\arguments{
\item{eval_hosp_data}{Hospital admissions data for evaluating against,
for all locations}

\item{first_forecast_date}{The first forecast date we are evaluating}

\item{last_forecast_date}{The last forecast date we are evaluating}
}
\value{
ggplot object containing total hospital admissions in the US
}
\description{
Make a figure of overall admissions for context
}
