% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_fig4_rel_crps_overall}
\alias{make_fig4_rel_crps_overall}
\title{Make figure that stratifies across location and forecast dates}
\usage{
make_fig4_rel_crps_overall(
  scores,
  horizons_to_show = c("nowcast", "1 wk", "4 wks", "overall")
)
}
\arguments{
\item{scores}{A tibble of scores by location, forecast date, date and model,
containing the outputs of \code{scoringutils::score()} on samples plus metadata
transformed into a tibble.}

\item{horizons_to_show}{A vector of strings indicating the names of the
\code{horizon} that we want to show on the plot, must be a subset of
\code{nowcast}, \verb{1 wk}, \verb{2 wks},\verb{3 wks}, \verb{4 wks} and \code{overall}}

\item{summarize_across_horizon}{Boolean indicating whether or not to
average the scores across the horizon, default is \code{FALSE} meaning
each day-forecast-date-location score is in the distribution}
}
\value{
A ggplot object containing plots of the distribution of relative
CRPS scores across location and forecast dates
}
\description{
Make figure that stratifies across location and forecast dates
}
