% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5}
\alias{make_fig5}
\title{Make Fig 5}
\usage{
make_fig5(
  fig5_plot_wis_t_real_time,
  fig5_density_real_time,
  fig5_plot_real_time_rel_wis,
  fig5_plot_wis_t_all_time,
  fig5_density_all_time,
  fig5_plot_all_time_rel_wis,
  fig5_heatmap_rel_wis_all_time,
  fig5_heatmap_rel_wis_feb_mar,
  fig5_qq_plot_all_time,
  fig5_qq_plot_feb_mar,
  fig5_std_rank_feb_mar,
  fig5_std_rank_all_time,
  fig_file_dir
)
}
\arguments{
\item{fig5_plot_wis_t_real_time}{average wis over time across locations for
each model in real-time (feb-mar)}

\item{fig5_density_real_time}{density plot of relative performance
across location, forecast_date, day, and model}

\item{fig5_plot_real_time_rel_wis}{heatmap of the relative wis between
the two models}

\item{fig5_plot_wis_t_all_time}{average wis over time across locations for
each model in real-time (feb-mar)}

\item{fig5_density_all_time}{density plot of relative performance
across location, forecast_date, day, and model}

\item{fig5_plot_all_time_rel_wis}{heatmap of the relative wis between
the two models}

\item{fig5_heatmap_rel_wis_all_time}{heatmap comparing WIS across
forecast dates for each location for all time}

\item{fig5_heatmap_rel_wis_feb_mar}{heatmap comparing WIS across
forecast dates for each location for the real-time period (Feb-Mar)}

\item{fig5_qq_plot_all_time}{qq plot comparing model coverage for all time}

\item{fig5_qq_plot_feb_mar}{qq plot comparing model coverage for the
real-time period (Feb-Mar)}

\item{fig5_std_rank_feb_mar}{comparison of standardized rank across models
for the real-time period (Feb-Mar)}

\item{fig5_std_rank_all_time}{comparison of standardized rank across models
for all time}

\item{fig_file_dir}{Path to save figures}
}
\value{
a ggplot object containing all the figures combined
}
\description{
Make Fig 5
}
