% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_average_wis}
\alias{make_fig5_average_wis}
\title{Get plot of WIS over time}
\usage{
make_fig5_average_wis(
  all_scores,
  cfa_real_time_scores = c(),
  models_to_show,
  time_period,
  horizon_time_in_weeks = NULL
)
}
\arguments{
\item{all_scores}{Scores from entire time period of interest, including
the retrospective cfa model, scored with \code{\link[scoringutils:score]{scoringutils::score()}} and
summarized across quantiles only with \code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}}}

\item{cfa_real_time_scores}{Real-time scores from Feb - Mar for the cfa ww
model submitted to the hub, scored with \code{\link[scoringutils:score]{scoringutils::score()}} and
summarized across quantiles only with \code{\link[scoringutils:summarise_scores]{scoringutils::summarise_scores()}},
default is \code{c()}}

\item{models_to_show}{A vector of charcter strings indicating which models
from the COVID-19 forecast hub to include in the plot.}

\item{time_period}{String indicating the time period this plot pertains to}

\item{horizon_time_in_weeks}{horizon time in weeks to summarize over, default
is \code{NULL} which means that the scores are summarized over the nowcast period
and the 4 week forecast period}
}
\value{
a ggplot object of WIS scores over time colored by model, for the
real-time cfa model from Feb - Mar and the retrospective CFA model over
all time points
}
\description{
Get plot of WIS over time
}
