% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_average_wis}
\alias{make_fig5_average_wis}
\title{Get plot of WIS over time}
\usage{
make_fig5_average_wis(
  all_scores,
  cfa_real_time_scores,
  horizon_time_in_weeks = NULL
)
}
\arguments{
\item{all_scores}{Scores from entire time period of interest, including
the retrospective cfa model}

\item{cfa_real_time_scores}{Real-time scores from Feb - Mar for the cfa ww
model submitted to the hub}

\item{horizon_time_in_weeks}{horizon time in weeks to summarize over, default
is \code{NULL} which means that the scores are summarized over 4 weeks}
}
\value{
a ggplot object of WIS scores over time colored by model, for the
real-time cfa model from Feb - Mar and the retrospective CFA model over
all time points
}
\description{
Get plot of WIS over time
}
