% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_bar_chart}
\alias{make_fig5_bar_chart}
\title{Make plot of WIS scores in Hub models overall}
\usage{
make_fig5_bar_chart(scores, time_period)
}
\arguments{
\item{scores}{quantile based scores from the hub}

\item{time_period}{string indicating which time period to make the plot for}

\item{fig_file_dir}{directory to save figure}
}
\value{
A plot ordered by average wis over the time period
}
\description{
Make plot of WIS scores in Hub models overall
}
