% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{make_fig5_density_rank}
\alias{make_fig5_density_rank}
\title{Make a figure of the distribution of standardized WIS rank}
\usage{
make_fig5_density_rank(
  scores,
  models_to_show,
  time_period,
  tp_fp,
  fig_file_dir
)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{models_to_show}{A vector of charcter strings indicating which models
from the COVID-19 forecast hub to include in the plot.}

\item{time_period}{time period that scores are summarized over}

\item{tp_fp}{string indicating short name for time period to save figure}

\item{fig_file_dir}{directory to save figure}
}
\value{
A ggplot object containing geomridges plots colored by density,
indicating the standardized rank for each location-date combo
}
\description{
Adapted from https://www.pnas.org/doi/10.1073/pnas.2113561119
and https://github.com/reichlab/covid19-forecast-evals/blob/b741b6a24e40c7f2a8ddc41da40c95b23db6df4e/code/figure-model-ranks.R#L11 #nolint
}
