% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_density_rank}
\alias{make_fig5_density_rank}
\title{Make a figure of the distribution of standardized WIS rank}
\usage{
make_fig5_density_rank(scores, time_period)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{time_period}{time period that scores are summarized over}
}
\value{
A ggplot object containing geomridges plots colored by density,
indicating the standardized rank for each location-date combo
}
\description{
Adapted from https://www.pnas.org/doi/10.1073/pnas.2113561119
and https://github.com/reichlab/covid19-forecast-evals/blob/b741b6a24e40c7f2a8ddc41da40c95b23db6df4e/code/figure-model-ranks.R#L11 #nolint
}
