% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_qq_plot}
\alias{make_fig5_qq_plot}
\title{Make a quantile quantile plot for the Hub}
\usage{
make_fig5_qq_plot(scores, time_period)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations, forecast
dates and horizons}

\item{time_period}{time period that scores are summarized over}
}
\value{
a ggplot object containing a plot of the proportion of data within
each interval for each model.
}
\description{
Make a quantile quantile plot for the Hub
}
