% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig5.R
\name{make_fig5_table_and_plot}
\alias{make_fig5_table_and_plot}
\title{Make summary table of WIS scores in Hub models overall}
\usage{
make_fig5_table_and_plot(scores, time_period, fig_file_dir)
}
\arguments{
\item{scores}{quantile based scores from the hub}

\item{fig_file_dir}{directory to save figure}
}
\value{
A table with the average scores of each model over the time period
}
\description{
Make summary table of WIS scores in Hub models overall
}
