% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig4.R
\name{make_qq_plot_overall}
\alias{make_qq_plot_overall}
\title{Make a qq plot}
\usage{
make_qq_plot_overall(scores_quantiles)
}
\arguments{
\item{scores_quantiles}{A tibble of scores by location, forecast date,
date and model, containing the outputs of \code{scoringutils::score()} on
quantiles plus metadata transformed into a tibble.}
}
\value{
a ggplot object with the overall QQ plot colored by model.
}
\description{
Using \code{\link[scoringutils:plot_interval_coverage]{scoringutils::plot_interval_coverage()}}
}
