% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generic.R
\name{model_type_eyeplot}
\alias{model_type_eyeplot}
\title{Side by side eyeplots by model type.}
\usage{
model_type_eyeplot(data, y)
}
\arguments{
\item{data}{Data frame to plot, with a \code{model_type} column.}

\item{y}{Column in \code{df} to plot on the y axis.}
}
\value{
The plot.
}
\description{
Side by side eyeplots by model type.
}
