% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generic.R
\name{model_type_slabinterval}
\alias{model_type_slabinterval}
\title{Side by side slabinterval plot by model type.}
\usage{
model_type_slabinterval(
  data,
  y,
  geom = "slabinterval",
  shape = 21,
  point_size = 10,
  interval_size_range = c(2, 5),
  ...
)
}
\arguments{
\item{data}{Data frame to plot, with a \code{model_type} column.}

\item{y}{Column in \code{data} to plot on the y axis.}

\item{geom}{\code{geom} argument passed to \code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}}.
Default \code{"slabinterval"}.}

\item{shape}{\code{shape} argument passed to
\code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}}. Default \code{21}.}

\item{point_size}{\code{point_size} argument passed to
\code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}}. Default \code{10}.}

\item{interval_size_range}{\code{interval_size_range} argument
passed to \code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}}. Default \code{c(2, 5)}.}

\item{...}{Additional keyword arguments passed to
\code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}} .}
}
\value{
The plot.
}
\description{
Wraps \code{\link[ggdist:stat_slabinterval]{ggdist::stat_slabinterval()}}.
Defaults to a half-eyeplot.
}
