% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms_fig3.R
\name{multi_location_crps_figure}
\alias{multi_location_crps_figure}
\title{Generate a bar plots of CRPS for each model in different locations}
\usage{
multi_location_crps_figure(scores, locs_to_plot, fig_file_dir)
}
\arguments{
\item{scores}{tibble of crps scores by location, forecast date, model,
horizon day}

\item{locs_to_plot}{the locations we want summaries for}

\item{fig_file_dir}{string indicating directory to save fig in}
}
\value{
Figure showing CRPS for multiple locations.
}
\description{
Generate a bar plots of CRPS for each model in different locations
}
