% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecasttools.R
\name{nhsn_soda_query}
\alias{nhsn_soda_query}
\title{Construct a Socrata open data
API (SODA) query for the NSHN
dataset}
\usage{
nhsn_soda_query(
  api_endpoint,
  start_date = NULL,
  end_date = NULL,
  columns = NULL,
  states = NULL,
  limit = 1e+05,
  order_by = c("state", "date"),
  desc = FALSE,
  ...
)
}
\arguments{
\item{api_endpoint}{Base API endpoint URL to use
when constructing the query.}

\item{start_date}{Pull only rows with dates
greater than or equal to this date. If \code{NULL},
no minimum date. Default \code{NULL}.}

\item{end_date}{Pull only rows with dates
less than or equal to this date. If \code{NULL},
no maximum date. Default \code{NULL}.}

\item{columns}{Vector of columns to retrieve, in
addition to \code{date} and \code{state}, which are always
retrieved. If \code{NULL}, retrieve all columns.
Default \code{NULL}.}

\item{states}{Vector of states or territories to
retrieve, by two letter US postal service code.
If \code{NULL}, retrieve all. Default \code{NULL}.}

\item{limit}{limit to the number of rows to retrieve.
Default 1e5.}

\item{order_by}{Vector of columns by which to order the
results. Default \code{c("state", "date")}}

\item{desc}{whether to order descending instead of
ascending. Default \code{FALSE} (order ascending).}

\item{...}{additional arguments (ignored for now)}
}
\value{
the query as \code{\link[soql:soql]{soql::soql()}} output
}
\description{
Construct a Socrata open data
API (SODA) query for the NSHN
dataset
}
