% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_horizons.R, R/order_elements.R
\name{order_horizons}
\alias{order_horizons}
\title{Get the order of the categorical horizons for plotting}
\usage{
order_horizons(df)

order_horizons(df)
}
\arguments{
\item{df}{a dataframe containing the column name \code{horizon} which will
contain character strings indicating the horizon (by week, nowcast period,
or overall) of the score/quantile/sample.}
}
\value{
a dataframe containing the same columns as \code{df} with the \code{horizon}
column now an ordered factor with ordered levels as follows:
\verb{"overall", "calibration", "nowcast", "1 wk", "2 wks", "3 wks", "4 wks"}

a dataframe containing the same columns as \code{df} with the \code{horizon}
column now an ordered factor with ordered levels as follows:
\verb{"overall", "calibration", "nowcast", "1 wk", "2 wks", "3 wks", "4 wks"}
}
\description{
Get the order of the categorical horizons for plotting

Get the order of the categorical horizons for plotting
}
