% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_real_time_exclusions.R
\name{parse_real_time_exclusions}
\alias{parse_real_time_exclusions}
\title{Parse forecast exclusions from real-time run archived metadata}
\usage{
parse_real_time_exclusions(dir)
}
\arguments{
\item{dir}{Directory to parse, with subdirectories representing
forecast dates, each of which contains a \code{metadata.yaml} file.}
}
\value{
Table of the parsed exclusions, as a
tidy \code{\link[tibble:tibble]{tibble}}.
}
\description{
Parse forecast exclusions from real-time run archived metadata
}
