% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap_scores.R
\name{plot_bootstrapped_score_ratios}
\alias{plot_bootstrapped_score_ratios}
\title{Plot bootstrapped CRPS ratios as pointintervals}
\usage{
plot_bootstrapped_score_ratios(
  replicates,
  by = NULL,
  connect_points = FALSE,
  order_by_point_estimate = FALSE
)
}
\arguments{
\item{replicates}{Data frame of bootstrapped replicates,
as the output of \code{\link[=bootstrap_crps_values]{bootstrap_crps_values()}}.}

\item{by}{Stratification variable. Will become the x-axis
of the plot. Default \code{NULL} (plot a single point-interval.}

\item{connect_points}{Connect the points in the point intervals with lines?
Boolean, default \code{FALSE}.}

\item{order_by_point_estimate}{Order x-axis values by the value of the point estimate
(ascending)? Boolean, default \code{FALSE}.}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot bootstrapped CRPS ratios as pointintervals
}
