% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_fitting_clock_time}
\alias{plot_fitting_clock_time}
\title{Plot model fitting clock time.}
\usage{
plot_fitting_clock_time(chain_run_time)
}
\arguments{
\item{chain_run_time}{Data frame of chain run times summarized to
show only the slowest values per fit.}
}
\value{
The plot
}
\description{
Plot model fitting clock time.
}
