% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_fitting_clock_time_versus_sites}
\alias{plot_fitting_clock_time_versus_sites}
\title{Plot model fitting clock time as a function of number
of wastewater sampling sites.}
\usage{
plot_fitting_clock_time_versus_sites(clock_time, metadata)
}
\arguments{
\item{clock_time}{Data frame of chain run times, extracted
from a CmdStanR fit object and collated via \code{\link[=combine_outputs]{combine_outputs()}}.}

\item{metadata}{Data frame of wastewater metadata that gives the
number of sampling sites (\code{n_sites}) by \code{location} and \code{forecast_date}.}
}
\value{
The plot
}
\description{
Plot model fitting clock time as a function of number
of wastewater sampling sites.
}
