% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_fitting_clock_time_versus_sites}
\alias{plot_fitting_clock_time_versus_sites}
\title{Plot model fitting clock time as a function of number
of wastewater sampling sites.}
\usage{
plot_fitting_clock_time_versus_sites(clock_time, metadata)
}
\arguments{
\item{chain_run_time}{Data frame of chain run times, extracted
from a CmdStanR fit object and collated via \code{\link[=combine_outputs]{combine_outputs()}}.}
}
\value{
The plot
}
\description{
Plot model fitting clock time as a function of number
of wastewater sampling sites.
}
