% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{plot_hub_performance_by_period}
\alias{plot_hub_performance_by_period}
\title{Get plot of overall hub performance, grouped by period}
\usage{
plot_hub_performance_by_period(
  all_scores,
  cfa_real_time_scores,
  figure_file_path,
  all_time_period,
  real_time_period,
  models_to_show,
  summarize_across_horizon = FALSE,
  baseline_model = "COVIDhub-4_week_ensemble"
)
}
\arguments{
\item{all_scores}{df with granular (daily) scores from every model,
forecast_date, and location for the entire time period. Includes the
two retrospective models}

\item{cfa_real_time_scores}{df with granular (daily) scores from the
submitted cfa ww model for the time period when it was submitted.}

\item{figure_file_path}{path to directory to save figures}

\item{all_time_period}{string indicating the longer time frame we are
comparing, e.g. "Oct 2023-Mar 2024"}

\item{real_time_period}{string indicating the shorter time frame that
we submitted our model to the hub e.g. "Feb 2024-Mar 2024"}

\item{models_to_show}{A vector of charcter strings indicating which models
from the COVID-19 forecast hub to include in the plot.}

\item{summarize_across_horizon}{Boolean indicating whether or not to
average the scores across the horizon, default is \code{FALSE} meaning
each day-forecast-date-location score is in the distribution}

\item{baseline_model}{which model to compute relative WIS compared to, default
is \verb{COVIDhub-4_week_ensemble}}
}
\value{
a ggplot object containing distributions of WIS scores grouped by
model and the comaprison time period, with the mean plotted alongside the
full distribution
}
\description{
Get plot of overall hub performance, grouped by period
}
