% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_metadata.R
\name{plot_hub_submit_info_real_time}
\alias{plot_hub_submit_info_real_time}
\title{Plot model submission decisions in real-time}
\usage{
plot_hub_submit_info_real_time(dates, locations, exclusion_table)
}
\arguments{
\item{dates}{forecast dates to plot}

\item{locations}{locations to plot}

\item{hosp_subtitution_table}{table of exclusions,
as produced by \code{\link[=parse_real_time_exclusions]{parse_real_time_exclusions()}}.}
}
\value{
The plot, as a ggplot object.
}
\description{
Plot model submission decisions in real-time
}
