% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_generic.R
\name{plot_pointinterval_scatter}
\alias{plot_pointinterval_scatter}
\title{Generic plotting helper functions
Scatterplot of pointintervals, with both x and y intervals.}
\usage{
plot_pointinterval_scatter(data, x, y, ...)
}
\arguments{
\item{data}{Dataframe to plot, as the output of \code{\link[ggdist:mean_qi]{ggdist::mean_qi()}}
or a related function.}

\item{x}{Name of the x axis variable.}

\item{y}{Name of the y axis variable.}

\item{...}{keyword arguments passed to \code{\link[ggdist:geom_pointinterval]{ggdist::geom_pointinterval()}}.}
}
\value{
The plot as a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Generic plotting helper functions
Scatterplot of pointintervals, with both x and y intervals.
}
