% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_prior_posterior.R
\name{plot_prior_posterior}
\alias{plot_prior_posterior}
\title{Plot posterior draws for a given parameter against
a known marginal prior distribution.}
\usage{
plot_prior_posterior(
  draws_long,
  priors,
  fit_id_col = "forecast_date",
  variable_name = ".variable",
  value_name = ".value",
  prior_pdf_name = "prior_pdf",
  row_facet = "location"
)
}
\arguments{
\item{draws_long}{long-format posterior draws, as the
output of \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}}.}

\item{priors}{data frame of priors, with the same
faceting columns as \code{draws_long}.}

\item{fit_id_col}{Name of a column in \code{draws_long} that uniquely
identifies individual model fits within a facet. Default \code{"forecast_date"}.}

\item{variable_name}{Name for the variable column.
Default \code{".variable"}, matching  \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}}.}

\item{value_name}{Name for the value column.
Default \code{".value"}, matching  \code{\link[tidybayes:gather_draws]{tidybayes::gather_draws()}}.}

\item{prior_pdf_name}{Column containing prior pdfs. Default
\code{"pdf"}.}

\item{row_facet}{Column to facet rows by. Default \code{"location"}.}
}
\value{
The plot.
}
\description{
Plot posterior draws for a given parameter against
a known marginal prior distribution.
}
