% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_relative_scores.R
\name{plot_rel_score_dists}
\alias{plot_rel_score_dists}
\title{Make a dotsinterval plot of relative scores
stratified by specified variables.}
\usage{
plot_rel_score_dists(
  scores,
  target_models,
  baseline_model,
  metric_to_compare,
  x = NULL,
  by = NULL,
  order_x = FALSE
)
}
\arguments{
\item{scores}{Output of \code{\link[scoringutils:score]{scoringutils::score()}}.}

\item{target_models}{Model for which to plot relative scores}

\item{baseline_model}{Baseline model for the relative scores
computation.}

\item{metric_to_compare}{Metric for which to plot relative
scores. One of \code{"wis"} or \code{"crps"}. Not case-sensitive.}

\item{x}{Variable for the x axis. If \code{NULL}, plot a single
distribution.}

\item{by}{additional variables defining a single score to plot.
Passed along with the variable in \code{x} as the \code{by} argument to
\code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}. Default \code{NULL}
(no additional variables).}

\item{order_x}{Order the groups on the x-axis by their mean
value of the relative_metric (i.e. the point in the pointinterval)?
Default \code{FALSE}.}
}
\value{
a ggplot object with the distributions plotted by
forecast date.
}
\description{
Make a dotsinterval plot of relative scores
stratified by specified variables.
}
