% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_diagnostics.R
\name{plot_rhat_distribution}
\alias{plot_rhat_distribution}
\title{Plot a distribution of R-hat values.}
\usage{
plot_rhat_distribution(diagnostic_extrema)
}
\arguments{
\item{diagnostic_extrema}{data frame of diagnostic extrema,
as the output of \code{\link[=extract_diagnostic_extrema]{extract_diagnostic_extrema()}}, collated via
\code{\link[=combine_outputs]{combine_outputs()}}.}
}
\value{
The plot.
}
\description{
Plot a distribution of R-hat values.
}
