% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{plot_score_decomposed_bars}
\alias{plot_score_decomposed_bars}
\title{Barplot of WIS or CRPS decomposed into
overprediction, underprediction, and dispersion}
\usage{
plot_score_decomposed_bars(
  scores,
  x = "model",
  width = 0.8,
  position = "dodge2",
  ...
)
}
\arguments{
\item{scores}{quantile based scores from the hub}

\item{x}{Column containing x values. Default "model".}

\item{width}{width for the bars. Passed to \code{\link[=geom_decomposed_scores]{geom_decomposed_scores()}}.
Default 0.8.}

\item{position}{Position for the bars. Passed to
\code{\link[=geom_decomposed_scores]{geom_decomposed_scores()}}. Default \code{"dodge2"}.}

\item{...}{keyword arguments passed to \code{\link[=geom_decomposed_scores]{geom_decomposed_scores()}}.}
}
\value{
A plot of decomposed probabilistic scores.
}
\description{
Barplot of WIS or CRPS decomposed into
overprediction, underprediction, and dispersion
}
