% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{plot_score_scatter}
\alias{plot_score_scatter}
\title{Plot a scatterplot of scores comparing those of one model
to those of another, for a given stratification.}
\usage{
plot_score_scatter(
  scores,
  metric,
  model_x,
  model_y,
  by = NULL,
  label = NULL,
  nudge_x = 0,
  nudge_y = 0,
  label_size = 2,
  label_color = "black",
  ...
)
}
\arguments{
\item{scores}{Table of unsummarized scores, as the output of
[scoringutils::score().}

\item{metric}{Metric to plot.}

\item{model_x}{Model whose score should be plotted on the x axis.}

\item{model_y}{Model whose score should be plotted on the y axis.}

\item{by}{Summize scores by these columns. Passed as the
\code{by} argument to \code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.
Default \code{NULL}.}

\item{label}{Column to use for labeling points. If \code{NA} or \code{NULL},
do not label.}

\item{nudge_x}{Passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Default \code{0}.}

\item{nudge_y}{Passed to \code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Default \code{0}.}

\item{label_size}{Passed as the \code{"size"} argument to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Default \code{2}.}

\item{label_color}{Passed as the \code{"color"} argument to
\code{\link[ggplot2:geom_text]{ggplot2::geom_text()}}. Default \code{"black"}.}

\item{...}{Keyword arguments
passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\value{
The scatterplot, as ggplot object
}
\description{
Plot a scatterplot of scores comparing those of one model
to those of another, for a given stratification.
}
