% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{plot_score_scatter}
\alias{plot_score_scatter}
\title{Plot a scatterplot of scores comparing those of one model
to those of another, for a given stratification.}
\usage{
plot_score_scatter(scores, metric, model_x, model_y, by = NULL, ...)
}
\arguments{
\item{scores}{Table of unsummarized scores, as the output of
[scoringutils::score().}

\item{metric}{Metric to plot.}

\item{model_x}{Model whose score should be plotted on the x axis.}

\item{model_y}{Model whose score should be plotted on the y axis.}

\item{by}{Summize scores by these columns. Passed as the
\code{by} argument to \code{\link[forecasttools:summarise_scores_with_baseline]{forecasttools::summarise_scores_with_baseline()}}.
Default \code{NULL}.}

\item{...}{keyword arguments passed to \code{\link[ggplot2:geom_point]{ggplot2::geom_point()}}.}
}
\value{
The scatterplot, as ggplot object
}
\description{
Plot a scatterplot of scores comparing those of one model
to those of another, for a given stratification.
}
