% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_scores_by_scenario}
\alias{plot_scores_by_scenario}
\title{Barplot of scores by scenario}
\usage{
plot_scores_by_scenario(scenario_scores, score_metric = "crps")
}
\arguments{
\item{scenario_scores}{data frame of scenario scores to plot.}

\item{score_metric}{Metric to plot. Must be a column name in
\code{scenario_scores}. Default \code{"crps"}.}
}
\value{
The plot, as a ggplot object.
}
\description{
Barplot of scores by scenario
}
