% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{plot_std_rank_distribution}
\alias{plot_std_rank_distribution}
\title{Make a figure of the distribution of standardized WIS rank}
\usage{
plot_std_rank_distribution(scores, models_to_show)
}
\arguments{
\item{scores}{df of granular (daily) score across models, locations,
forecast dates and horizons}

\item{models_to_show}{Character vector indicating which models
from the COVID-19 forecast hub to include.}
}
\value{
A ggplot object containing geom denity ridges plots
colored by density, indicating the standardized rank for each
location-date combo
}
\description{
Adapted from https://www.pnas.org/doi/10.1073/pnas.2113561119
and https://github.com/reichlab/covid19-forecast-evals/blob/b741b6a24e40c7f2a8ddc41da40c95b23db6df4e/code/figure-model-ranks.R#L11 #nolint
}
