% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_admissions.R
\name{plot_total_admissions}
\alias{plot_total_admissions}
\title{Make a figure of overall hospital admissions eval data}
\usage{
plot_total_admissions(eval_hosp_data, first_date, last_date)
}
\arguments{
\item{eval_hosp_data}{Location-level admissions data to summarize}

\item{first_date}{First date to plot}

\item{last_date}{Last date to plot}
}
\value{
ggplot object displaying a timeseries of total
hospital admissions.
}
\description{
Make a figure of overall hospital admissions eval data
}
