% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_trend_versus_diff.R
\name{plot_trend_versus_diff}
\alias{plot_trend_versus_diff}
\title{Scatterplot of inferred data trends versus forecast differences.}
\usage{
plot_trend_versus_diff(data, trend_metric, diff_metric, fill_metric, ...)
}
\arguments{
\item{data}{Dataframe to plot, as the output of \code{\link[ggdist:point_interval]{ggdist::mean_qi()}}
or a related function.}

\item{trend_metric}{Name of the column containing the point estimate
of the trend.}

\item{diff_metric}{Name of the column containing the point estimate of
the difference.}

\item{fill_metric}{Name of the column to associate to the fill
aesthetic. If \code{NULL}, do not map the fill aesthetic.}

\item{...}{keyword arguments passed to \code{\link[ggdist:geom_pointinterval]{ggdist::geom_pointinterval()}}.}
}
\value{
The plot as a \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}} object.
}
\description{
Scatterplot of inferred data trends versus forecast differences.
}
