% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_recent_trend_fits.R
\name{process_recent_trend_fits}
\alias{process_recent_trend_fits}
\title{Process BRMS log-linear fits of recent trends}
\usage{
process_recent_trend_fits(
  hosp_fit,
  ww_fit,
  save_dir,
  figure_ext = "pdf",
  n_lab_sites_plot = 10
)
}
\arguments{
\item{hosp_fit}{\link[brms:brmsfit-class]{brms::brmsfit} object for the hospital admissions trend.}

\item{ww_fit}{\link[brms:brmsfit-class]{brms::brmsfit} object for the wastewater trend.}

\item{figure_ext}{File extension for figures, without the \code{.},
e.g. \code{"pdf"} or \code{"png"}. Default \code{"pdf"}.}

\item{n_lab_sites_plot}{Maximum number of lab-sites for which to
plot wastewater trends. Default \code{10}.}

\item{figure_save_dir}{Directory in which to save output
(figures and tidy posterior draws).}
}
\value{
NULL, invisibly saving postprocessing results to
disk as a side effect.
}
\description{
Process BRMS log-linear fits of recent trends
}
