% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_data.R
\name{pull_and_write_hosp_data}
\alias{pull_and_write_hosp_data}
\title{Pull and save the current state of NHSN data on HealthData.gov}
\usage{
pull_and_write_hosp_data(location_data_path, output_dir, force = FALSE)
}
\arguments{
\item{location_data_path}{Path to a csv file containing data
about constituent locations, with columns \code{"abbreviation"} and
\code{"population"}}

\item{output_dir}{Directory in which to save the output.}

\item{force}{Overwrite an existing \code{.csv} with the same date-stamp?
Boolean, default \code{FALSE}.}
}
\value{
NULL, saving the data to disk as a side effect.
}
\description{
Pull data using \code{\link[=pull_nhsn]{pull_nhsn()}}, format it for pipeline ingestion,
and write it to disk as a date-stamped \code{.csv} file.
}
\examples{
\dontrun{
# this will write a file named `<today's date>.csv` to a directory named
# `input/hosp_data/vintage_datasets`.
locs <- file.path("input", "locations.csv")
out <- file.path("input", "hosp_data", "vintage_datasets")
pull_and_write_hosp_data(locs, out)
}
}
