% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{pull_hub_forecasts}
\alias{pull_hub_forecasts}
\title{Pull hub submissions and create a scorable table
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.}
\usage{
pull_hub_forecasts(
  model_name,
  dates,
  locations = NULL,
  hub_subdir = NA,
  pull_from_github = TRUE,
  submissions_path =
    "https://raw.githubusercontent.com/reichlab/covid19-forecast-hub/master/data-processed/",
  truth_data_path =
    "https://media.githubusercontent.com/media/reichlab/covid19-forecast-hub/master/data-truth/truth-Incident\%20Hospitalizations.csv"
)
}
\arguments{
\item{model_name}{a vector of character strings indicating
the names of the models to score.}

\item{dates}{a vector of dates indicating the dates of the
submissions to score.}

\item{locations}{a vector of character strings indicating the locations
to score}

\item{hub_subdir}{path where the retrospective hub submissions are saved
locally since these are not on COVIDHub github}

\item{pull_from_github}{boolean indicating whether or not to pull
from github}

\item{submissions_path}{url pointing to the "data-processed" folder on
the COVIDhub github, which is where team's submissions are located}

\item{truth_data_path}{the path to the truth data used by the hub for
evaluation}
}
\value{
a dataframe containing all of the scores for all models,
forecast dates (indicated by dates), locations, target end dates, and
quantiles
}
\description{
Pull hub submissions and create a scorable table
as the output of \code{\link[scoringutils:as_forecast_quantile]{scoringutils::as_forecast_quantile()}}.
}
