% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{relative_wis_histogram}
\alias{relative_wis_histogram}
\title{Plot a histogram of individual forecast relative WIS values}
\usage{
relative_wis_histogram(
  scores,
  models_to_show,
  baseline_model = "COVIDhub-4_week_ensemble"
)
}
\arguments{
\item{scores}{Table of raw scores to plot.}

\item{models_to_show}{Character vector of models to plot.}

\item{baseline_model}{Name of the baseline model against
which to compute relative scores. Default \code{"COVIDhub-4_week_ensemble"}.}
}
\description{
Plot a histogram of individual forecast relative WIS values
}
