% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_table_sufficient_ww.R
\name{save_only_ww_data}
\alias{save_only_ww_data}
\title{Save only the wastewater data from the output quantiles}
\usage{
save_only_ww_data(ww_quantiles, ww_data_dir)
}
\arguments{
\item{ww_quantiles}{The full set of wastewater quantiles for each
date, forecast date, state}

\item{ww_data_dir}{directory where wastewater data is saved}
}
\description{
Save only the wastewater data from the output quantiles
}
