% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess.R
\name{save_table}
\alias{save_table}
\title{Save table}
\usage{
save_table(
  data_to_save,
  type_of_output,
  output_dir,
  scenario,
  forecast_date,
  model_type = c("ww", "hosp"),
  location
)
}
\arguments{
\item{data_to_save}{The dataframe/tibble to save}

\item{type_of_output}{The name of the type of output e.g. scores, quantiles,
etc.}

\item{output_dir}{The upper level directory to save these outputs in}

\item{scenario}{A  string indicating the scenario under which the
model was run}

\item{forecast_date}{A string indicating the date of the forecast,
in YYYY-MM-DD}

\item{model_type}{A string indicating the type of model (either \code{ww} or \code{hosp})}

\item{location}{A string indicating the location (e.g. 2 letter abbreviation
for the state)}
}
\description{
This helper function is a wrapper to save intermediate outputs
running within the model fit loop as tsvs in the following file strucuture:
scenario > forecast_date > model_type > location > type_of_output
}
