% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_alpha_score_component}
\alias{scale_alpha_score_component}
\title{ggplot scale for fill alphas by name of
WIS/CRPS score component.}
\usage{
scale_alpha_score_component(...)
}
\arguments{
\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_alpha_manual]{ggplot2::scale_alpha_manual()}}}
}
\value{
the ggplot scale
}
\description{
ggplot scale for fill alphas by name of
WIS/CRPS score component.
}
