% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_model}
\alias{scale_fill_model}
\alias{scale_color_model}
\alias{scale_shape_model}
\title{ggplot discrete scales for forecast models}
\usage{
scale_fill_model(name = "Model", ...)

scale_color_model(name = "Model", ...)

scale_shape_model(name = "Model", ...)
}
\arguments{
\item{name}{Display name for the scale. Default \code{"Model"}.}

\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_manual]{ggplot2::scale_*_manual()}}.}
}
\value{
the ggplot scale
}
\description{
ggplot discrete scales for forecast models
}
