% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_style.R
\name{scale_fill_score_ratio}
\alias{scale_fill_score_ratio}
\title{ggplot fill scale for score ratios}
\usage{
scale_fill_score_ratio(...)
}
\arguments{
\item{...}{Keyword arguments passed to
\code{\link[ggplot2:scale_gradient]{ggplot2::scale_fill_gradient2()}}}
}
\description{
ggplot fill scale for score ratios
}
