% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{score_hub_forecasts}
\alias{score_hub_forecasts}
\title{Score hub forecasts}
\usage{
score_hub_forecasts(hub_forecasts)
}
\arguments{
\item{hub_forecasts}{hub quantile forecasts, as the output
of \code{\link[=pull_hub_forecasts]{pull_hub_forecasts()}}.}
}
\value{
Data frame of scores, as the output of \code{\link[scoringutils:score]{scoringutils::score()}}.
}
\description{
Score hub forecasts
}
