% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_real_time_outputs.R
\name{score_real_time_outputs}
\alias{score_real_time_outputs}
\title{Load in and score the real-time outputs}
\usage{
score_real_time_outputs(
  score_type,
  real_time_output_dir,
  table_of_run_ids,
  locations,
  dates,
  eval_data,
  hosp_only = FALSE
)
}
\arguments{
\item{score_type}{A string indicating which score to generate, either
"crps" or "wis". Note, if using crps, will score draws from nowcast and
forecast. If using wis, will score only the forecasts.}

\item{real_time_output_dir}{A string indicating the upper level directory
where the real-time outputs live locally}

\item{table_of_run_ids}{A tibble containing the forecast date, run id,
and date run for each of the production runs}

\item{locations}{A vector of character strings indicating the locations to
pull, this should be all jurisdictions}

\item{dates}{A vector of forecast dates to pull}

\item{eval_data}{a tibble of hospital admissions evaluation data to be used
for scoring.}

\item{hosp_only}{boolean indicating if we should only pull the hospital
admissions model}
}
\value{
A large tibble containing crps scores for every location and
forecast date, conditioned on the presence of wastewater and model
convergence
}
\description{
Load in and score the real-time outputs
}
