% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub.R
\name{select_hub_models}
\alias{select_hub_models}
\title{Query Zoltar for models to include in the analysis}
\usage{
select_hub_models(
  min_submissions_per_model,
  min_locations_per_submission,
  forecast_dates,
  locations,
  project_name = "COVID-19 Forecasts",
  excluded_models = c("COVIDhub_CDC-ensemble", "COVIDhub-trained_ensemble")
)
}
\arguments{
\item{min_submissions_per_model}{Number of forecast dates with a
submission for a model to be included in the analysis.}

\item{min_locations_per_submission}{Minimum number of included locations
for a given submsission to count toward \code{min_submissions_per_model}.}

\item{forecast_dates}{vector of dates formatted in ISO8601 convention
(YYYY-MM-DD) indicating the forecast dates for the analysis}

\item{locations}{vector of state abbreviations that we want to
ensure the submitting teams have produced forecasts for.}

\item{project_name}{name of the Zoltar project, default is
\code{"COVID-19 Forecasts"}.}

\item{excluded_models}{Models to exclude even if they meet the other
criteria. Defaults to
\code{c("COVIDhub_CDC-ensemble", "COVIDhub-trained_ensemble")}
(Two non-individual models that are not the Hub ensemble or the
Hub baseline).}
}
\value{
a vector of character strings indicating the
unique model names that fit the inclusion criteria
}
\description{
Query Zoltar for models to include in the analysis
}
