% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hub_comparison_figure.R
\name{summarize_std_rank}
\alias{summarize_std_rank}
\title{Summarize standardized ranks with medians and 25th,75th percentiles}
\usage{
summarize_std_rank(scores)
}
\arguments{
\item{scores}{A tibble of the individual day and location's scores}
}
\value{
A table with median, 25th, and 75th percentiles of standard
ranking for each model
}
\description{
Summarize standardized ranks with medians and 25th,75th percentiles
}
