% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summarize_ww_data_quality.R
\name{summarize_ww_data_quality}
\alias{summarize_ww_data_quality}
\title{Get table of location-forecast dates with available and
sufficient wastewater}
\usage{
summarize_ww_data_quality(ww_quality_table)
}
\arguments{
\item{ww_quality_table}{Data frame of flags for each run,
as the row-bound output of one or more calls to
\code{\link[=compute_ww_data_quality_table]{compute_ww_data_quality_table()}}.}
}
\value{
a summarized table with locations, forecast dates, a
\code{wwsufficient} boolean column indicating whether there is valid
wastewater data and a \code{status} column indicating the reason
for data (in)validity.
#' @export
}
\description{
This function takes in a large tibble with all the combined wastewater
data flags from each of the model runs, and produces a table with a
row for each model run indicating the wastewater data status for that
run.
}
