% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_absolute_scores.R
\name{wis_barplot}
\alias{wis_barplot}
\title{Make plot of WIS scores in Hub models overall}
\usage{
wis_barplot(scores)
}
\arguments{
\item{scores}{quantile based scores from the hub}
}
\value{
A plot ordered by WIS over the time period
}
\description{
Make plot of WIS scores in Hub models overall
}
