% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{with_dependencies}
\alias{with_dependencies}
\title{Helper function to note targets dependencies not explicitly
noted in the function call}
\usage{
with_dependencies(x, ...)
}
\arguments{
\item{...}{additional arguments, which can be used to indicate additional upstream targets
to treat as dependencies when this function is used in a target \code{command}}

\item{output}{to be passed along (typically the output of a function call}
}
\value{
the first argument
}
\description{
Helper function to note targets dependencies not explicitly
noted in the function call
}
