% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wweval-package.R
\docType{package}
\name{wweval-package}
\alias{wweval}
\alias{wweval-package}
\title{wweval: Evaluation of wastewater informed COVID-19 hospital admissions forecasting}
\description{
This package provides helper functions designed to preprocess, postprocess, and analyze outputs from fitting a semi-mechanistic renewal model with and without incorporating wastewater.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/}
  \item Report bugs at \url{https://github.com/cdcgov/wastewater-informed-covid-forecasting/issues/}
}

}
\author{
\strong{Maintainer}: Kaitlyn Johnson \email{johnsonkaitlyne9@gmail.com} (\href{https://orcid.org/0000-0001-8011-0012}{ORCID})

Authors:
\itemize{
  \item Sam Abbott \email{contact@samabbott.co.uk} (\href{https://orcid.org/0000-0001-8057-8037}{ORCID})
  \item Zachary Susswein \email{utb2@cdc.gov}
  \item Andrew Magee \email{rzg0@cdc.gov}
  \item Dylan Morris \email{dylan@dylanhmorris.com} (\href{https://orcid.org/0000-0002-3655-406X}{ORCID})
  \item Scott Olesen \email{ulp7@cdc.gov}
  \item Damon Bayer \email{xum8@cdc.gov}
}

Other contributors:
\itemize{
  \item George Vega Yon \email{g.vegayon@gmail.com} (\href{https://orcid.org/0000-0002-3171-0844}{ORCID}) [contributor]
}

}
\keyword{internal}
